% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\docType{class}
\name{AbstractConstraintSettings-class}
\alias{AbstractConstraintSettings-class}
\alias{AbstractConstraintSettings}
\title{AbstractClass for Constraint Settings.}
\value{
An \code{AbstractConstraintSettings} object.
}
\description{
The \code{ConstraintSettings} class encapsulates the constraints
on the physicochemical properties of primers.
}
\section{Slots}{

\describe{
\item{\code{status}}{Named boolean vector indicating 
which of the possible constraints are active (\code{TRUE})
and which are not (\code{FALSE}).}

\item{\code{settings}}{Named list containing the 
settings of the active constraints}
}}

\keyword{internal}
