% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_ILP.R
\name{add.coverage.constraints}
\alias{add.coverage.constraints}
\title{Addition of Coverage Constraints.}
\usage{
add.coverage.constraints(lprec, covered.templates, template.coverage)
}
\arguments{
\item{lprec}{An ILP instance.}

\item{covered.templates}{Indices of covered template sequences.}

\item{template.coverage}{List containing the indices of covering primers for each template.}
}
\value{
\code{lprec} with coverage constraints.
}
\description{
Adds coverage constraints to ILP instance.
}
\keyword{internal}
