% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_ILP.R
\name{add.dimerization.constraints}
\alias{add.dimerization.constraints}
\title{Addition of Dimerization Constraints}
\usage{
add.dimerization.constraints(lprec, D.idx, indices)
}
\arguments{
\item{lprec}{An ILP instance.}

\item{D.idx}{Data frame giving the indices of dimerizing primer pairs.}

\item{indices}{Row indices for setting the dimerization constraints in \code{lprec}.}
}
\value{
\code{lprec} with added dimerization constraints.
}
\description{
Updates ILP formulation with dimerization constraints.
}
\keyword{internal}
