% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{add.uniform.leaders.to.seqs}
\alias{add.uniform.leaders.to.seqs}
\title{Add Uniform Binding Regions.}
\usage{
add.uniform.leaders.to.seqs(lex.seq, leaders)
}
\arguments{
\item{lex.seq}{Template data frame}

\item{leaders}{Data frame with uniform binding regions.}
}
\value{
Template data frame with updated binding regions.
}
\description{
Augments a template data frame with uniform binding regions.
}
\keyword{internal}
