% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stop_codons.R
\name{adjust.ORF.start}
\alias{adjust.ORF.start}
\title{Adjust ORF position}
\usage{
adjust.ORF.start(ORFs, seq.start)
}
\arguments{
\item{ORFs}{the reading frames (0,1,2).}

\item{seq.start}{the position where a sequence is extracted .}
}
\value{
The adjusted reading frames for the given start positions.
}
\description{
Adjusts the reading frame according to the position at which we consider a
subsequence.
}
\keyword{internal}
