% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{adjust_binding_regions}
\alias{adjust_binding_regions}
\title{Adjustment of Existing Binding Regions.}
\usage{
adjust_binding_regions(template.df, region.fw, region.rev)
}
\arguments{
\item{template.df}{A \code{Templates} object providing the template sequences
for which the binding regions shall be adjusted.}

\item{region.fw}{Interval of new binding regions relative to the forward binding region defined in \code{template.df}.}

\item{region.rev}{Interval of new binding regions relative to the reverse binding region defined in \code{template.df}}
}
\value{
A \code{Templates} object with updated binding regions.
}
\description{
Adjusts the existing annotation of binding regions by specifying
a new binding interval relative to the existing binding region.
}
\details{
The new binding intervals provided by \code{fw} and \code{rev}
for forward and reverse primers, respectively, are provided
relative to the existing definition of binding regions in \code{template.df},
which can be set using \code{\link{assign_binding_regions}}.
When specifying relative positions, position \code{0} is defined as the first position after the end of the existing
binding region. Hence, negative positions relate to regions within the
existing binding region while non-negative values 
extend the binding region further.
}
\examples{
data(Ippolito)
# Extend the binding region by one position
relative.interval <- c(-max(template.df$Allowed_End_fw), 0)
template.df.adj <- adjust_binding_regions(template.df, relative.interval)
# compare old and new annotations:
head(cbind(template.df$Allowed_Start_fw, template.df$Allowed_End_fw))
head(cbind(template.df.adj$Allowed_Start_fw, template.df.adj$Allowed_End_fw))
}
\keyword{Templates}
