% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_primers_tree.R
\name{align.seqs}
\alias{align.seqs}
\title{Multiple Sequence Alignment}
\usage{
align.seqs(seqs, names)
}
\arguments{
\item{seqs}{The sequences to be aligned.}

\item{names}{The identifiers of the sequences.}
}
\value{
An alignment object as created from seqinr's read.alignment method.
}
\description{
Computes a multiple sequence alignment using MAFFT.
}
\references{
Katoh, Misawa, Kuma, Miyata 2002 (Nucleic Acids Res. 30:3059-3066) 
MAFFT: a novel method for rapid multiple sequence alignment based
on fast Fourier transform.
}
\keyword{internal}
