% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{align.structures}
\alias{align.structures}
\title{Formatting of Dimerization Structures.}
\usage{
align.structures(structs)
}
\arguments{
\item{s1}{strutcs A character vector, where a block of 4 elements
contains: sequence 1 (with removed overlaps), part of sequence 1
overlappign with sequence 2, part of sequence 2 
overapping with sequence 1, and sequence 2 (with removed overlaps).}
}
\value{
A list of two elements givin the conformation of the first
and the second sequence, respectively.
}
\description{
Formats the given dimerization structures nicely.
}
\keyword{internal}
