% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_primers_tree.R
\name{ancestor_of}
\alias{ancestor_of}
\title{Tree Ancestry}
\usage{
ancestor_of(tree, ancestor.node, test.node)
}
\arguments{
\item{tree}{The phylogenetic tree to be tested.}

\item{ancestor.node}{A node to be checked for being an ancestor to \code{test.node}.}

\item{test.node}{Possible descendants of \code{ancestor.node}.}
}
\value{
TRUE, if \code{ancestor.node} is an ancestor to any node in \code{test.node}.
}
\description{
Checks whether \code{ancestor.node} is an ancestor to the nodes 
specified in \code{test.node}.
}
\keyword{internal}
