% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_annealing_temperature.R
\name{annealing.temp.rule.of.thumb}
\alias{annealing.temp.rule.of.thumb}
\title{Rule of thumb for annealing temperature}
\usage{
annealing.temp.rule.of.thumb(melting.temp)
}
\arguments{
\item{melting.temp}{Melting temperatures of primers}
}
\value{
The annealing temperature corresponding to the input melting temperature.
}
\description{
Computes the annealing temperature using a rule of thumb
}
\keyword{internal}
