% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{annotate.binding.events}
\alias{annotate.binding.events}
\title{Annotation of Primer Binding Events.}
\usage{
annotate.binding.events(fw.binding, allowed.range, nbr.primers,
  allowed.region.definition = c("within", "any"))
}
\arguments{
\item{fw.binding}{IRanges with coverage information.}

\item{allowed.range}{IRanges of the allowed binding ranges in the templates.}

\item{nbr.primers}{Number of primers to consider.}

\item{allowed.region.definition}{Definition of the allowed binding region}
}
\value{
IRanges with annotations of (preliminary) specificity and allowed binding. 
The field \code{all_binding} contains all binding regions, \code{on_target} contains
all events in the target region, and \code{off_target} contains all off-target
binding events.
}
\description{
Annotates whether primer binding events are in the allowed binding region or not.
}
\keyword{internal}
