% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_eval.R
\name{apply.constraint}
\alias{apply.constraint}
\title{Application of Constraints}
\usage{
apply.constraint(gc.ratio.fw, gc.ratio.rev, min.GC, max.GC, fw.idx, rev.idx,
  mode.directionality = c("fw", "rev", "both"))
}
\arguments{
\item{gc.ratio.fw}{Forward values.}

\item{gc.ratio.rev}{Reverse values.}

\item{min.GC}{Minimal allowed value.}

\item{max.GC}{Maximal allowed value.}

\item{fw.idx}{Indices of forward values to consider.}

\item{rev.idx}{Indices of reverse values to consider.}

\item{mode.directionality}{Direction of primers}
}
\value{
Data frame with \code{TRUE} for values fulfilling the constraints, 
\code{FALSE} otherwise.  Also returns \code{FALSE} if a data point is not available.
}
\description{
Checks whether the input values are within the specified limits.
}
\keyword{internal}
