% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{assign_binding_regions.character}
\alias{assign_binding_regions.character}
\title{Character Assignment of Binding Regions}
\usage{
\method{assign_binding_regions}{character}(template.df, fw, rev,
  optimize.region = FALSE, primer.length = 20, gap.char = "-")
}
\arguments{
\item{template.df}{Template data frame.}

\item{fw}{FASTA file specifying the forward binding regions.}

\item{rev}{FASTA file specifying the reverse binding regions.}

\item{optimize.region}{Should the primer binding region be optimized using secondary structure prediction?}

\item{primer.length}{Probe length for optimizing template secondary structure.}

\item{gap.char}{The gap character for aligned sequences.}
}
\value{
Template data frame with assigned binding regions.
}
\description{
Generic method for assigning the binding region using individual binding regions.
}
\keyword{internal}
