% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{assign_binding_regions.numeric}
\alias{assign_binding_regions.numeric}
\title{Numeric Assignment of Binding Regions.}
\usage{
\method{assign_binding_regions}{numeric}(template.df, fw, rev,
  optimize.region = FALSE, primer.length = 20, gap.char = "-")
}
\arguments{
\item{template.df}{Template data frame.}

\item{fw}{Binding region data forward primers.}

\item{rev}{Binding region data for reverse primers.}

\item{optimize.region}{Should the primer binding region be optimized using secondary structure prediction?}

\item{primer.length}{Probe length for optimizing template secondary structure.}
}
\value{
The template data frame with assigned binding regions.
}
\description{
Numeric S3 generic case for assigning binding regions.
}
\keyword{internal}
