% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{assign_binding_regions}
\alias{assign_binding_regions}
\title{Assignment of Template Binding Regions.}
\usage{
assign_binding_regions(template.df, fw = NULL, rev = NULL,
  optimize.region = FALSE, primer.length = 20, gap.char = "-")
}
\arguments{
\item{template.df}{A \code{Templates} object containing the sequences
for which primer binding regions should be annotated.}

\item{fw}{Binding regions for forward primers. Either a numeric interval indicating a uniform
binding range relative to the template 5' end or a path to a FASTA file providing
binding sequences for every template. If \code{fw} is missing, only
\code{rev} is considered.}

\item{rev}{Binding regions for reverse primers. Either a numeric interval indicating a uniform
binding range relative to the template 3' end or the path to a FASTA file providing
binding sequences for every template. If \code{rev} is missing,
only \code{fw} is considered.}

\item{optimize.region}{If \code{TRUE}, the binding regions
specified via \code{fw} and \code{rev} are 
adjusted such that binding regions that may form secondary structures are 
avoided. This feature requires ViennaRNA (see notes). If \code{FALSE}
(the default), the input binding regions are not modified.}

\item{primer.length}{A numeric scalar providing the probe length that is used for
for adjusting the primer binding regions when \code{optimize.region} is \code{TRUE}.}

\item{gap.char}{The character indicating gaps in aligned sequences.
The default is "-".}
}
\value{
An object of class \code{Templates} with assigned binding regions.
}
\description{
Assigns the primer target binding regions to a set of template sequences.
}
\details{
The arguments \code{fw} and \code{rev} provide data describing
the binding regions of the forward and reverse primers, respectively.
To specify binding regions for each template
individually, \code{fw} and \code{rev} should provide the paths to FASTA files. The headers of these
FASTA file should match the headers of the loaded \code{template.df} 
and the sequences in the files specified by \code{fw} and \code{rev} should indicate
the target binding regions.

To specify uniform binding regions,
\code{fw} and \code{rev} should be numeric intervals indicating the allowed
binding range for primers in the templates. The \code{fw} interval is 
specified with respect to the 5' end, while the \code{rev} interval is
specified with respect to the template 3' end.
If \code{optimize.region} is \code{TRUE}, the input binding region is
adjusted such that regions forming secondary structures are avoided.
}
\note{
The program ViennaRNA (https://www.tbi.univie.ac.at/RNA/) must be
installed to perform the computations that are triggered
when \code{optimize.region} is set to \code{TRUE}.
}
\examples{
data(Ippolito)
# Assignment of individual binding regions
l.fasta.file <- system.file("extdata", "IMGT_data", "templates", 
     "Homo_sapiens_IGH_functional_leader.fasta", package = "openPrimeR")
template.df.individual <- assign_binding_regions(template.df, l.fasta.file, NULL)
# Assign the first/last 30 bases as forward/reverse binding regions
template.df.uniform <- assign_binding_regions(template.df, c(1,30), c(1,30))
# Optimization of binding regions (requires ViennaRNA)
\dontrun{template.df.opti <- assign_binding_regions(template.df, c(1,30), c(1,30),
                     optimize.region = TRUE, primer.length = 20)}
}
\seealso{
Other template functions: \code{\link{Templates-class}},
  \code{\link{read_templates}},
  \code{\link{update_template_cvg}},
  \code{\link{write_templates}}
}
\keyword{Templates}
