% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{batchify}
\alias{batchify}
\title{Creates multiple Batches for Parallelization.}
\usage{
batchify(tasks, annealing.temps = NULL)
}
\arguments{
\item{tasks}{An integer vector with indices representing individual computations.}

\item{annealing.temps}{Temperatures according to which to batchify.}
}
\value{
A list of lists containing indices corresponding to \code{tasks}, each list gives a batch.
}
\description{
Creates multiple Batches for Parallelization.
}
\keyword{internal}
