% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{batchify.simple}
\alias{batchify.simple}
\title{Simple Batchification}
\usage{
batchify.simple(tasks)
}
\arguments{
\item{tasks}{The tasks to assign to individual batches.}
}
\value{
A list of lists containing indices corresponding to \code{tasks}, each list gives a batch.
}
\description{
Simple Batchification
}
\keyword{internal}
