% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{batchify.temp}
\alias{batchify.temp}
\title{Batchification by Temperature.}
\usage{
batchify.temp(tasks, annealing.temp)
}
\arguments{
\item{tasks}{The tasks to assign to individual batches.}

\item{annealing.temp}{The annealing temperatures according to which batches are to be created.}
}
\value{
A list of lists containing indices corresponding to \code{tasks}, each list gives a batch.
}
\description{
Batchification by Temperature.
}
\keyword{internal}
