% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{build.gain.df}
\alias{build.gain.df}
\title{Gain of Coverage by Excluded Primers.}
\usage{
build.gain.df(candidate.df, constraint.settings, constraint.limits, relax.df)
}
\arguments{
\item{candidate.df}{An object of class \code{Primers} containing
excluded primers that are considered for addition to \code{filtered.df}.}

\item{constraint.settings}{A list with the current constraint settings.}

\item{constraint.limits}{The current constraint limits.}

\item{relax.df}{Data frame with count of relaxations per constraint.}
}
\value{
A data frame with exclusion data.
}
\description{
Computes a data frame on the excluded sequences per constraint.
}
\keyword{internal}
