% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{build.tool.overview}
\alias{build.tool.overview}
\title{Creation of an Overview of Third-Party Tools.}
\usage{
build.tool.overview(AVAILABLE.TOOLS, for.shiny = FALSE)
}
\arguments{
\item{AVAILBLE.TOOLS}{A vector whose names give the required tools and whose
entries give their installation status as logicals.}

\item{If}{\code{for.shiny} is \code{TRUE}, provide the URLs for the tool using HTML.}
}
\value{
A data frame with information on third-part tools.
}
\description{
Creates a table of required third-party tools and their installation status.
}
\keyword{internal}
