% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{build_leader_df}
\alias{build_leader_df}
\title{Building of Leader Data Frame.}
\usage{
build_leader_df(direction = c("fw", "rev"), leader, start, end, ali.start,
  ali.end)
}
\arguments{
\item{direction}{The primer direction for which we are annotating binding regions.}

\item{leader}{The binding region sequence.}

\item{start}{The start positions of the binding region.}

\item{end}{The end positions of the binding region.}

\item{ali.start}{The start positions of the binding region in the aligned input.}

\item{ali.end}{The end positions of the binding region in the aligned input.}
}
\value{
A data frame with binding region information.
}
\description{
Constructs the leader data frame.
}
\keyword{internal}
