% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{cbind.Primers}
\alias{cbind.Primers}
\title{cbind for Primers class.}
\usage{
\method{cbind}{Primers}(...)
}
\arguments{
\item{...}{Parameters for cbind function.}
}
\value{
Column binded Primers data frame.
}
\description{
Ensures that the cbind result has the appropriate class.
}
\examples{
data(Ippolito)
primer.df <- cbind(primer.df, primer.df)
}
\keyword{internal}
