% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{cbind.Templates}
\alias{cbind.Templates}
\title{cbind for Template class.}
\usage{
\method{cbind}{Templates}(...)
}
\arguments{
\item{...}{Parameters for cbind function.}
}
\value{
Column binded Templates data frame.
}
\description{
Ensures that the cbind result has the appropriate class.
}
\examples{
data(Ippolito)
template.df <- cbind(template.df, seq_len(nrow(template.df)))
}
\keyword{internal}
