% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{check.3prime.hexamers}
\alias{check.3prime.hexamers}
\title{3' Hexamer Check.}
\usage{
check.3prime.hexamers(template.df, primer.df, mode.directionality = c("fw",
  "rev", "both"))
}
\arguments{
\item{template.df}{Template data frame.}

\item{primer.df}{Primer data frame.}

\item{mode.directionality}{Primer directionality.}
}
\value{
Returns \code{TRUE} if the 3' hexamer of a primer is fully complementary
to the corresponding template region and \code{FALSE} otherwise.
}
\description{
Check whether the 3' hexamer of a primer is fully complementary
to the corresponding region in the template.
}
\keyword{internal}
