% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_primers.R
\name{check.init.primer.length}
\alias{check.init.primer.length}
\title{Primer Length Check.}
\usage{
check.init.primer.length(template.df, allowed.region.definition = c("within",
  "any"), primer.lengths, mode.directionality = c("fw", "rev", "both"))
}
\arguments{
\item{template.df}{Template data frame.}

\item{allowed.region.definition}{Definition of allowed binding regions.}

\item{primer.lengths}{The desired lengths of the priemrs.}

\item{mode.directionality}{The primer directionality.}
}
\value{
TRUE, if primers of the desired length can be constructed,
}
\description{
Checks whether it is possible to construct primers of the desired length.
}
\keyword{FALSE}
\keyword{internal}
\keyword{otherwise.}
