% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_primers.R
\name{check.init.primer.length.single}
\alias{check.init.primer.length.single}
\title{Primer Length Check.}
\usage{
check.init.primer.length.single(allowed,
  allowed.region.definition = c("within", "any"), min.len)
}
\arguments{
\item{allowed}{String containing the allowed binding sequence.}

\item{allowed.region.definition}{Definition of allowed binding regions.}

\item{min.len}{Minimal desired primer lengths.}
}
\value{
TRUE if primers of the desired length can be constructed,
FALSE otherwise.
}
\description{
Checks whether it is possible to construct primers of the desired length.
}
\keyword{internal}
