% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_eval.R
\name{check.template.constraints}
\alias{check.template.constraints}
\title{Check Constraints on Templates}
\usage{
check.template.constraints(template.constraints)
}
\arguments{
\item{template.constraints}{Strings with comma-separated values to be turned to logical.}
}
\value{
List with boolean values
}
\description{
Transforms the comma-separated input strings to a boolean representation.
}
\keyword{internal}
