% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check.tool.function}
\alias{check.tool.function}
\title{Check Functionality of Third-Party Tools.}
\usage{
check.tool.function(frontend = FALSE)
}
\arguments{
\item{frontend}{Whether tool functionality shall be checked for the frontend.}
}
\value{
\code{TRUE} for each functioning tool, \code{FALSE} for non-functioning tools.
}
\description{
Checks whether all required tools should work.
}
\keyword{internal}
