% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check.tool.installation}
\alias{check.tool.installation}
\title{Check Tool Installation}
\usage{
check.tool.installation(frontend = FALSE)
}
\arguments{
\item{frontend}{Whether tool installation shall be checked for the frontend.
If \code{TRUE}, dependencies that are required only by the frontend are considered additionally.}
}
\value{
\code{TRUE} for each installed tool, \code{FALSE} otherwise.
}
\description{
Checks whether all required tools are installed.
}
\keyword{internal}
