% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{check_constraint_settings_validity}
\alias{check_constraint_settings_validity}
\title{Check the Validity of the Constraint Settings.}
\usage{
check_constraint_settings_validity(object)
}
\arguments{
\item{object}{An \code{AbstractConstraintSettings} object.}
}
\value{
Error messages in case of errors, otherwise \code{TRUE}.
}
\description{
Checks whether the status and the active constraints match.
Determines whether the constraints are allowed/known.
}
\keyword{internal}
