% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{check_interval}
\alias{check_interval}
\title{Check Constraint Intervals}
\usage{
check_interval(constraints)
}
\arguments{
\item{constraints}{A list with constraint settings.}
}
\value{
\code{TRUE}, if all constraints specificy valid intervals, 
\code{FALSE} otherwise.
}
\description{
Checks the validity of constraint intervals.
}
\keyword{internal}
