% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{check_limit_value}
\alias{check_limit_value}
\title{Check of limit correctness.}
\usage{
check_limit_value(setting, limit)
}
\arguments{
\item{setting}{A single constraint setting.}

\item{limit}{A single constraint limit.}
}
\value{
A vector containing \code{TRUE} if the limit is more general
than the constraint setting and \code{FALSE} otherwise.
}
\description{
Checks whether a constraint limit is more general than the setting.
}
\keyword{internal}
