% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{check_limits}
\alias{check_limits}
\title{Validity Check for Limits.}
\usage{
check_limits(constraint.settings, constraint.limits)
}
\arguments{
\item{constraint.settings}{A list with the constraint settings.}

\item{constraint.limits}{A list with the constraint relaxation limits.}
}
\value{
\code{TRUE} if the limits are at least as wide as the constraints, 
\code{FALSE} otherwise.
}
\description{
Checks whether the constraint limits are at least as
general as the constraint settings. This ensures that
the relaxation works in the proper direction.
}
\keyword{internal}
