% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{check_names}
\alias{check_names}
\title{Check Setting Names.}
\usage{
check_names(known.options, input.options)
}
\arguments{
\item{known.options}{Allowed setting names.}

\item{input.options}{Input setting names}
}
\value{
Mapping of \code{input.options} to \code{known.options} or \code{NULL} if invalid.
}
\description{
Checks whether the specified settings hvae the correct names.
}
\keyword{internal}
