% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{check_report_deps}
\alias{check_report_deps}
\title{Check for Report Dependencies.}
\usage{
check_report_deps()
}
\value{
A logical vector giving the dependency availability status.
}
\description{
Checks whether the dependencies for rmarkdown::render() are fulfilled.
}
\keyword{internal}
