% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{check_restriction_sites}
\alias{check_restriction_sites}
\title{Identification of Sequence Restriction Sites.}
\usage{
check_restriction_sites(primer.df, template.df, adapter.action = c("warn",
  "rm"), selected = NULL, only.confident.calls = TRUE,
  updateProgress = NULL)
}
\arguments{
\item{primer.df}{A \code{Primers} object containing the primer nucleotide
sequences to be checked for restriction sites.}

\item{template.df}{An object of class \code{Templates} containing
the templates corresponding to \code{primer.df}.}

\item{adapter.action}{The action to be performed when adapter sequences
are found. Either "warn" to issue a warning about adapter sequences or
"rm" to remove identified adapter sequences. Currently, only
the default setting ("warn") is supported.}

\item{selected}{Names of restriction sites that are to be checked.
By default \code{selected} is \code{NULL} in which case all REBASE 
restriction sites are taken into account.}

\item{only.confident.calls}{Whether only confident calls
of restriction sites are returned.
All restriction site call is considered \emph{confident} if the restriction site
is located in a region that does not match the template sequences.
Note that this classification requires that the provided primers
are somehow complementary to the provided templates.
In contrast, non-confident restriction site calls are 
based solely on the primer sequences and do not take the templates
into account, resulting in more false positive calls of restriction sites.}

\item{updateProgress}{A Shiny progress callback function. The default
is \code{NULL} meaning that no progress is tracked via the Shiny app.}
}
\value{
A data frame with possible restriction sites found in every primer.
}
\description{
Checks a set of primers for the presence of
restriction sites. To reduce the number of possible restriction sites,
only unambiguous restriction sites are taken into account and 
only common (typically used) restriction sites are checked if a common
restriction site can be found in a sequence.
}
\examples{
data(Ippolito)
site.df <- check_restriction_sites(primer.df, template.df)
}
\references{
Roberts, R.J., Vincze, T., Posfai, J., Macelis, D. (2010) REBASE–a database for DNA restriction
and modification: enzymes, genes and genomes. Nucl. Acids Res. 38: D234-D236. http://rebase.neb.com
}
\seealso{
Other primer functions: \code{\link{Primers-class}},
  \code{\link{check_constraints}},
  \code{\link{create_report}},
  \code{\link{design_primers}},
  \code{\link{filter_primers}},
  \code{\link{get_initial_primers}},
  \code{\link{primer_significance}},
  \code{\link{score_degen}}, \code{\link{write_primers}}
}
