% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{check_setting}
\alias{check_setting}
\title{Check Setting Validity.}
\usage{
check_setting(known.options, options, mandatory.options = NULL)
}
\arguments{
\item{known.options}{Vector with names and classes of allowed options.}

\item{options}{Active options to be checked.}

\item{mandatory.options}{Fields that have to be present.}
}
\value{
\code{TRUE} if the setting is valid, \code{FALSE} otherwise.
}
\description{
Checks whether the input settings are valid or not.
}
\keyword{internal}
