% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{check_settings_validity}
\alias{check_settings_validity}
\title{Validity Check for DesignSettings.}
\usage{
check_settings_validity(object)
}
\arguments{
\item{object}{A DesignSettings object to be checked for validity.}
}
\value{
\code{TRUE} if \code{object} is valid, FALSE otherwise.
}
\description{
Validates whether a DesignSettings object has the correct structure.
}
\keyword{internal}
