% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{classify_design_problem}
\alias{classify_design_problem}
\title{Classification of the Difficulty of a Primer Design Task.}
\usage{
classify_design_problem(template.df, mode.directionality = c("both", "fw",
  "rev"), primer.length = 18, primer.estimate = FALSE, required.cvg = 1)
}
\arguments{
\item{template.df}{A \code{Templates} object providing
the template sequences for which the difficulty of designing
primers shall be estimated.}

\item{mode.directionality}{The directionality of the
primers that are to be designed. Either
\code{fw} for forward primers, \code{rev} for reverse primers,
or \code{both} for primers of both directions. By default,
both directions are considered.}

\item{primer.length}{A scalar numeric providing the 
target length of the designed primers. The default length 
of generated primers is set to \code{18}.}

\item{primer.estimate}{Whether the number of required primers shall be estimated. By default (\code{FALSE}), the number of required primers is not estimated.}

\item{required.cvg}{A scalar numeric in the range [0,1] providing the target coverage ratio for designing primers. 
The \code{required.cvg} is used only when \code{primer.estimate} is set to \code{TRUE} such that a solution to the set cover problem is required.}
}
\value{
A list with the following fields:
\describe{
\item{\code{Classification}}{The estimated difficulty of the primer design task.}
\item{\code{Class-Distances}}{The total variance distance of the fitted
beta distribution to the reference distribution.}
\item{\code{Confidence}}{The confidence in the estimate of the
design tasks' difficulty as based on the class distances.}
\item{\code{Uncertain}}{Whether the classification is highly uncertain, that is
low-confidence.}
\item{\code{Nbr_primers_fw} and \code{Nbr_primers_rev}}{The respective number of 
required forward and reverse primers if \code{primer.estimate} was set to \code{TRUE}.}
}
}
\description{
Uses reference beta distributions of primer coverage ratios to
classify a primer design task into the groups ranging from \emph{easy} 
to \emph{hard}. For \emph{easy} tasks, it should
not be a problem to design a small primer set. For
\emph{hard} tasks, however, a small set of primers may not be
achievable.
}
\details{
The difficulty of a primer design task is evaluated by
estimating the distribution of coverage ratios per primer
by performing exact string matching with 
primers of length \code{primer.length}, which are constructed
by extracting template subsequences. Next, a beta distribution
is fitted to the estimated coverage distribution, which is
then compare to reference distributions representing
primer design problems of different difficulties via the
total variance distance. The difficulty of the input primer design
problem is found by selecting the class of the 
reference distributions that has the smallest distance
to the estimated coverage distribution.
An estimate of the required number of primers to reach a given
\code{required.cvg} can be computed by setting
\code{primer.estimate} to \code{TRUE}. Since this estimate
is based solely on perfect matching primers, the number of
primers that would actually be required is typically less.
}
\examples{
data(Ippolito)
design.estimate <- classify_design_problem(template.df)
# Estimate the number of required primers
design.estimate.nbr <- classify_design_problem(template.df, mode.directionality = "fw",
                         primer.length = 20, primer.estimate = TRUE)
}
