% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{combine.binding.events}
\alias{combine.binding.events}
\title{Combination of Binding Events.}
\usage{
combine.binding.events(my.binding.fw, my.binding.rev, fw.m, rev.m)
}
\arguments{
\item{my.binding.fw}{Forward binding events of individual primers.}

\item{my.binding.rev}{Reverse binding events of individual primers.}

\item{fw.m}{Forward binding events of paired primers.}

\item{rev.m}{Reverse binding events of paired primers.}
}
\value{
IRanges of all binding events.
}
\description{
Appends all binding events.
}
\keyword{internal}
