% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{combine.strings}
\alias{combine.strings}
\title{Combination of OligoArrayAux Structure Sequences.}
\usage{
combine.strings(s1, s2)
}
\arguments{
\item{s1}{A character vector to be combined with \code{s2}.}

\item{s2}{A character vector to be included into \code{s1}.}
}
\value{
A character vector.
}
\description{
Combines the input strings.
}
\keyword{internal}
