% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguity.R
\name{comp}
\alias{comp}
\title{Sequence complement}
\usage{
comp(seq, forceToLower = TRUE, ambiguous = FALSE)
}
\arguments{
\item{seq}{Input char vector.}

\item{forceToLower}{if TRUE the input is transformed to lower case.}

\item{ambiguous}{if TRUE ambiguous IUPAC nucleotides are complemented.}
}
\value{
The complement of \code{seq}.
}
\description{
Complements the input sequence (re-write of seqinr comp function for gap support)
}
\keyword{internal}
