% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{compare.constraints}
\alias{compare.constraints}
\title{Constraint list comparison}
\usage{
compare.constraints(A, B)
}
\arguments{
\item{A}{First constraint list.}

\item{B}{Second constraint list.}
}
\value{
TRUE if the constraints are identical, FALSE else.
}
\description{
Determines whether two list with constraints are identical.
}
\keyword{internal}
