% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_comparison.R
\name{comparison.cvg}
\alias{comparison.cvg}
\title{Comparison Coverage Stats.}
\usage{
comparison.cvg(primer.data, template.data)
}
\arguments{
\item{primer.data}{List with primer data frames.}

\item{template.data}{List with template data frames.}
}
\value{
Coverage statistics for comparing primers.
}
\description{
Computes coverage stats for primer comparison.
}
\keyword{internal}
