% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparison.R
\docType{data}
\name{Comparison}
\alias{Comparison}
\alias{primer.data}
\alias{template.data}
\title{Evaluated Primer Data for Comparison.}
\format{\code{primer.data} and \code{template.data} are
lists of \code{Primers} and \code{Templates} objects, respectively.}
\usage{
data(Comparison)
}
\description{
Evaluated primer sets targeting the functional human IGH immunoglobulin
genes. The sets were generated using the default evaluation settings
of openPrimeR. The primer sets were gathered from
IMGT and the literature.
}
\examples{
# Load 'primer.data' and 'template.data'
data(Comparison)
# Explore the first entry of the primer and template data:
primer.data[[1]]
template.data[[1]]
# Summarize the primer properties:
get_comparison_table(template.data, primer.data)
}
\references{
IMGT®, the international ImMunoGeneTics information system® http://www.imgt.org (founder and director: Marie-Paule Lefranc, Montpellier, France).
}
\keyword{datasets}
