% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_comparison.R
\name{comparison.stats.raw}
\alias{comparison.stats.raw}
\title{Computation of Raw Stats for Primer Comparison}
\usage{
comparison.stats.raw(primer.data, template.data)
}
\arguments{
\item{primer.data}{List with primer data frames.}

\item{template.data}{List with template data frames.}
}
\value{
Raw statistics for primer comparison.
}
\description{
Computes raw stats for primer comparison.
}
\keyword{internal}
