% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguity.R
\name{complement.sequence}
\alias{complement.sequence}
\title{Sequence complement}
\usage{
complement.sequence(seq)
}
\arguments{
\item{seq}{The input sequence strings.}
}
\value{
The complements of the input sequences.
}
\description{
Computes the complement of the input sequence.
}
\keyword{internal}
