% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{compute.all.cross.dimers.frontend}
\alias{compute.all.cross.dimers.frontend}
\title{Cross Dimerization.}
\usage{
compute.all.cross.dimers.frontend(primer.df, primer_conc, na_salt_conc,
  mg_salt_conc, k_salt_conc, tris_salt_conc, annealing.temp,
  for.shiny = FALSE, no.structures = FALSE)
}
\arguments{
\item{primer.df}{Input primer data frame.}

\item{primer_conc}{Primer concentration.}

\item{na_salt_conc}{Sodium ion concentration.}

\item{mg_salt_conc}{Magensium ion concentration.}

\item{k_salt_conc}{Potassium ion concentration.}

\item{tris_salt_conc}{Tris ion concentration.}

\item{annealing.temp}{The PCR annealing temperature.}

\item{for.shiny}{Whether to format the table for HTML output.}

\item{no.structures}{Whether to compute structures of dimers.}
}
\value{
A formatted data frame with cross-dimerization infos
}
\description{
Computes all cross dimers in a user-formatted way.
}
\keyword{internal}
