% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{compute.all.cross.dimers.unfiltered}
\alias{compute.all.cross.dimers.unfiltered}
\title{Cross dimerization}
\usage{
compute.all.cross.dimers.unfiltered(primer.df, primer_conc, na_salt_conc,
  mg_salt_conc, k_salt_conc, tris_salt_conc, annealing.temp, check.idx = NULL,
  for.shiny = FALSE, no.structures = FALSE)
}
\arguments{
\item{primer.df}{Input primers data frame.}

\item{primer_conc}{Primer concentration.}

\item{na_salt_conc}{Sodium ion concentration.}

\item{mg_salt_conc}{Magensium ion concentration.}

\item{k_salt_conc}{Potassium ion concentration.}

\item{tris_salt_conc}{Tris ion concentration.}

\item{annealing.temp}{The PCR annealing temperature.}

\item{check.idx}{Indices of primers for checking cross-dimerization.}

\item{for.shiny}{Whether to format for HTML output.}

\item{no.structures}{Whether dimer structures shall not be determined.
If \code{TRUE}, structures are not computed resulting in faster runtimes.}
}
\value{
All cross dimers.
}
\description{
Compute DeltaG data frame for possible primer cross-dimers.
}
\keyword{internal}
