% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_algo.R
\name{compute.all.primer.subsets.ILP}
\alias{compute.all.primer.subsets.ILP}
\title{Computation of Primer Subsets}
\usage{
compute.all.primer.subsets.ILP(primer.df, template.df, k, groups,
  cur.results.loc, required.cvg = 1)
}
\arguments{
\item{primer.df}{Primer data frame.}

\item{template.df}{Template data frame.}

\item{k}{Subset size-increment.}

\item{groups}{Identifiers of template groups in order to limit coverage to
certain groups of template sequences.}

\item{cur.results.loc}{Location for storing the results.}

\item{required.cvg}{The required coverage ratio.}
}
\value{
Write-out of results.
}
\description{
Computes all optimal primer subsets and stores their plots.
}
\keyword{internal}
