% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{compute.all.self.dimers}
\alias{compute.all.self.dimers}
\title{Self dimers}
\usage{
compute.all.self.dimers(primer.df, primer_conc, na_salt_conc, mg_salt_conc,
  k_salt_conc, tris_salt_conc, annealing.temp, for.shiny = FALSE,
  no.structures = FALSE)
}
\arguments{
\item{primer.df}{Input primer data frame.}

\item{primer_conc}{Primer concentration.}

\item{na_salt_conc}{Sodium ion concentration.}

\item{mg_salt_conc}{Magensium ion concentration.}

\item{k_salt_conc}{Potassium ion concentration.}

\item{tris_salt_conc}{Tris ion concentration.}

\item{annealing.temp}{The PCR annealing temperature.}

\item{for.shiny}{Whether the output is to be formatted for HTML.}

\item{no.structures}{Whether dimerization structures shall be outputted.}
}
\value{
Data frame with thermodynamic information on all self dimers.
}
\description{
Computes all possible self dimers for the primers in the input data frame.
}
\keyword{internal}
