% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_annealing_temperature.R
\name{compute.covered.Ta}
\alias{compute.covered.Ta}
\title{Annealing temperature}
\usage{
compute.covered.Ta(primer.df, mode.directionality = c("fw", "rev", "both"),
  template.df, na_salt_conc, mg_salt_conc, k_salt_conc, tris_salt_conc,
  primer_conc)
}
\arguments{
\item{primer.df}{Primer data frame.}

\item{mode.directionality}{Primer directionality.}

\item{template.df}{Template data frame}

\item{na_salt_conc}{Sodium ion concentration.}

\item{mg_salt_conc}{Magensium ion concentration.}

\item{k_salt_conc}{Potassium ion concentration.}

\item{tris_salt_conc}{Tris ion concentration.}

\item{primer_conc}{Primer concentration.}
}
\value{
The recommended annealing temperature.
}
\description{
Computes the annealing temperature using all binding events.
}
\keyword{internal}
