% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{compute.dimer.matrix}
\alias{compute.dimer.matrix}
\title{Dimerization matrix}
\usage{
compute.dimer.matrix(G, deltaG.cutoff = -7)
}
\arguments{
\item{G}{Matrix with free energies of all considered primer pairs.}

\item{deltaG.cutoff}{Primers with free energies below the cutoff are considered dimerizing.}
}
\value{
Binary matrix with dimerization events according to the \code{deltaG.cutoff}.
Contains '1' if primers (i,j) dimerize and '0' else.
}
\description{
Computes a matrix indicating all dimerizing primers according to a DeltaG cutoff.
}
\keyword{internal}
