% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_melting_temperature.R
\name{compute.empiric.melting.temp}
\alias{compute.empiric.melting.temp}
\title{Non-Thermodynamic Computation of Melting Temperatures.}
\usage{
compute.empiric.melting.temp(primer.df)
}
\arguments{
\item{primer.df}{A \code{Primers} object.}
}
\value{
A data frame with melting temperature information for the primers.
}
\description{
Computes the melting temperature of primers from an empiric formula.
}
\keyword{internal}
