% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_gc_ratio.R
\name{compute.gc.ratio}
\alias{compute.gc.ratio}
\title{GC ratio}
\usage{
compute.gc.ratio(x)
}
\arguments{
\item{x}{Input sequence.}
}
\value{
The fraction of G/Cs in \code{x}.
}
\description{
Computes the ratio of G/Cs in a sequence.
}
\details{
In case of ambiguities, the mean GC ratio of all possible sequences is computed.
}
\keyword{internal}
