% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_melting_temperature.R
\name{compute.melting.temps.thermo}
\alias{compute.melting.temps.thermo}
\title{Computation of Thermodynamic Melting Temperatures.}
\usage{
compute.melting.temps.thermo(primer.df, primer_conc, na_salt_conc, mg_salt_conc,
  k_salt_conc, tris_salt_conc, mode.directionality = c("fw", "rev", "both"))
}
\arguments{
\item{primer.df}{Primer data frame.}

\item{primer_conc}{Primer concentration.}

\item{na_salt_conc}{Sodium ion concentration.}

\item{mg_salt_conc}{Magensium ion concentration.}

\item{k_salt_conc}{Potassium ion concentration.}

\item{tris_salt_conc}{Tris ion concentration.}

\item{mode.directionality}{Direction of primers}
}
\value{
Data frame with melting temperature info for the input primers.
}
\description{
Use nearest-neighbor thermodynamic computations to find melting temperatures.
}
\keyword{internal}
