% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stop_codons.R
\name{compute.mismatch.table}
\alias{compute.mismatch.table}
\title{Mismatch overview table}
\usage{
compute.mismatch.table(primer.df, template.df, mode.directionality = c("fw",
  "rev"))
}
\arguments{
\item{primer.df}{Primer data frame.}

\item{template.df}{Template data.}

\item{mode.directionality}{Direction of primers.}
}
\value{
: A data frame summarizing all mismatches of the input primers with the input templates.
}
\description{
Computes a table summarizing all of the mismatches caused by the primers in the
input data frame.
}
\keyword{internal}
