% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_secondary_structures.R
\name{compute.secondary.structures}
\alias{compute.secondary.structures}
\title{Secondary Structure Computations.}
\usage{
compute.secondary.structures(primer.df, mode.directionality = c("fw", "rev",
  "both"), annealing.temperature)
}
\arguments{
\item{primer.df}{Primer data frame.}

\item{mode.directionality}{Direction of primers.}

\item{annealing.temperature}{Temperatures at which to compute secondary structures for every primer}
}
\value{
Data frame with secondary structure information.
}
\description{
Computes the secondary structures of the input primers using ViennaRNA.
}
\references{
Lorenz, Ronny and Bernhart, Stephan H. and Höner zu Siederdissen, 
Christian and Tafer, Hakim and Flamm, Christoph and Stadler, Peter F. and Hofacker, Ivo L.
ViennaRNA Package 2.0
Algorithms for Molecular Biology, 6:1 26, 2011, doi:10.1186/1748-7188-6-26
}
\keyword{internal}
