% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_secondary_structures.R
\name{compute.structure.vienna}
\alias{compute.structure.vienna}
\title{Computation of Secondary Structures with ViennaRNA.}
\usage{
compute.structure.vienna(seqs, annealing.temperature,
  folding.constraints = NULL, id = "")
}
\arguments{
\item{seqs}{The input sequences for which structures shall be computed.}

\item{annealing.temperature}{The temperature in degree Celsius 
at which to compute secondary structures.}

\item{folding.constraints}{Character vector specifying the folding
conditions for every input sequence.
For example the constraint \code{xxxxxx...} would forbid 
folding in the first 6 bases and allow folding in the last 3 bases.}

\item{id}{An identifier for storing the files}

\item{constraints}{If provide}
}
\value{
A data frame with secondary structures.
}
\description{
Computes secondary structures using ViennaRNA.
}
\keyword{internal}
