% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_template_secondary_structures.R
\name{compute.template.secondary.structures}
\alias{compute.template.secondary.structures}
\title{Template Secondary Structures}
\usage{
compute.template.secondary.structures(template.df, annealing.temperature,
  regions = NULL, constraints = NULL)
}
\arguments{
\item{template.df}{Template data frame.}

\item{annealing.temperature}{Temperature [C] at which 
to compute secondary structures.}

\item{regions}{List containing the positional intervals for which the template
secondary structure should be computed.}

\item{constraints}{String giving secondary structure constraints.
For example \code{xxxxxx...} would forbid folding in the first 6 bases
of a template with length 9 and allow folding in its last 3 bases.}
}
\value{
Data frame with info on template secondary structures.
}
\description{
Computes template secondary structures.
}
\keyword{internal}
