% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_melting_temperature.R
\name{compute.Tm.baldino}
\alias{compute.Tm.baldino}
\title{Baldino Formula}
\usage{
compute.Tm.baldino(sequences, na_salt_conc, mg_salt_conc, k_salt_conc,
  tris_salt_conc, primer_conc)
}
\arguments{
\item{sequences}{Input sequence strings.}

\item{na_salt_conc}{Sodium ion concentration.}

\item{mg_salt_conc}{Magensium ion concentration.}

\item{k_salt_conc}{Potassium ion concentration.}

\item{tris_salt_conc}{Tris ion concentration.}

\item{primer_conc}{Primer concentration.}
}
\value{
The melting temperature for the input sequences.
}
\description{
Computes the melting temperature using the formulation by Baldino.
}
\references{
Rychlik, W. J. S. W., W. J. Spencer, and R. E. Rhoads. "Optimization of the annealing temperature for DNA amplification in vitro." Nucleic acids research 18.21 (1990): 6409-6412.
}
\keyword{internal}
