% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{compute.unique.covered.idx}
\alias{compute.unique.covered.idx}
\title{Unique Coverage Indices
Computes the indices of templates that are covered uniquely 
covered by an individual primer.}
\usage{
compute.unique.covered.idx(primer.df, template.df)
}
\arguments{
\item{primer.df}{Primer data frame.}

\item{template.df}{Template data frame.}
}
\value{
Index of templates uniquely covered per input primer.
}
\description{
Unique Coverage Indices
Computes the indices of templates that are covered uniquely 
covered by an individual primer.
}
\keyword{internal}
