% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{con_select}
\alias{con_select}
\title{Quick Selection of Constraints.}
\usage{
con_select(active.constraints)
}
\arguments{
\item{active.constraints}{Identifiers of constraints.}
}
\value{
The identifiers of constraints that can be computed.
}
\description{
Select constraints that can be used according to third-party tools quickly.
}
\keyword{internal}
