% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{condition}
\alias{condition}
\title{Condition Constructor}
\usage{
condition(subclass, message, call = sys.call(-1), ...)
}
\arguments{
\item{subclass}{String giving the specific error.}

\item{message}{String giving the user message.}

\item{call}{Environment object.}

\item{...}{Other arguments for the output structure.}
}
\value{
A condition structure.
}
\description{
Constructs a condition for custom errors.
}
\keyword{internal}
