% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{methods}
\name{conOptions}
\alias{conOptions}
\alias{conOptions,DesignSettings-method}
\alias{conOptions<-}
\alias{conOptions<-,DesignSettings-method}
\alias{conOptions,DesignSettings-method}
\title{Getter/Setter for Constraint Options.}
\usage{
conOptions(x)

\S4method{conOptions}{DesignSettings}(x)

conOptions(x) <- value

\S4method{conOptions}{DesignSettings}(x) <- value
}
\arguments{
\item{x}{A \code{DesignSettings} object.}

\item{value}{A list with constraint options. The permissible
fields of the list and their types are documented in the 
\code{\link{ConstraintOptions}} class.}
}
\description{
Gets the constraint settings of the provided 
\code{DesignSettings} object \code{x}.

Sets the constraint settings of the provided
\code{DesignSettings} object \code{x}.
}
\examples{
# Load some settings
data(Ippolito)
# View the active constraint options
conOptions(settings)
# Prevent mismatch binding events
conOptions(settings)$allowed_mismatches <- 0
# View available constraint options
settings
}
\seealso{
Other settings functions: \code{\link{ConstraintOptions-class}},
  \code{\link{ConstraintSettings-class}},
  \code{\link{CoverageConstraints-class}},
  \code{\link{DesignSettings-class}},
  \code{\link{PCR_Conditions-class}}, \code{\link{PCR}},
  \code{\link{constraintLimits}},
  \code{\link{constraints}}, \code{\link{cvg_constraints}},
  \code{\link{read_settings}}
}
\keyword{Settings}
