% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_gc_clamp.R
\name{consecutive.GC.count}
\alias{consecutive.GC.count}
\title{Consecutive GCs}
\usage{
consecutive.GC.count(y, len)
}
\arguments{
\item{y}{Positions where G/C occurs.
Positions are numbered from 1 to 5 where 5 is the end of the primer.}

\item{len}{Is the number of bases from the primer end considered.}
}
\value{
The maximal number of consecutive G/Cs.
}
\description{
Determines the maximum number of consecutive G/Cs
}
\keyword{internal}
