% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{methods}
\name{constraintLimits}
\alias{constraintLimits}
\alias{constraintLimits,DesignSettings-method}
\alias{constraintLimits<-}
\alias{constraintLimits<-,DesignSettings-method}
\alias{constraintLimits,DesignSettings-method}
\title{Getter/Setter for Constraint Limits.}
\usage{
constraintLimits(x)

\S4method{constraintLimits}{DesignSettings}(x)

constraintLimits(x) <- value

\S4method{constraintLimits}{DesignSettings}(x) <- value
}
\arguments{
\item{x}{A \code{DesignSettings} object whose constraint limits
are to be modified.}

\item{value}{A list with constraint boundaries. The permissible fields 
of the list are provided in \code{\link{ConstraintSettings}}.}
}
\description{
Gets the constraint limits that are defined in the provided 
\code{DesignSettings} object \code{x}.

Sets the constraint limits of the provided 
\code{DesignSettings} object \code{x}.
}
\examples{
# Load some settings
data(Ippolito)
# View the active constraint limits
constraintLimits(settings)
# Extend the GC relaxation limit
constraintLimits(settings)$gc_clamp <- c("min" = 0, "max" = 6)
# View available constraints
settings
}
\seealso{
Other settings functions: \code{\link{ConstraintOptions-class}},
  \code{\link{ConstraintSettings-class}},
  \code{\link{CoverageConstraints-class}},
  \code{\link{DesignSettings-class}},
  \code{\link{PCR_Conditions-class}}, \code{\link{PCR}},
  \code{\link{conOptions}}, \code{\link{constraints}},
  \code{\link{cvg_constraints}},
  \code{\link{read_settings}}
}
\keyword{Settings}
