% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{methods}
\name{constraints}
\alias{constraints}
\alias{constraints,DesignSettings-method}
\alias{constraints,AbstractConstraintSettings-method}
\alias{constraints,DesignSettings-method}
\alias{constraints<-}
\alias{constraints<-,DesignSettings,ANY-method}
\alias{constraints,DesignSettings-method}
\alias{constraints<-,AbstractConstraintSettings,list-method}
\alias{constraints,DesignSettings-method}
\title{Getter/Setter for Constraints.}
\usage{
constraints(x)

\S4method{constraints}{DesignSettings}(x)

\S4method{constraints}{AbstractConstraintSettings}(x)

constraints(x) <- value

\S4method{constraints}{DesignSettings,ANY}(x) <- value

\S4method{constraints}{AbstractConstraintSettings,list}(x) <- value
}
\arguments{
\item{x}{A \code{DesignSettings} object.}

\item{value}{A list with constraint settings. Each list entry
should have a permissible name and consist of at most two
values providing the minimal and/or maximal allowed values, which
have to be denominated via \code{min} and \code{max}.}
}
\description{
Gets the active constraints of the provided
\code{DesignSettings} object \code{x}.

Sets the active constraints of the provided
\code{DesignSettings} object \code{x}.
}
\details{
For an overview of permissible constraints,
please consider the \code{\link{ConstraintSettings}} documentation.
}
\examples{
# Load some settings
data(Ippolito)
# View the active constraints
constraints(settings)
# Require a minimal GC clamp extent of 0
constraints(settings)$gc_clamp["min"] <- 0
# View available constraints
settings
}
\seealso{
Other settings functions: \code{\link{ConstraintOptions-class}},
  \code{\link{ConstraintSettings-class}},
  \code{\link{CoverageConstraints-class}},
  \code{\link{DesignSettings-class}},
  \code{\link{PCR_Conditions-class}}, \code{\link{PCR}},
  \code{\link{conOptions}}, \code{\link{constraintLimits}},
  \code{\link{cvg_constraints}},
  \code{\link{read_settings}}
}
\keyword{Settings}
