% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{constraints.to.df}
\alias{constraints.to.df}
\title{Conversion of Constraints List to Data Frame.}
\usage{
constraints.to.df(limit.constraints, out.names, format.type = c("backend",
  "shiny", "report"))
}
\arguments{
\item{limit.constraints}{A list with constraints.}

\item{out.names}{The desired column names.}

\item{format.type}{The type of formatting to be performed on the table}
}
\value{
A data frame giving an overview of the constraints.
}
\description{
Converts the input constraints to a data frame representation.
}
\keyword{internal}
