% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{constraints.xml.format}
\alias{constraints.xml.format}
\title{Constraint XML Format.}
\usage{
constraints.xml.format(constraints, set.name)
}
\arguments{
\item{constraints}{List with constraint settings.}

\item{set.name}{Identifier for the constraint settings.}
}
\value{
XML string containing the constraint settings.
}
\description{
Format constraint settings for XML output.
}
\keyword{internal}
