% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{constraints_to_unit}
\alias{constraints_to_unit}
\title{Mapping of Constraints to Units.}
\usage{
constraints_to_unit(constraint, use.unit = TRUE, format.type = c("backend",
  "HTML", "report"))
}
\arguments{
\item{constraint}{The names of the constraints to convert to their plot identifiers (units).}

\item{use.unit}{Whether constraint names should be annotated with their units.}

\item{use.HTML}{Whether constraint units should be annotated with HTML units.}
}
\value{
A list of constraint names.
}
\description{
Maps constraints to units for plotting.
}
\keyword{internal}
