% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\docType{class}
\name{ConstraintSettings-class}
\alias{ConstraintSettings-class}
\alias{ConstraintSettings}
\title{Class for Constraint Settings.}
\value{
A \code{ConstraintSettings} object.
}
\description{
The \code{ConstraintSettings} class encapsulates the constraints
on the physicochemical properties of primers.
}
\details{
You can check whether the constraint settings are fulfilled using
\code{\link{check_constraints}} or filter accordingly using
\code{\link{filter_primers}}.
}
\section{Slots}{

\describe{
\item{\code{status}}{Named boolean vector indicating 
which of the possible constraints are active (\code{TRUE})
and which are not (\code{FALSE}).}

\item{\code{settings}}{Named list containing the 
settings of the active constraints:
The list may contain the following fields:
\describe{
\item{primer_coverage:}{The required number of covered template sequences per primer.}
\item{primer_specificity:}{The required required specificity of 
                           primers in terms of a ratio in the interval [0,1].}
\item{primer_length:}{The required lengths of primer sequences.}
\item{gc_clamp:}{The desired number of GCs at primer 3' termini.}
\item{gc_ratio:}{The desired ratio of GCs in primers 
                 in terms of numbers in the interval [0,1].}
\item{no_runs:}{The accepted length homopolymer runs in a primer.}
\item{no_repeats:}{The accepted length of dinucleotide repeats in a primer.}
\item{self_dimerization:}{The lowest acceptable free energy [kcal/mol] for the 
                          interaction of a primer with itself. The identification
                           of self dimers requires the software \emph{OligoArrayAux} (see notes).}
\item{melting_temp_range:}{The desired melting temperature (Celsius) of primers.
                          The accurate computation of melting temperatures requires the software \emph{MELTING} (see notes).}
\item{melting_temp_diff:}{The maximal allowed difference between the melting temperatures (Celsius)
                          of primers contained in the same set. The accurate computation of melting temperatures
                          requires the software \emph{MELTING} (see notes).}
\item{cross_dimerization:}{The lowest acceptable free energy [kcal/mol] for the
                           interaction of a primer with another primer. The identification
                           of cross dimers requires the software \emph{OligoArrayAux} (see notes).}
\item{secondary_structure:}{The lowest acceptable free energy [kcal/mol] for the
                            formation of primer secondary structures. Secondary structures are determined
                            using the software \emph{ViennaRNA} (see notes).}
}}
}}

\note{
External programs are required for computing the following constraints:
\describe{
\item{MELTING (http://www.ebi.ac.uk/biomodels/tools/melting/):}{Thermodynamic computations (optional) for determining melting temperatures for the constraints \code{melting_temp_diff} and \code{melting_temp_range}}
\item{OligoArrayAux (http://unafold.rna.albany.edu/OligoArrayAux.php):}{Thermodynamic computations used for computing \code{self_dimerization} and \code{cross_dimerization}.
Also required for computing \code{primer_coverage} when a constraint based on the free energy of annealing is active.}
\item{ViennaRNA (http://www.tbi.univie.ac.at/RNA/):}{Secondary structure predictions used for the constraint \code{secondary_structure}}
}
}
\section{primer_coverage}{

Computing the primer coverage involves identifying
which templates are expected to be amplified (covered) by which primers.
The \code{primer_coverage} constraint
determines the minimal and maximal number of coverage events per primer
that are required. The computation of primer coverage is governed
by the coverage constraints postulated via \code{\link{CoverageConstraints}}
and the constraint options defined via \code{\link{ConstraintOptions}}.
}

\section{primer_specificity}{

Primer specificity is automatically determined during the primer coverage
computations but the constraint is only checked when the \code{primer_specificity}
field is available. The specificity of a primer is defined as its ratio of
on-target vs total coverage events (including off-target coverage). Low-specificity
primers should be excluded as they may not amplify the target region effectively.
}

\section{primer_length}{

The length of a primer is defined by its number of bases. Typical
primers have lengths between 18 and 22. Longer primers may guarantee higher
specificities.
}

\section{gc_clamp}{

The GC clamp refers to the presence of GCs at the 3' end of a primer.
For the \code{gc_clamp} constraint, we consider the number of 3' terminal GCs.
For example, the primer \emph{actgaaatttcaccg} has a GC clamp of length 3.
The presence of a GC clamp is supposed to aid the stability of the
polymerase complex. At the same time, long GC clamps should be avoided.
}

\section{no_runs}{

Homopolymer runs (e.g. the primer \emph{aaaaa} has a run of 5 A's)
may lead to secondary structure formation and unspecific binding
and should therefore be avoided.
}

\section{no_repeats}{

Dinucleotide repeats (e.g. the primer \emph{tatata} has 3 TA repeats)
should be avoided for the same reason a long homopolymer runs.
}

\section{self_dimerization}{

Self dimerization refers to a primer that binds to itself rather than
to one of the templates. Primers exhibiting self dimers should
be avoided as they may prevent the primer from amplifying the templates.
Therefore primers with small free energies of dimerization should 
be avoided.
}

\section{melting_temp_range}{

The melting temperature is the temperature at which 50% of the primers
are in duplex with templates and 50% are still present as random coils.
Hence, primers exhibiting high melting temperatures have high affinities
to the templates, while primers with small melting temperatures
have small affinities. The melting temperatures of the primers
determine the annealing temperature of the PCR, which is why
the melting temperatures of the primers should not deviate too much (see
\code{melting_temp_diff}).
}

\section{melting_temp_diff}{

The differences between the melting temperatures of primers in a
set of primers should not deviate too much as the annealing temperaturte
of a PCR should be based on the smallest melting temperature
of a primer in the set. If there are other primers in the set 
exhibiting considerably higher melting temperatures,
these primers may bind inspecifically due to the low annealing temperature.
}

\section{cross_dimerization}{

When two different primers bind to each each other rather than
to the templates, this is called cross dimerization.
Cross dimerization should be prevent at all costs because
such primers cannot effectively amplify their target templates.
Cross dimerizing primers can be excluding primers
exhibiting small free energies of cross dimerization.
}

\section{secondary_structure}{

When a primer exhibits secondary structure, this may prevent it from
binding to the templates. To prevent this, 
primers with low free energies of secondary structure formation
can be excluded.
}

\examples{
# Initializing a new 'ConstraintSettings' object:
constraint.settings <- new("ConstraintSettings")
# Retrieving the constraint settings from a 'DesignSettings' object:
data(Ippolito) # loads a 'DesignSettings' object into 'settings'
constraints(settings)
# Modifying the constraint settings:
constraints(settings)$no_runs["max"] <- 10
constraints(settings) <- constraints(settings)[names(constraints(settings)) != "gc_clamp"]
}
\seealso{
Other settings functions: \code{\link{ConstraintOptions-class}},
  \code{\link{CoverageConstraints-class}},
  \code{\link{DesignSettings-class}},
  \code{\link{PCR_Conditions-class}}, \code{\link{PCR}},
  \code{\link{conOptions}}, \code{\link{constraintLimits}},
  \code{\link{constraints}}, \code{\link{cvg_constraints}},
  \code{\link{read_settings}}
}
\keyword{Classes}
