% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguity.R
\name{convert.from.iupac}
\alias{convert.from.iupac}
\title{Conversion from IUPAC nucleotides}
\usage{
convert.from.iupac(seqs)
}
\arguments{
\item{seqs}{A vector of strings.}
}
\value{
lists containing the disambiguated input sequences
}
\description{
Convert sequences with IUPAC ambiguity codes to all possible sequences without ambiguities.
}
\keyword{internal}
