% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{convert.PCR.units}
\alias{convert.PCR.units}
\title{Conversion of PCR Units}
\usage{
convert.PCR.units(pcr.settings, to.mol = TRUE)
}
\arguments{
\item{pcr.settings}{List with several PCR settings (concentrations).}

\item{to.mol}{If \code{TRUE}, convert to the molar concentration.
If \code{FALSE} convert to the unit representation in the XML.}
}
\value{
List with concentrations for usage in the backend.
}
\description{
Converts frontend PCR concentration units to the units used for the backend.
}
\keyword{internal}
