% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_melting_temperature.R
\name{convert.temperature}
\alias{convert.temperature}
\title{Conversion between Celsius and Kelvin}
\usage{
convert.temperature(temp, temp.scale = c("K", "C"))
}
\arguments{
\item{temp}{The input temperature.}

\item{temp.scale}{The desired unit of the output temperature.}
}
\value{
Transforms the input temperature to the specified \code{temp.scale}.
}
\description{
Converts the input from Kelvin to Celsius or from Celsius to Kelvin.
}
\details{
If \code{temp.scale} is 'K', \code{T_m} is transformed from Celsius
to Kelvin. If \code{temp.scale} is 'C', \code{T_m} is transformed from
Kelvin to Celsius. The default is to transform from Celsius to Kelvin.
}
\keyword{internal}
