% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguity.R
\name{convert.to.iupac}
\alias{convert.to.iupac}
\title{Merge sequences.}
\usage{
convert.to.iupac(seqs)
}
\arguments{
\item{seqs}{Vector of strings.}
}
\value{
Consensus sequence of seqs.
}
\description{
Merges the input sequences to one sequence containing IUPAC ambiguity codes.
}
\keyword{internal}
