% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{copy.melt.config}
\alias{copy.melt.config}
\title{Copy MELTING Config File}
\usage{
copy.melt.config(melt.bin = NULL)
}
\value{
TRUE if the file is available in the MELTING folder, FALSE otherwise.
}
\description{
Copies modified MELTING tandem mismatch file to the MELTING data folder.
}
\keyword{internal}
