% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\docType{class}
\name{CoverageConstraints-class}
\alias{CoverageConstraints-class}
\alias{CoverageConstraints}
\title{Class for Coverage Constraints.}
\value{
A \code{CoverageConstraints} object.
}
\description{
The \code{CoverageConstraints} class encapsulates the conditions
under which the coverage is evaluated.
}
\section{Slots}{

\describe{
\item{\code{status}}{Named boolean vector indicating 
which of the possible options are active (\code{TRUE})
and which are not (\code{FALSE}).}

\item{\code{settings}}{Named list with constraint options. Each 
list entry should have an entry \code{min} and/or \code{max}
in order to indicate the minimal and maximal allowed values,
respectively. 
The following identifiers can be used as coverage constraints:
\describe{
\item{\code{primer_efficiency}:}{The desired efficiencies of primer-template amplification events 
in order to be considered as \emph{covered}. \code{primer_efficiency} provides a value in the interval [0,1],
which is based on \pkg{DECIPHER}'s thermodynamic model, which considers the impact of 3' terminal mismatches.}
\item{\code{annealing_DeltaG}:}{The desired free energies of annealing for putative
coverage events between primers and templates. Typically, one would 
limit the maximally allowed free energy.}
\item{\code{stop_codon}:}{Whether coverage events introducing
stop codons into the amplicons should be allowed or discarded. 
Here, a value of \code{1} indicates coverage events that induce stop codons.
As such, setting both minimum and maximium to zero will disregard
coverage events inducing stop codons, while setting the minimum to zero
and the maximum to 1 will allow coverage events that induce stop codons.}
\item{\code{substitution}:}{Whether coverage events introducing
substitutions into the amino acid sequence are considered or discarded.
The same encoding as for \code{stop_codon} is used, that is,
the value \code{1} indicates coverage events
inducing substitutions. Hence, to prevent substitutions,
the maximal value of \code{substitution} can be set to zero.}
\item{\code{terminal_mismatch_pos}:}{The position relative to 
the primer 3' terminal end for which mismatch binding events should be allowed,
where the last base in a primer is indicated by position \code{1}.
For example, setting the minimal value of \code{terminal_mismatch_pos} 
to \code{7} means that only coverage events that do not have a terminal mismatch
within the last 6 bases of the primer are allowed.}
\item{\code{coverage_model}:}{Use a logistic regression model combining the free energy of annealing and 3' terminal mismatch positions
to determine the expected rate of false positive coverage calls. 
Using \code{coverage_model}, you can specify the allowed ratio of falsely predicted coverage events.
Typically, one would limit the maximal allowed rate of false positives. Note that setting a
small false positive rate will reduce the sensitivity of the coverage calls (i.e. true positives will be missed).}
}}
}}

\note{
External programs are required for computing the following constraints:
\describe{
\item{OligoArrayAux (http://unafold.rna.albany.edu/OligoArrayAux.php):}{Thermodynamic computations used for 
computing the coverage constraints \code{annealing_DeltaG}, \code{primer_efficiency}, and \code{coverage_model}}
}
}
\examples{
# Initialize a new 'CoverageConstraints' object:
cvg.constraints <- new("CoverageConstraints")
# Retrieving the coverage constraints from a 'DesignSettings' object:
data(Ippolito) # loads a 'DesignSettings' object into 'settings'
cvg_constraints(settings)
# Modifying the coverage constraints
cvg_constraints(settings)$primer_efficiency["min"] <- 0.001
}
\seealso{
Other settings functions: \code{\link{ConstraintOptions-class}},
  \code{\link{ConstraintSettings-class}},
  \code{\link{DesignSettings-class}},
  \code{\link{PCR_Conditions-class}}, \code{\link{PCR}},
  \code{\link{conOptions}}, \code{\link{constraintLimits}},
  \code{\link{constraints}}, \code{\link{cvg_constraints}},
  \code{\link{read_settings}}
}
\keyword{Classes}
