% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{covered.primers.to.ID.string}
\alias{covered.primers.to.ID.string}
\title{Conversion of Primer Indices to ID string}
\usage{
covered.primers.to.ID.string(covered.primers, primer.df)
}
\arguments{
\item{covered.primers}{Identifiers of primers covering sequences.}

\item{primer.df}{Primer data frame.}
}
\value{
String containing the covered template IDs.

A string containing the IDs of covering primers.
}
\description{
Converts the input coverage indices to a comma-separated string with the template IDs.
}
\keyword{internal}
