% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{covered.seqs.to.ID.string}
\alias{covered.seqs.to.ID.string}
\title{Conversion of Template Coverage Indices to ID string}
\usage{
covered.seqs.to.ID.string(covered.seqs, template.df)
}
\arguments{
\item{covered.seqs}{Indices of covered template sequences.}

\item{template.df}{Template data frame.}
}
\value{
A string containing the covered template IDs.
}
\description{
Converts the input coverage indices to a comma-separated string with the template IDs.
}
\keyword{internal}
