% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{covered.seqs.to.idx}
\alias{covered.seqs.to.idx}
\title{Conversion of Coverage Strings to Indices.}
\usage{
covered.seqs.to.idx(covered.seqs, template.df)
}
\arguments{
\item{covered.seqs}{Strings of covered sequences to be converted.}

\item{template.df}{Template data frame containing the identifiers of templates}
}
\value{
Indices of covered templates.
}
\description{
Converts the input coverage strings (comma separated template Identifiers) into indices.
}
\keyword{internal}
