% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{create.constraint.table}
\alias{create.constraint.table}
\title{Output a Constraint Overview Table}
\usage{
create.constraint.table(constraints, constraint.limits = NULL,
  constraints.used.fw = NULL, constraints.used.rev = NULL,
  format.type = c("backend", "shiny", "report"))
}
\arguments{
\item{constraints}{List with constraint settings.}

\item{constraint.limits}{List with constraint limits.}

\item{constraints.used.fw}{Constraints used for forward primer design.}

\item{constraints.used.rev}{Constraints used for reverse primer design.}

\item{format.type}{The type of formatting to be performed on entries.}
}
\value{
Data frame with summary of constraints.
}
\description{
Outputs a table showing the values of constraints.
}
\keyword{internal}
