% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{create.constraint.XML}
\alias{create.constraint.XML}
\title{XML Output of Constraints}
\usage{
create.constraint.XML(filtering.constraints, c.f.lim, cvg.constraints,
  PCR.settings, constraint.settings)
}
\arguments{
\item{filtering.constraints}{List with constraint settings for filtering.}

\item{c.f.lim}{Relaxation limits for the filtering constraints.}

\item{cvg.constraints}{List with constraints for coverage computations.}

\item{PCR.settings}{Settings for the PCR.}

\item{constraint.settings}{Other settings of constraints (e.g. coverage).}
}
\value{
String in XML format containing all constraint settings.
}
\description{
Creates an XML summarizing all settings.
}
\keyword{internal}
