% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{create.cvg.text}
\alias{create.cvg.text}
\title{Coverage Info Text}
\usage{
create.cvg.text(stats, selected.group = NULL, ident = NULL)
}
\arguments{
\item{stats}{Data frame with coverage statistics.}

\item{selected.group}{Retrieve information for a subgroup of templates only.}

\item{ident}{An identifier for the coverage.}
}
\value{
A string with information on the coverage.
}
\description{
Creates a string with information on the coverage
}
\keyword{internal}
