% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{create.G.matrix}
\alias{create.G.matrix}
\title{Create free energy matrix}
\usage{
create.G.matrix(primer.df, G.df, primer.df.2 = NULL)
}
\arguments{
\item{primer.df}{Primer data frame.}

\item{G.df}{Free energy data for the primers.}

\item{primer.df.2}{Optional second primer data frame}
}
\value{
Matrix with the smallest free dimerization energy for every primer pair.
}
\description{
Creates a matrix giving the deltaG values of all primers.
}
\keyword{internal}
