% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{create.k.mers}
\alias{create.k.mers}
\title{Creation of k-mers for multiple sequences.}
\usage{
create.k.mers(seqs, k)
}
\arguments{
\item{k}{The size of the k-mer.}

\item{seq}{A character vector.}
}
\value{
A list with named character vectors, containing the k-mers.
}
\description{
Creation of k-mers for multiple sequences.
}
\keyword{internal}
