% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{create.kmer}
\alias{create.kmer}
\title{Creation of k-mers of a single sequence.}
\usage{
create.kmer(seq, k)
}
\arguments{
\item{seq}{A character vector.}

\item{k}{The size of the k-mer.}
}
\value{
A names character vector, where the names 
are the relative positions of the k-mers and the values
give the character vector of the k-mer.
}
\description{
Creation of k-mers of a single sequence.
}
\keyword{internal}
