% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{create.other.table}
\alias{create.other.table}
\title{Creation of a Table for Other Constraint Settings.}
\usage{
create.other.table(other.settings, col.names, format.type)
}
\arguments{
\item{other.settings}{List with other constraint settings.}

\item{format.type}{How the table shall be formatted.}
}
\value{
A data frame.
}
\description{
Creation of a Table for Other Constraint Settings.
}
\keyword{internal}
