% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{create.PCR.table}
\alias{create.PCR.table}
\title{Creation of a Table for PCR Conditions}
\usage{
create.PCR.table(other.settings, format.type = c("backend", "shiny",
  "report"))
}
\arguments{
\item{other.settings}{List with PCR settings.}

\item{format.tyep}{How the table shall be formatted.}
}
\value{
A data frame.
}
\description{
Creation of a Table for PCR Conditions
}
\keyword{internal}
