% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_primers_tree.R
\name{create.primer.ranges}
\alias{create.primer.ranges}
\title{Ranges for Initial Primers.}
\usage{
create.primer.ranges(end.position, p.lens, start.position, step.size = 1,
  groups = NULL)
}
\arguments{
\item{end.position}{End positions of primers.}

\item{p.lens}{Desired primer lengths.}

\item{step.size}{A numeric giving the steps with which start
positions are cycled. Should be 1 for primer design (evaluate all positions)
and higher values can be used for windowing.}

\item{groups}{Character vector with group annotation.}

\item{start.posiion}{The start positions of primers.}
}
\value{
Data frame with ranges for initial primers.
}
\description{
Creates a data frame indicating primer starts and ends.
}
\keyword{internal}
