% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_algo.R
\name{create.Tm.brackets}
\alias{create.Tm.brackets}
\title{Creation of Melting Temperature Groups}
\usage{
create.Tm.brackets(primers, template.df, settings, target.temps = NULL)
}
\arguments{
\item{primers}{An object of class \code{Primers} for which to create groups based on melting temperatures.}

\item{template.df}{An object of class \code{Templates} corresponding to the primers.}

\item{settings}{A \code{DesignSettings} objects.}

\item{target.temps}{Pre-defined target melting temperatures to use instead
of automatically determining groups from the \code{primers}.}
}
\value{
Data frame with target melting temperatures for individual primer sets.
}
\description{
Creates a data frame identifying target melting temperatures of individual primer sets.
}
\keyword{internal}
