% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{create.uniform.leaders}
\alias{create.uniform.leaders}
\title{Uniform Binding Ranges.}
\usage{
create.uniform.leaders(fw.interval, rev.interval, template.df, gap.char)
}
\arguments{
\item{fw.interval}{Binding region for forward templates.}

\item{rev.interval}{Binding region for reverse templates.}

\item{template.df}{Template data frame.}

\item{gap.char}{The character for gaps in alignments.}
}
\value{
Data frame with binding region information.
}
\description{
Creates uniform binding regions for all templates.
}
\keyword{internal}
