% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{create_coverage_xls}
\alias{create_coverage_xls}
\title{Creation of a Coverage XLS Spreadsheet.}
\usage{
create_coverage_xls(primer.df, template.df, filename, settings)
}
\arguments{
\item{primer.df}{A \code{Primers} object containg primers
with evaluated coverage.}

\item{template.df}{A \code{Templates} object containing the templates
corresponding to \code{primer.df}.}

\item{filename}{A character vector providing the filename for storing the XLS file.}

\item{settings}{A \code{DesignSettings} object providing
the coverage conditions that are to be shown in the spreadsheet.}
}
\value{
Creates a spreadsheet visualizing the primer coverage
and stores it under \code{filename}.
}
\description{
Creation of an XLS spreadsheet providing an overview of the covered
template sequences for each primer. Each cell in the spreadsheet
indicates a coverage event between a primer and template using
color codes. Identified coverage events are indicated by green, while
primer-template pairs without coverage are indicated by red.
In case that a primer binding condition (see \code{\link{CoverageConstraints}})
was active when computing the coverage, the numeric value of the
coverage condition is annotated for each cell.
}
\examples{
data(Ippolito)
filename <- tempfile("cvg_overview", fileext = ".xls")
# store coverage xls file to disk
create_coverage_xls(primer.df, template.df, filename, settings)
}
