% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\docType{methods}
\name{create_report,Primers,Templates-method}
\alias{create_report,Primers,Templates-method}
\title{Creation of a PDF report.}
\usage{
\S4method{create_report}{Primers,Templates}(primers, templates, out.file,
  settings, sample.name, used.settings, required.cvg = 1)
}
\arguments{
\item{primers}{An evaluated \code{Primers} object.}

\item{templates}{A \code{Templates} object.}

\item{out.file}{A character vector giving the file to store the report in.}

\item{settings}{A \code{DesignSettings} object.}

\item{sample.name}{An identifier for your analysis.}

\item{used.settings}{A named list (with fields "fw" and "rev") containing the forward/reverse used design settings.}

\item{required.cvg}{The required coverage ratio.}
}
\value{
Creates a PDF file reporting on the input primers.
}
\description{
Creates a PDF report for a set of primers.
}
\note{
Creating the report requires the external programs Pandoc (http://pandoc.org)
and LaTeX (http://latex-project.org).
}
\keyword{internal}
