% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{create_report}
\alias{create_report}
\title{Creation of a Primer PDF Report.}
\usage{
create_report(primers, templates, out.file, settings, sample.name = NULL,
  used.settings = NULL, ...)
}
\arguments{
\item{primers}{To create a report for a single primer set, please provide
an evaluated \code{Primers} object.
For creating a report comparing multiple primer sets, please provide
a list of \code{Primers} objects.}

\item{templates}{If \code{primers} is a \code{Primers} object, \code{templates} should be a \code{Templates} object.
If \code{primers} is a list of \code{Primers} objects, \code{templates}
should be a list of \code{Templates} objects of the same length as \code{primers}.}

\item{out.file}{A character vector giving the path to the file where the report should be stored.}

\item{settings}{The \code{DesignSettings} object that was used for analyzing
the input primers.}

\item{sample.name}{An identifier for your analysis. By default ( 
\code{NULL}), the sample identifier is selected from the
\code{Run} column of the input templates.}

\item{used.settings}{A named list (with fields \code{fw} and \code{rev}) containing the relaxed settings
for designing forward/reverse primers. By default (\code{NULL}), 
the relaxed settings are not shown in the report.}

\item{...}{\code{required.cvg} (optional, default: 1), the desired coverage ratio if \code{primers} is a single primer set.}
}
\value{
Creates a PDF file summarizing the results from analyzing one or 
multiple sets of primers.
}
\description{
Creates a PDF report for analyzed primer sets.
}
\note{
Creating the report requires the external programs Pandoc (http://pandoc.org)
and LaTeX (http://latex-project.org).
}
\examples{
setting.xml <- system.file("extdata", "settings", 
                 "C_Taq_PCR_high_stringency.xml", package = "openPrimeR")
settings <- read_settings(setting.xml)
# Creation of a report for a single primer set
data(Ippolito)
out.file.single <- tempfile("evaluation_report", fileext = ".pdf")
create_report(primer.df, template.df, out.file.single, settings)
# Creation of a report for multiple primer sets.
data(Comparison)
out.file.comp <- tempfile("comparison_report", fileext = ".pdf")
create_report(primer.data[1:3], template.data[1:3], out.file.comp, settings)
}
\seealso{
Other primer functions: \code{\link{Primers-class}},
  \code{\link{check_constraints}},
  \code{\link{check_restriction_sites}},
  \code{\link{design_primers}},
  \code{\link{filter_primers}},
  \code{\link{get_initial_primers}},
  \code{\link{primer_significance}},
  \code{\link{score_degen}}, \code{\link{write_primers}}
}
\keyword{Primers}
