% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{methods}
\name{cvg_constraints}
\alias{cvg_constraints}
\alias{cvg_constraints,DesignSettings-method}
\alias{cvg_constraints<-}
\alias{cvg_constraints<-,DesignSettings-method}
\alias{cvg_constraints,DesignSettings-method}
\title{Getter/Setter for Coverage Constraints.}
\usage{
cvg_constraints(x)

\S4method{cvg_constraints}{DesignSettings}(x)

cvg_constraints(x) <- value

\S4method{cvg_constraints}{DesignSettings}(x) <- value
}
\arguments{
\item{x}{A \code{DesignSettings} object.}

\item{value}{A list with coverage constraints. Each
list entry must have a permissible name and
contain a numeric vector with at most two components
describing the minimal and/or maximal required values that
are to be indicated via \code{min} and \code{max}.
The permissible contraint identifiers are documented in the
\code{\link{CoverageConstraints}} class.}
}
\description{
Gets the coverage constraints of the provided
\code{DesignSettings} object \code{x}.

Sets the coverage constraints of the provided
\code{DesignSettings} object \code{x}.
}
\examples{
# Load some settings
data(Ippolito)
# View all active coverage constraints
cvg_constraints(settings)
# Increase the maximal false positive rate to increase the sensitiviity of coverage predictions
cvg_constraints(settings)$coverage_model <- c("max" = 0.1) 
# View available coverage constraints:
settings
}
\seealso{
Other settings functions: \code{\link{ConstraintOptions-class}},
  \code{\link{ConstraintSettings-class}},
  \code{\link{CoverageConstraints-class}},
  \code{\link{DesignSettings-class}},
  \code{\link{PCR_Conditions-class}}, \code{\link{PCR}},
  \code{\link{conOptions}}, \code{\link{constraintLimits}},
  \code{\link{constraints}}, \code{\link{read_settings}}
}
\keyword{Settings}
