% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{class}
\name{DesignSettings-class}
\alias{DesignSettings-class}
\alias{DesignSettings}
\title{Class for Primer Design Settings.}
\value{
A \code{DesignSettings} object.
}
\description{
The \code{DesignSettings} class encapsulates all settings 
for designing and evaluating primer sets. 
Upon loading an XML file, the \code{DesignSettings} class checks whether
the defined constraints can be applied by identifying whether 
the requirements for external programs are fulfilled. 
If the requirements are not fulfilled, the affected constraints 
are removed from the loaded \code{DesignSettings} object
and a warning is issued.
The loaded constraints are automatically ordered according to
the option \emph{openPrimeR.constraint_order} such that 
the runtime of the \code{\link{design_primers}} and \code{\link{filter_primers}}
functions is optimized.
}
\details{
Note that the fields \code{Input_Constraints}, \code{Input_Constraint_Boundaries}, and \code{Coverage_Constraints} should 
contain entries with at most two components using the fields \code{min} and/or \code{max}.

The \code{Input_Constraint_Boundaries} should always be at least as general as the
specified \code{Input_Constraints}.
}
\section{Slots}{

\describe{
\item{\code{Input_Constraints}}{A \code{\link{ConstraintSettings}} object specifying the 
desired target value ranges for primer properties.}

\item{\code{Input_Constraint_Boundaries}}{A \code{\link{ConstraintSettings}} object specifying
the limits for relaxing the constraints during the primer design procedure.
This slot may contain the same fields as the \code{Input_Constraints} slot,
but the specified desired ranges should be at least as general as
those specified in the \code{Input_Constraints} slot.}

\item{\code{Coverage_Constraints}}{A \code{\link{CoverageConstraints}} object specifying
the constraints for computing the primer coverage.}

\item{\code{PCR_conditions}}{A \code{\link{PCR_Conditions}} object specifying the PCR-related settings.}

\item{\code{constraint_settings}}{A \code{\link{ConstraintSettings}} object providing options 
for the computation of individual physicochemical properties.}
}}

\examples{
# Load a settings object
filename <- system.file("extdata", "settings", 
                 "C_Taq_PCR_high_stringency.xml", package = "openPrimeR")
settings <- read_settings(filename)
# Modify the constraints
constraints(settings)$gc_clamp["min"] <- 0
# Modify the constraint limits for designing primers
constraintLimits(settings)$gc_clamp["max"] <- 6
# Modify the coverage constraints
cvg_constraints(settings)$primer_efficiency["min"] <- 0.001
# Modify the PCR conditions
PCR(settings)$Na_concentration <- 0.0001
# Modify the constraint options
conOptions(settings)$allowed_mismatches <- 0
}
\seealso{
\code{\link{read_settings}} for reading settings from XML files,
\code{\link{write_settings}} for storing settings as XML files,
\code{\link{constraints}} for accessing constraints,
\code{\link{constraintLimits}} for accessing constraint boundaries,
\code{\link{cvg_constraints}} for accessing coverage constraints,
\code{\link{conOptions}} for accessing constraint options,
\code{\link{PCR}} for accessing the PCR conditions.

Other settings functions: \code{\link{ConstraintOptions-class}},
  \code{\link{ConstraintSettings-class}},
  \code{\link{CoverageConstraints-class}},
  \code{\link{PCR_Conditions-class}}, \code{\link{PCR}},
  \code{\link{conOptions}}, \code{\link{constraintLimits}},
  \code{\link{constraints}}, \code{\link{cvg_constraints}},
  \code{\link{read_settings}}
}
\keyword{Settings}
