% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{detect.gap.columns}
\alias{detect.gap.columns}
\title{Identification of Gappy Columns in Alignments.}
\usage{
detect.gap.columns(bins, gap.cutoff = 0.95, gap.char = "-")
}
\arguments{
\item{bins}{A list of \code{DNABin} alignments.}

\item{gap.cutoff}{The required percentage of gaps
for consideration as a gap column.}

\item{gap.char}{The gap character in the alignments.}
}
\value{
A list with indices giving the gap columns
for every alignment in \code{bins}.
}
\description{
Identification of Gappy Columns in Alignments.
}
\keyword{internal}
