% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\name{dimerization.table}
\alias{dimerization.table}
\title{Dimerization Table.}
\usage{
dimerization.table(dimer.data, deltaG.cutoff,
  dimer.type = c("Self-Dimerization", "Cross-Dimerization"))
}
\arguments{
\item{dimer.data}{Data frame with dimerization data.}

\item{deltaG.cutoff}{Free energy cutoff for dimerization.}

\item{dimer.type}{String identifying whether \code{dimer.data} refers to cross-dimers or self-dimers?}
}
\value{
Data frame with dimer counts.
}
\description{
Summarizes how often individual primers dimerize according to the \code{deltaG.cutoff}.
}
\keyword{internal}
