% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{dir.copy}
\alias{dir.copy}
\title{Copy Directories.}
\usage{
dir.copy(src.dir, dest.dir, overwrite)
}
\arguments{
\item{src.dir}{The directory to be copied.}

\item{dest.dir}{The target directory.}

\item{overwrite}{Overwrite existing files in \code{dest.dir}.}
}
\value{
TRUE if copying was successful, FALSE otherwise.
}
\description{
Copies a directory to another location.
}
\keyword{internal}
