% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{disambiguate.primers}
\alias{disambiguate.primers}
\title{Disambiguation of Primers.}
\usage{
disambiguate.primers(p.df)
}
\arguments{
\item{p.df}{Primer data frame.}
}
\value{
Data frame with disambiguated primers.
}
\description{
Disambiguates ambiguous primer sequences into all possible sequences.
}
\keyword{internal}
