% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{estimate.cvg}
\alias{estimate.cvg}
\title{Estimation of Primer Coverage.}
\usage{
estimate.cvg(lex.df, k = 18, mode.directionality, sample = "")
}
\arguments{
\item{k}{A numeric giving the size of the primers.}

\item{mode.directionality}{Estimation of coverage for forward/reverse/both?}

\item{sample}{An optional identifier for the sample.}

\item{seqs}{A character vector of sequences to evaluate coverage for.}
}
\value{
A list with entries \code{fw} and \code{rev} giving
data frames for forward/reverse binding.
}
\description{
Estimates the possible coverage of primers using
probes of size \code{k} and only considering perfect
matches without consideration of ambiguities.
}
\keyword{internal}
