% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_eval.R
\name{eval.comparison.primers}
\alias{eval.comparison.primers}
\title{Evaluation of Primers for Comparison}
\usage{
eval.comparison.primers(primer.data, constraint.settings)
}
\arguments{
\item{primer.data}{List with primer data frames.}

\item{constraint.settings}{List with constraint.settings.}
}
\value{
List with evaluated primer data frames.
}
\description{
Evaluate multiple primer sets according to the input constraint settings.
}
\keyword{internal}
