% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{evaluate.cvg}
\alias{evaluate.cvg}
\title{Evaluation of Coverage.}
\usage{
evaluate.cvg(template.seqs, primers, mode.directionality = c("fw", "rev"),
  allowed.mismatches, updateProgress = NULL)
}
\arguments{
\item{template.seqs}{Template sequences as a \code{DNAStringSet}.}

\item{primers}{Primer sequences as a \code{DNAStringSet}.}

\item{mode.directionality}{Directionality of primres}

\item{allowed.mismatches}{Allowed number of mismatches between a primer and a template.}

\item{updateProgress}{Progress function for shiny}
}
\value{
IRanges object with primer coverage information.
}
\description{
Evaluates primer coverage.
}
\keyword{internal}
