% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{evaluate.diff.primer.cvg}
\alias{evaluate.diff.primer.cvg}
\title{Evaluation of Coverage.}
\usage{
evaluate.diff.primer.cvg(primers, excluded.seqs, template.df)
}
\arguments{
\item{primers}{Primer data frame.}

\item{excluded.seqs}{Identifiers of templates to be excluded.}

\item{template.df}{Template data frame}
}
\value{
Primer data frame with updated coverage under the exclusion of \code{excluded.seeqs}.
}
\description{
Re-evaluates the coverage of primers under exclusion of certain templates.
}
\details{
This function requires that \code{primers} was already annotated with primer coverage before.
}
\keyword{internal}
