% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_algo.R
\name{evaluate.fw.rev.combinations}
\alias{evaluate.fw.rev.combinations}
\title{Evaluation of Set Combinations}
\usage{
evaluate.fw.rev.combinations(opti.fw, opti.rev, compatible.df, template.df)
}
\arguments{
\item{opti.fw}{List with forward optimal primer sets.}

\item{opti.rev}{List with reverse optimal primer sets.}

\item{compatible.df}{Data frame containing the indicices
of temperature-compatible forward and reverse primers sets.}

\item{template.df}{Template data frame for which primers were designed.}

\item{opti.rev.indices}{Indices for accessing \code{opti.rev}.}
}
\value{
List with information on the combinations of forward and reverse primers
as well as the combined data frames themselves.
}
\description{
Evaluates the combinations of forward and reverse primer sets.
}
\keyword{internal}
