% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_gc_clamp.R
\name{evaluate.GC.clamp}
\alias{evaluate.GC.clamp}
\title{GC clamp}
\usage{
evaluate.GC.clamp(y)
}
\arguments{
\item{y}{Pirmer sequence from 5' to 3'.}
}
\value{
The length of the GC clamp.
}
\description{
Determines the number of consecutive G/Cs at the 3' end.
}
\keyword{internal}
