% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_eval.R
\name{evaluate.template.constraints}
\alias{evaluate.template.constraints}
\title{Evaluation of Template Constraints.}
\usage{
evaluate.template.constraints(constraint.values, constraint.settings,
  active.constraints, mode.directionality = c("fw", "rev", "both"))
}
\arguments{
\item{constraint.values}{Data frame with template constraints}

\item{constraint.settings}{List specifying the allowed values for constraint evaluation.}

\item{active.constraints}{Strings specifying the constraints to check.}

\item{mode.directionality}{Direction of primers.}
}
\value{
List indicating which template constraints where fulfilled or not (TRUE/FALSE).
}
\description{
Evaluates the input template constraints.
}
\keyword{internal}
