% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_view.R
\name{exclude.cols}
\alias{exclude.cols}
\title{Exclusion of Columns}
\usage{
exclude.cols(excl.col, template.df)
}
\arguments{
\item{excl.col}{Names of columns in \code{template.df} to be removed.}

\item{template.df}{Data frame for which columns in \code{excl.col} should be removed.}
}
\value{
\code{template.df} with removed columns as specified in \code{excl.col}.
}
\description{
Removes columns from a data frame.
}
\keyword{internal}
