% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter.by.constraints}
\alias{filter.by.constraints}
\title{Filter By Constraints}
\usage{
filter.by.constraints(filtered.df, constraint.df, current.constraints,
  active.constraints, mode.directionality = c("fw", "rev", "both"),
  template.df)
}
\arguments{
\item{filtered.df}{Primer data frame.}

\item{constraint.df}{Data frame with constraint values.}

\item{current.constraints}{List with constraint settings.}

\item{active.constraints}{Strings giving the names of active constraints.}

\item{mode.directionality}{Direction of primers}

\item{template.df}{Template data frame.}
}
\value{
A list containing the filtered primer data frame, as well as a 
data frame of the excluded primers and the used filtering settings.
}
\description{
Remove primers that do not fulfill the current constraints (evaluate all primers).
}
\keyword{internal}
