% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_primers_tree.R
\name{filter.primer.candidates}
\alias{filter.primer.candidates}
\title{Filtering of Primer Candidates}
\usage{
filter.primer.candidates(primer.candidates, min.len)
}
\arguments{
\item{primer.candidates}{Alignment of candidate primers.}

\item{min.len}{Minimal required length of primers.}
}
\value{
Filtered alignment of candidate primers.
}
\description{
Filters primer candidates according to length and duplications.
}
\keyword{internal}
