% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_greedy.R
\name{filter_primers.by.Tm.delta}
\alias{filter_primers.by.Tm.delta}
\title{Filter by Melting Temperature Difference}
\usage{
filter_primers.by.Tm.delta(target.temp, selected.primers, max.Tm.delta)
}
\arguments{
\item{target.temp}{Target melting temperature in Celsius.}

\item{selected.primers}{Current candidate primer data frame.}

\item{max.Tm.delta}{Maximum allowed difference of primer melting temperatures to target temperature.}
}
\value{
Filtered primer data frame.
}
\description{
Filters primers by melting temperature differences.
}
\keyword{internal}
