% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_primers}
\alias{filter_primers}
\title{Filter a Set of Primers.}
\usage{
filter_primers(primer.df, template.df, settings,
  active.constraints = names(constraints(settings)))
}
\arguments{
\item{primer.df}{A \code{Primers} object containing the primers
to be filtered.}

\item{template.df}{A \code{Templates} object with the template sequences 
that are to be covered by \code{primer.df}.}

\item{settings}{A \code{DesignSettings} object specifying the parameters 
for filtering the primers.}

\item{active.constraints}{The constraints that are to be used for filtering
primers. By default, \code{active.constraints} is set to \code{NULL} such that
all active constraints are used.}
}
\value{
A \code{Primers} object containing only those primers 
fulfilling all specified constraints.
}
\description{
Filters a primer set according to the constraints specified via 
\code{settings} and \code{active.constraints} such that all primers
that do not fulfill the constraints are removed from \code{primer.df}.
}
\note{
Please note that some constraints can only be computed if additional software is installed,
please see \code{\link{DesignSettings}} for an overview.
}
\examples{
data(Ippolito)
filename <- system.file("extdata", "settings", 
             "C_Taq_PCR_high_stringency.xml", package = "openPrimeR")
settings <- read_settings(filename)
# Only retain the primers fulfilling the GC clamp constraint:
filtered.df <- filter_primers(primer.df, template.df, settings,
                 active.constraints = c("gc_ratio"))
}
\seealso{
Other primer functions: \code{\link{Primers-class}},
  \code{\link{check_constraints}},
  \code{\link{check_restriction_sites}},
  \code{\link{create_report}},
  \code{\link{design_primers}},
  \code{\link{get_initial_primers}},
  \code{\link{primer_significance}},
  \code{\link{score_degen}}, \code{\link{write_primers}}
}
\keyword{Primers}
