% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{methods}
\name{filterLimits}
\alias{filterLimits}
\alias{filterLimits,DesignSettings-method}
\title{Getter for Filtering Constraint Limits.}
\usage{
filterLimits(x)

\S4method{filterLimits}{DesignSettings}(x)
}
\arguments{
\item{x}{A \code{DesignSettings} object.}
}
\description{
Gets the limits on the constraints
that are used for the filtering procedure when designing primers
using the \code{Input_Constraint_Boundaries} slot of the provided
\code{DesignSettings} object \code{x}.
}
\keyword{internal}
