% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{filters}
\alias{filters}
\alias{filters,DesignSettings-method}
\title{Getter for Filtering Constraints.}
\usage{
filters(x)

\S4method{filters}{DesignSettings}(x)
}
\arguments{
\item{x}{A \code{DesignSettings} object.}
}
\value{
Gets the list of filtering constraints.
}
\description{
Gets the constraints on the physicochemical properties
that are used for the filtering procedure when designing primers
using the \code{Input_Constraints} slot of the provided
\code{DesignSettings} object \code{x}.
}
\keyword{internal}
