% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{fix_constraint_boundaries}
\alias{fix_constraint_boundaries}
\title{Correction of Constraint Boundaries.}
\usage{
fix_constraint_boundaries(constraints, constraint.limits, fix.limit = TRUE)
}
\arguments{
\item{constraints}{A list with constraint settings.}

\item{constraint.limits}{A list with constraint limits.}

\item{fix.limit}{Whether the constraint limits should be
adjusted. If \code{FALSE}, the constraint settings are adjusted.}
}
\value{
The corrected constraint limits.
}
\description{
Fixes the constraint boundaries if they are more narrow than the current settings.
}
\keyword{internal}
