% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{format.constraints}
\alias{format.constraints}
\title{Format Constraint Names.}
\usage{
\method{format}{constraints}(constraints)
}
\arguments{
\item{constraints}{The character vector of constraints to transform.}
}
\value{
A character vector with formatted constraint names.
}
\description{
Formats constraint names for frontend output.
}
\keyword{internal}
