% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stop_codons.R
\name{format.seq.ali}
\alias{format.seq.ali}
\title{Format mismatches}
\usage{
\method{format}{seq.ali}(seq, pos, format.type)
}
\arguments{
\item{seq}{The input sequence.}

\item{pos}{The mismatch positions to be formatted.}

\item{format.type}{Vector of giving the style (bold/italics) for each pos.}
}
\value{
The input sequence with highlighted mismatch positions.
}
\description{
Formats a sequence for highlighting mismatches in an alignment.
}
\keyword{internal}
