% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_view.R
\name{format.seqs.tex}
\alias{format.seqs.tex}
\title{Format a Sequence for LateX output.}
\usage{
\method{format}{seqs.tex}(seqs)
}
\arguments{
\item{seqs}{Character vector of sequences.}
}
\value{
Formatted sequences.
}
\description{
Formats a sequence for LateX report output.
}
\keyword{internal}
