% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{get.3prime.mismatch.pos}
\alias{get.3prime.mismatch.pos}
\title{Identification of 3' Mismatches.}
\usage{
get.3prime.mismatch.pos(primers, mismatches)
}
\arguments{
\item{primers}{Primer sequence strings.}

\item{mismatches}{Comma-separated strings containing the primer mismatch positions.}
}
\value{
The closest position of a mismatch relative to the 3' end of the primer.
Here, 1 indicates the terminal position, 2 the penultimate position, and so on.
No mismatch is indicated by an infinite value.
}
\description{
Computes the lastmost position of a 3' mismatches of a primer with a template.
}
\keyword{internal}
