% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{get.analysis.mode}
\alias{get.analysis.mode}
\title{Direction of Primers.}
\usage{
get.analysis.mode(primers)
}
\arguments{
\item{primers}{A primer data frame.}
}
\value{
\code{both} if both, forward and reverse primers exist in \code{primers}.
Otherwise, if either only forward primers or reverse primers exist, returns \code{fw}
or \code{rev}, respectively.
}
\description{
Identifies the directionality of the input primers.
}
\keyword{internal}
