% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_primers_tree.R
\name{get.consensus.seq}
\alias{get.consensus.seq}
\title{Computation of Consensus.}
\usage{
get.consensus.seq(ali)
}
\arguments{
\item{ali}{An alignment object.}
}
\value{
A consensus sequence without gap characters.
}
\description{
Computes the consensus of the sequences in the input alignment.
}
\keyword{internal}
