% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_eval.R
\name{get.constraint.value.idx}
\alias{get.constraint.value.idx}
\title{Retrive Constraint Indices.}
\usage{
get.constraint.value.idx(active.constraints, constraint.df)
}
\arguments{
\item{active.constraints}{The names of the constraints for which to find the indices in \code{constraint.df}.}

\item{constraint.df}{The primer data frame where the \code{active.constraints} should be found.}
}
\value{
Indices of \code{active.constraints} in \code{constraint.df}.
}
\description{
Gets the index of the required constraint columns in the primer data frame.
}
\keyword{internal}
