% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_ILP.R
\name{get.coverage.matrix}
\alias{get.coverage.matrix}
\title{Coverage Matrix}
\usage{
get.coverage.matrix(primer.df, template.df, constraints = NULL)
}
\arguments{
\item{primer.df}{Primer data frame.}

\item{template.df}{Template data frame.}

\item{constraints}{A character vector of coverage constraints
to be used as entries for the coverage matrix instead of the 0/1 encoding.
At its default setting (\code{NULL}), the 0/1 encoding is used.}
}
\value{
The binary coverage matrix.
}
\description{
Constructs a coverage matrix where rows indicate templates and columns indicate primers.
}
\details{
Entry (i,j) in the matrix is equal to 1 if primer j covers template i and otherwise 0.
}
\keyword{internal}
