% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{get.covered.templates}
\alias{get.covered.templates}
\title{Covered Templates}
\usage{
get.covered.templates(Tm.set, template.df)
}
\arguments{
\item{Tm.set}{Primer data frame.}

\item{template.df}{Template data set.}
}
\value{
Index of templates that are covered by the primers in \code{Tm.set}.
}
\description{
Get the indices of covered templates.
}
\keyword{internal}
