% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_melting_temperature.R
\name{get.delta.G}
\alias{get.delta.G}
\title{Change in Free Energy.}
\usage{
get.delta.G(delta.H, delta.S, temp = 37)
}
\arguments{
\item{delta.H}{Change in enthalpy in cal/mol.}

\item{delta.S}{Change in entropy cal/mol*K.}

\item{temp}{Temperature in Celsius for which to compute free energy change.}
}
\value{
The change in free energy in kcal/mol.
}
\description{
Computes the change in free energy.
}
\keyword{internal}
