% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{get.dimer.data}
\alias{get.dimer.data}
\title{Retrieval of dimerization energies.}
\usage{
get.dimer.data(s1, s2, annealing.temp, ions, no.structures)
}
\arguments{
\item{s1}{Nucleotide character vectors (5' to 3')}

\item{s2}{Nucleotide character vectors  (5' to 3')}

\item{annealing.temp}{The PCR annealing temperature in Celsius.}

\item{ions}{The sodium-equivalent ions used in the PCR.}

\item{no.structures}{Whether to compute structures of dimers.}
}
\value{
A data frame containing free energies in the field
\code{DeltaG} and the dimerization structure in \code{Structure}.
}
\description{
Uses OligoArrayAux to compute dimerization candidates.
}
\keyword{internal}
