% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\name{get.eval.cols}
\alias{get.eval.cols}
\title{Retrieval of Evaluation Columns.}
\usage{
get.eval.cols(primer.data, constraint.settings)
}
\arguments{
\item{primer.data}{A list with \code{Primers} objects.}

\item{constraint.settings}{A list with constraint settings.}
}
\value{
A character vector with EVAL-columns.
}
\description{
Retrieves the evaluation columns by intersecting
the already evaluated constraints in \code{primer.data}
as well as the constraints specified in input constraint settings.
}
\keyword{internal}
