% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get.extension}
\alias{get.extension}
\title{Identification of File extension.}
\usage{
get.extension(x)
}
\arguments{
\item{x}{A string for a filename.}
}
\value{
The extension of \code{x}.
}
\description{
Identifies the file extension of \code{x}.
}
\keyword{internal}
