% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_ILP.R
\name{get.ILP.vars}
\alias{get.ILP.vars}
\title{Retrieval of ILP Decisions}
\usage{
get.ILP.vars(ILP, original.dim = NULL)
}
\arguments{
\item{ILP}{A solved ILP instance.}

\item{original.dim}{Dimension of \code{ILP} before using presolve.}
}
\value{
The ILP decision variables.
}
\description{
Retrieves ILP decision variables.
}
\details{
The original dimension of the ILP is required to determine the correct decisions
when presolve has been active and dimensions of the ILP might have changed.
}
\keyword{internal}
