% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_algo.R
\name{get.init.file.name}
\alias{get.init.file.name}
\title{File Name for Initialized Primers.}
\usage{
get.init.file.name(cur.results.loc, GROUP, primer.lengths, mode.directionality,
  allowed.region.definition, init.algo, max.degen, conservation)
}
\arguments{
\item{cur.results.loc}{Directory where the file should be stored.}

\item{GROUP}{Sample name of templates.}

\item{primer.lengths}{Interval of desired primer lengths.}

\item{mode.directionality}{Directionality of the primers}

\item{allowed.region.definition}{Definition of the allowed region.}

\item{init.algo}{Initialization algorithm identifier.}

\item{max.degen}{Maximum degeneracy of primers.}

\item{conservation}{Required ratio of primer conservation.}
}
\value{
A filename for the initialized primers.
}
\description{
Constructs a filename for initialized primers.
}
\keyword{internal}
