% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{get.leader.exon.regions}
\alias{get.leader.exon.regions}
\title{Assign Binding Regions}
\usage{
get.leader.exon.regions(lex.seqs, uni.leaders)
}
\arguments{
\item{lex.seqs}{Data frame with template sequences.}

\item{uni.leaders}{Data frame with individual allowed binding regions.}
}
\value{
Template data frame with annotated binding regions.
}
\description{
Augments a template data frame with individual binding regions.
}
\keyword{internal}
