% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{get.matches}
\alias{get.matches}
\title{Identification of Sequence Matches.}
\usage{
get.matches(s1, s2)
}
\arguments{
\item{s1}{The aligned nucleotide sequence character vector.}

\item{s2}{The aligned, matching substring of \code{s1}.}
}
\value{
A match vector (\code{M} for matches, \code{X} for mismatches).
}
\description{
Identifies matches between two strings
provided by OligoArrayAux.
}
\keyword{internal}
