% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_melting_temperature.R
\name{get.melting.temp.diff}
\alias{get.melting.temp.diff}
\title{Computation of Maximal Melting Temperature Differences.}
\usage{
get.melting.temp.diff(Tm.fw, Tm.rev)
}
\arguments{
\item{Tm.fw}{The melting temperatures of forward primers.}

\item{Tmr.rev}{The melting temperatures of reverse primers.}
}
\value{
The worst-case melting temperature difference, for every primer.
}
\description{
Computation of Maximal Melting Temperature Differences.
}
\keyword{internal}
