% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguity.R
\name{get.merge.idx}
\alias{get.merge.idx}
\title{Indices for merging sequences}
\usage{
get.merge.idx(seqs, max.degeneracy)
}
\arguments{
\item{seqs}{The input sequence strings.}

\item{max.degeneracy}{The maximal allowed degeneracy of a merged seq.}
}
\value{
A list of lists containing the indices of seqs to be merged. 
  For example [[1,2,3]] would indicate to merge primers 1, 2, and 3.
}
\description{
Identifies the indices of similar input sequences to be merged.
}
\keyword{internal}
