% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{get.missing.df}
\alias{get.missing.df}
\title{Uncovered Templates.}
\usage{
get.missing.df(filtered.df, template.df, Tm.brackets, settings,
  mode.directionality)
}
\arguments{
\item{filtered.df}{An object of class \code{Primers}.}

\item{template.df}{An object of class \code{Templates}.}
}
\value{
A Templates data frame containing the missing templates.
}
\description{
Computes a data frame containing the templates that are
not yet covered.
}
\keyword{internal}
