% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stop_codons.R
\name{get.ORFs}
\alias{get.ORFs}
\title{Identification of ORFs.}
\usage{
get.ORFs(template.df)
}
\arguments{
\item{template.df}{template data frame.}
}
\value{
Returns a data frame containing the shift of the ORF (either 0,1, or 2) 
for every sequence, as well as a comment in case of problems.
}
\description{
Given a template data frame, identify the exon reading frames in the sequences.
}
\keyword{internal}
