% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{get.other.constraint.settings}
\alias{get.other.constraint.settings}
\title{Gather all Other Constraints (for Shiny frontend).}
\usage{
get.other.constraint.settings(allowed_mismatches, allowed_other_binding_ratio,
  allowed_region_definition)
}
\arguments{
\item{allowed_mismatches}{Allowed mismatches for primers binding events.}

\item{allowed_other_binding_ratio}{Ratio of primers allowed to bind to non-target regions.}

\item{allowed_region_definition}{The definition of the allowed region.}
}
\value{
List with all other constraint settings.
}
\description{
Constructor for other constraint settings (non-PCR, non-filtering, non-optimization).
}
\keyword{internal}
