% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{get.PCR.settings}
\alias{get.PCR.settings}
\title{Gather all PCR settings.}
\usage{
get.PCR.settings(use_taq_polymerase, annealing_temp, Na_concentration,
  Mg_concentration, K_concentration, Tris_concentration, primer_concentration,
  template_concentration, nbr.cycles)
}
\arguments{
\item{annealing_temp}{Annealing temperature in Celsius.}

\item{Na_concentration}{Sodium ion concentration.}

\item{Mg_concentration}{Magensium ion concentration.}

\item{K_concentration}{Potassium ion concentration.}

\item{Tris_concentration}{Tris ion concentration.}

\item{primer_concentration}{Primer concentration.}

\item{template_concentration}{Template concentration.}
}
\value{
List with all PCR settings.
}
\description{
Gathers all PCR settings (e.g. for XML output).
}
\keyword{internal}
