% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get.plot.height}
\alias{get.plot.height}
\title{Plot Extent}
\usage{
get.plot.height(N, px.per.n = 50, min.size = 300, max.size = 1500000)
}
\arguments{
\item{N}{Number of observations to plot.}

\item{px.per.n}{Pixels required per observations.}

\item{min.size}{Minimal extent of plot in pixels.}

\item{max.size}{Maximal extent of plot in pixels.}
}
\value{
The extent of the plot.
}
\description{
Returns the extent of a plot.
}
\keyword{internal}
