% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{get.primer.binding.idx}
\alias{get.primer.binding.idx}
\title{Retrieval of Allowed Binding Indices.}
\usage{
get.primer.binding.idx(binding, primer.type = c("fw", "rev", "both"), x,
  allowed.other.binding.ratio)
}
\arguments{
\item{binding}{IRanges binding information.}

\item{primer.type}{Direction of primer.}

\item{x}{Index of primer in the primer data frame.}

\item{allowed.other.binding.ratio}{The ratio of allowed off-target binding events.}
}
\value{
Indices in binding for primer with index code{x} that are allowed.
}
\description{
Retrieves the indices of allowed binding events in \code{binding}
for the primer with index \code{x} and type \code{primer.type}.
}
\keyword{internal}
