% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_primers.R
\name{get.primer.identifier.string}
\alias{get.primer.identifier.string}
\title{Primer Identifier Creation.}
\usage{
get.primer.identifier.string(sample, seq.IDs, seq.identifier, all.starts,
  all.ends, identifier, seq.primers)
}
\arguments{
\item{sample}{Sample name of the templates.}

\item{seq.IDs}{Identifiers of the templates.}

\item{all.starts}{Primer positions (start).}

\item{all.ends}{Primer positions (end).}

\item{identifier}{Direction keyword.}

\item{seq.primers}{The primer sequences as strings.}

\item{seq.identifiers}{The index of the seq.}
}
\value{
Identifiers for each primer.
}
\description{
Creates identifiers for generated primers.
}
\keyword{internal}
