% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_ILP.R
\name{get.redundant.cols}
\alias{get.redundant.cols}
\title{Identification of Redudant Primers.}
\usage{
get.redundant.cols(cvg.matrix)
}
\arguments{
\item{cvg.matrix}{Binary matrix of coverage events.}
}
\value{
TRUE for redundant primers, FALSE otherwise.
}
\description{
Identifies primers that are redundant.
}
\details{
Redundant primers do not reduce the coverage when removed.
}
\keyword{internal}
