% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{get.relative.binding.pos}
\alias{get.relative.binding.pos}
\title{Retrieval of Relative Binding Positions.}
\usage{
get.relative.binding.pos(allowed, primer.pos, direction, covered.seqs.idx)
}
\arguments{
\item{allowed}{Positions where binding is allowed in the templates.}

\item{primer.pos}{Binding position of primer (absolute).}

\item{direction}{Direction (either fw/rev).}

\item{covered.seqs.idx}{Indices of covered templates.}
}
\value{
Numeric of relative binding position to allowed region.
}
\description{
Retrieves primer binding position relative to allowed regions of either
forward or reverse primers, as specified by \code{direction}.
}
\keyword{internal}
