% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{get.run.names}
\alias{get.run.names}
\title{Getter for Run Names.}
\usage{
get.run.names(primer.data)
}
\arguments{
\item{primer.data}{A list with \code{Primers} or \code{Templates}.}
}
\value{
A vector with identifiers for every set.
}
\description{
Retrieves the run names of the input data.
}
\keyword{internal}
