% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{get.self.dimers}
\alias{get.self.dimers}
\title{Self dimerization}
\usage{
get.self.dimers(primers.1, primers.2, ions, annealing.temp,
  no.structures = FALSE)
}
\arguments{
\item{primers.1}{Input primers}

\item{primers.2}{(Copy/reverse) of the input primers}

\item{ions}{Sodium-equivalent ionic concentration.}

\item{annealing.temp}{The annealing temperature.}

\item{no.structures}{Whether the dimerization structure shall be computed.}
}
\value{
Possible self-dimer conformations.
}
\description{
Computes possible self-dimers.
}
\keyword{internal}
