% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{get.static.tool.info}
\alias{get.static.tool.info}
\title{Retrieval of Tool Information.}
\usage{
get.static.tool.info()
}
\value{
A data frame with information about the required tools.
}
\description{
Constructs a data frame containing information about the tools.
}
\keyword{internal}
