% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get.unlist.idx}
\alias{get.unlist.idx}
\title{Index for Unlisting.}
\usage{
get.unlist.idx(primer.start, primer.data.idx)
}
\arguments{
\item{primer.start}{Numeric vector.}

\item{primer.data.idx}{Selection indices.}
}
\value{
Indices.
}
\description{
Determines indices for unlisting.
}
\keyword{internal}
