% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_comparison.R
\name{get_comparison_table}
\alias{get_comparison_table}
\title{Overview of Primer Set Properties.}
\usage{
get_comparison_table(template.data, primer.data, sample.name = NULL)
}
\arguments{
\item{template.data}{List with \code{Template} objects corresponding
to \code{primer.data}.}

\item{primer.data}{List with evaluated \code{Primers} objects whose properties
are to be summarized.}

\item{sample.name}{Either a single identifier or identifiers for every
\code{Templates} object in \code{template.data}.
By default, \code{sample.name} is \code{NULL} such that the
\code{Run} annotations in the \code{Templates} objects provided
by \code{template.data} are used.}
}
\value{
A data frame summarizing the properties of each primer set.
}
\description{
Creates an overview of the properties of multiple primer sets by providing
the inter-quartile range of primer properties in bracket notation.
}
\examples{
data(Comparison)
tab <- get_comparison_table(template.data[1:3], primer.data[1:3], "IGH")
}
