% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\name{get_constraint_deviation_data}
\alias{get_constraint_deviation_data}
\title{Retrieve data for Constraint Deviations.}
\usage{
get_constraint_deviation_data(constraint.df, constraint.settings)
}
\arguments{
\item{constraint.df}{An evaluated object of class \code{Primers}.}

\item{constraint.settings}{A list with settings for the constraints
that are to be evalated.}
}
\value{
A data frame providing primer-specific 
information on deviations of primer properties from
the desired properties.
}
\description{
Retrieve data for Constraint Deviations.
}
\keyword{internal}
