% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{get_cvg_ratio}
\alias{get_cvg_ratio}
\title{Determination of the Ratio of Covered Templates.}
\usage{
get_cvg_ratio(primers, template.df, allowed.mismatches = NULL,
  cvg.definition = c("constrained", "basic"), mode.directionality = NULL,
  as.char = FALSE)
}
\arguments{
\item{primers}{A \code{Primers} object containing the primers
for which the coverage should be evaluated.}

\item{template.df}{A \code{Templates} object containing
the template sequences corresponding to \code{primers}.}

\item{allowed.mismatches}{The number of allowed mismatches
for determining the coverage of the templates. By default,
\code{allowed.mismatches} is set to \code{NULL} such that
all annotated coverage events are considered.}

\item{cvg.definition}{Whether to output the coverage obtained
from string matching ("basic") or the expected coverage ("constrained"),
which is constructed by applying the coverage constraints. By default,
\code{cvg.definition} is set to "constrained".}

\item{mode.directionality}{If \code{mode.directionality} is provided,
the coverage of templates is computed for a specific direction of primers.
Either "fw" (forward coverage only), "rev" (reverse coverage only), or "both" for both directions. By default, \code{mode.directionality} is \code{NULL}
such that the directionality of the primers is determined automatically.}

\item{as.char}{Whether the coverage ratio should
be outputted as a percentage-formatted character vector. By default,
\code{as.char} is set to \code{FALSE} such that a numeric is returned.}
}
\value{
By default, a numeric providing the expected primer coverage ratio. 
If \code{as.char} is \code{TRUE}, the output is provided as a
percentage-formatted character vector. 
The attributes \code{no_covered}, \code{no_templates}, 
and \code{covered_templates} provide the number of covered templates, 
the total number of templates, and the IDs of covered templates, respectively.
}
\description{
Determines the ratio of template sequences 
that are covered by the evaluated input primers. The ratio
is in the interval [0,1] where 0 indicates 0\% coverage (no templates
covered) and 1 indicates 100\% coverage (all templates covered).
}
\details{
The manner in which the coverage ratio is evaluated depends on 
the directionality of the input primers.
If either only forward or reverse primers are inputted, the individual
coverage of each primer is used to determine the overall coverage. 
If, however, forward and reverse primers are inputted at the same time, 
the coverage is defined by the intersection of binding events from both,
forward and reverse primers.
}
\examples{
data(Ippolito)
cvg.ratio <- get_cvg_ratio(primer.df, template.df)
# determine the identitity coverage ratio 
cvg.ratio.0 <- get_cvg_ratio(primer.df, template.df, allowed.mismatches = 0)
}
