% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{get_cvg_stats}
\alias{get_cvg_stats}
\title{Coverage Ratios per Group of Templates.}
\usage{
get_cvg_stats(primers, templates, for.viewing = FALSE,
  total.percentages = FALSE, allowed.mismatches = Inf,
  cvg.definition = c("constrained", "basic"))
}
\arguments{
\item{primers}{To retrieve coverage statistics for a single primer set, 
please provide an object of class \code{Primers} containing primers with evaluated coverage.
To retrieve coverage statistics for multiple primer sets, pelase provide
a list with evaluated \code{Primers} objects.}

\item{templates}{If \code{primers} is an object of class \code{Primers},
please provide an object of class \code{Templates} containing the
template sequences targeted by \code{primers}. If \code{primers} is a list,
\code{templates} should be a list of \code{Template} objects.}

\item{for.viewing}{Whether the table should be formatted
to be human-readable. By default, \code{for.viewing} is \code{FALSE}.}

\item{total.percentages}{Whether group coverage percentages
should be computed in relation to the total number of template sequences
or in relation to the number of templates belonging to a specific group.
By default, \code{total.percentages} is \code{FALSE} suc that the
percentages are group-specific.}

\item{allowed.mismatches}{The maximal allowed number of mismatches.
By default, \code{allowed.mismatches} is set to \code{Inf} such that the number of mismatches is not restricted additionally.}

\item{cvg.definition}{If \code{cvg.definition} is set to
"constrained", the statistics for the expected
coverage (after applying the coverage constraints) are retrieved.
If \code{cvg.definition} is set to "basic", the coverage is determined 
solely by string matching (i.e. without applying the coverage constraints).
By default, \code{cvg.definition} is set to "constrained".}
}
\value{
Data frame whose entries provide the coverage of templates belonging to a specific group.
}
\description{
Retrieve statistics on covered templates, either for a single primer set
or for multiple primer sets.
}
\examples{
# Coverage statistics for a single primer set
data(Ippolito)
cvg.stats <- get_cvg_stats(primer.df, template.df)
# Coverage statistics for multiple primer sets
data(Comparison)
cvg.stats.comp <- get_cvg_stats(primer.data, template.data)
}
