% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{get_cvg_stats_primer}
\alias{get_cvg_stats_primer}
\title{Statistics on the Number of Coverage Events per Primer.}
\usage{
get_cvg_stats_primer(primer.df, template.df, cvg.definition = c("constrained",
  "basic"))
}
\arguments{
\item{primer.df}{A \code{Primers} object with evaluated coverage providing
the set of primers for which the coverage statistics shall be computed.}

\item{template.df}{A \code{Templates} object providing
the template sequences for which the primer coverage
has been computed.}

\item{cvg.definition}{If \code{cvg.definition} is set to
"constrained", the statistics for the expected
coverage (after applying the coverage constraints) are retrieved.
If \code{cvg.definition} is set to "basic", the coverage is determined 
solely by string matching (i.e. without applying the coverage constraints).
By default, \code{cvg.definition} is set to "constrained".}
}
\value{
A data frame listing the number of binding events
broken down according to the number of expected mismatches between
primers and templates.
}
\description{
Creates a table summarizing the coverage events
of each primer according to the number of mismatches
between primers and templates.
}
\details{
Entries in numeric table columns
indicate the percentage of coverage events occurring
with a certain number of mismatches. For example
column \emph{3} provides all coverage events
with exactly three mismatches between primers and templates.
The column \emph{Group_Coverage} provides a listing
of the percentage of covered templates per group.
}
\examples{
data(Ippolito)
primer.cvg.stats <- get_cvg_stats_primer(primer.df, template.df)
}
