% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{get_max_set_coverage}
\alias{get_max_set_coverage}
\title{Determination of Maximal Coverage.}
\usage{
get_max_set_coverage(primer.df, template.df, Tm.brackets, settings,
  mode.directionality, max.only = TRUE)
}
\arguments{
\item{primer.df}{An object of class \code{Primers}.}

\item{template.df}{An objectc of class \code{Templates}.}

\item{Tm.brackets}{A data frame with temperature information.}

\item{settings}{A \code{DesignSettings} object.}

\item{mode.directionality}{The direction of the primers.}

\item{max.only}{Whether only the maxium coverage shall be returned.
If \code{max.only} is \code{FALSE}, the coverage ratios of all 
melting temperature sets according to \code{Tm.brackets} are returned.}
}
\value{
The maximal coverage ratio of a primer set if \code{max.only} is \code{TRUE}
or the coverages of all melting temperature seets if \code{max.only} is \code{FALSE}.
}
\description{
Determines the maximal coverage ratio of a set of primers 
for primer subsets valid for a certain temperature range.
a certain melting temperature range.
}
\keyword{internal}
