% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{get_report_fname}
\alias{get_report_fname}
\title{Creation of a Filename for Reports.}
\usage{
get_report_fname(report.name, sample.name)
}
\arguments{
\item{report.name}{The identifier for the report type.}

\item{sample.name}{The identifier of the sample that was analyzed.}
}
\value{
A character vector.
}
\description{
Creates the filename for reports.
}
\keyword{internal}
