% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{get_template_cvg_data}
\alias{get_template_cvg_data}
\title{Retrieval of Template Coverage Data.}
\usage{
get_template_cvg_data(primer.df, template.df)
}
\arguments{
\item{primer.df}{A \code{Primers} object.}

\item{template.df}{A \code{Templates} object.}
}
\value{
Computes a data frame providing the coverage of the 
templates for the basic as well as expected (constrained) coverage.
}
\description{
Determines the coverage of the templates for individual
allowed mismatch settings and coverage definitions.
}
\keyword{internal}
