% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_primers_tree.R
\name{hclust.tree}
\alias{hclust.tree}
\title{Hierarchical Clustering.}
\usage{
hclust.tree(primer.candidates)
}
\arguments{
\item{primer.candidates}{Alignment of primer candidates.}
}
\value{
Phylogeny of the input \code{primer.candidates}.
}
\description{
Performs hierarchical clustering on aligned primer sequences.
}
\details{
The clustering is performed to identify similar groups of primer candidates
that can be merged to form degenerate primers.
}
\keyword{internal}
