% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stop_codons.R
\name{highlight.mismatch}
\alias{highlight.mismatch}
\title{Highlight mismatches}
\usage{
highlight.mismatch(seq, mm.seq)
}
\arguments{
\item{seq}{character vector of the original sequence}

\item{mm.seq}{character vector of the mutated sequence}
}
\value{
A list highlighting the mutations and additional information 
(mutation type, number, etc.)
}
\description{
Collects information on the mutations present in the input and highlights the
mutations in the sequence.
}
\keyword{internal}
