% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{html.format.structure}
\alias{html.format.structure}
\title{Formats a Dimerization Structure for HTML.}
\usage{
html.format.structure(structures)
}
\arguments{
\item{structures}{A character vector of dimerization structures.}
}
\value{
HTML-formatted character vectors.
}
\description{
Formats a Dimerization Structure for HTML.
}
\keyword{internal}
