% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_ILP.R
\name{I.cvg}
\alias{I.cvg}
\title{Primer Coverage.}
\usage{
I.cvg(cvg.matrix)
}
\arguments{
\item{cvg.matrix}{Binary matrix of covering events.}
}
\value{
A list with covered templates for every primer.
}
\description{
Determines the indices of covered templates for every primer.
}
\keyword{internal}
