% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_ILP.R
\name{ILPConstrained}
\alias{ILPConstrained}
\title{Construct Coverage ILP.}
\usage{
ILPConstrained(D, cvg.matrix, time.limit = NULL, presolve.active = FALSE)
}
\arguments{
\item{D}{Binary dimerization matrix.}

\item{cvg.matrix}{Binary coverage matrix.}

\item{time.limit}{Time limit for ILP optimization in seconds.}

\item{presolve.active}{Whether the ILP presolver should be used.
This is set to \code{FALSE} by default, since presolving may lead
to inferior solutions. However, for large problems presolving
might be useful.}
}
\value{
An instance of the set cover ILP.
}
\description{
Constructs an ILP modeling the primer set cover problem.
}
\keyword{internal}
