% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stop_codons.R
\name{insert_str}
\alias{insert_str}
\title{String Insertion.}
\usage{
insert_str(target, insert, index)
}
\arguments{
\item{target}{The string to be modified.}

\item{insert}{The string to be inserted.}

\item{index}{The position where the insertion should take place.}
}
\value{
A string where \code{insert} is inserted into \code{target} at position \code{index}.
}
\description{
Inserts a string into another string at the speficied position.
}
\keyword{internal}
