% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stop_codons.R
\name{interleave}
\alias{interleave}
\title{Interleave strings
Combines the input vectors in an interleaved fashion.}
\usage{
interleave(v1, v2)
}
\arguments{
\item{v1}{Input string.}

\item{v2}{Input string.}
}
\value{
The interleaved combination of \code{v1} and \code{v2}.
}
\description{
Interleave strings
Combines the input vectors in an interleaved fashion.
}
\keyword{internal}
