% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_ILP.R
\name{J.cvg}
\alias{J.cvg}
\title{Template Coverage.}
\usage{
J.cvg(cvg.matrix)
}
\arguments{
\item{cvg.matrix}{Binary matrix of covering events.}
}
\value{
A list with covering primers for every template.
}
\description{
Determines the indices of covering primers for every template.
}
\keyword{internal}
