% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_melting_temperature.R
\name{joule.to.cal}
\alias{joule.to.cal}
\title{Conversion from J to cal}
\usage{
joule.to.cal(val.J)
}
\arguments{
\item{val.J}{Numeric Joule value.}
}
\value{
The value correspdoning to \code{val.J} in calories.
}
\description{
Converts the input from Joule to calories.
}
\keyword{internal}
