% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguity.R
\name{merge.ambig.primers}
\alias{merge.ambig.primers}
\title{Merge similar primers}
\usage{
\method{merge}{ambig.primers}(primer.df, mode.directionality = c("fw", "rev",
  "both"), max.degeneracy)
}
\arguments{
\item{primer.df}{Primer data frame.}

\item{mode.directionality}{Analysis direction.}

\item{max.degeneracy}{Maximal degeneracy of merged primers.}
}
\value{
A primer data frame where similar primers are merged into one entry.
}
\description{
Merges similar primers contained in the input primer data frame.
}
\keyword{internal}
