% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{merge.binding.information}
\alias{merge.binding.information}
\title{Merge of Forward/Reverse Binding Information.}
\usage{
\method{merge}{binding.information}(primers, fw.binding.filtered,
  rev.binding.filtered, mode.directionality = c("fw", "rev", "both"), idx.fw,
  idx.rev)
}
\arguments{
\item{primers}{The primer data frame.}

\item{mode.directionality}{Primer directionality.}

\item{idx.fw}{Index of fw primers.}

\item{idx.rev}{Index of rev primers.}

\item{fw.binding}{IRanges object with binding events of fw primers.}

\item{rev.binding}{IRanges object with binding events of rev primers.}
}
\value{
IRanges with correct binding events.
}
\description{
Determines binding events of individual and pairs of primers.
}
\keyword{internal}
