% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguity.R
\name{merge.primer.entries}
\alias{merge.primer.entries}
\title{Merge similar primers}
\usage{
\method{merge}{primer.entries}(opti.result, merge.idx,
  mode.directionality = c("fw", "rev", "both"))
}
\arguments{
\item{opti.result}{Input primer data frame.}

\item{merge.idx}{List of lists with merge indices (get.merge.idx).}

\item{mode.directionality}{Direction of primers.}
}
\value{
A primer data frame where entries of similar primers are merged.
}
\description{
Merges the entries of similar entries in the input primer data frame, given a list with merge indices.
}
\keyword{internal}
