% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguity.R
\name{merge.primer.entries.single}
\alias{merge.primer.entries.single}
\title{Merge input sequences}
\usage{
\method{merge}{primer.entries.single}(seqs, merge.idx)
}
\arguments{
\item{seqs}{The input sequences.}

\item{merge.idx}{List of list with merge indices.}
}
\value{
Merged input sequences according to the input merge indices.
}
\description{
Merges the input sequences given a list with merge indices.
}
\keyword{internal}
