% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguity.R
\name{merge.select}
\alias{merge.select}
\title{Select merge indices}
\usage{
\method{merge}{select}(merge.idx)
}
\arguments{
\item{merge.idx}{list of lists containing the indices of possible merges}
}
\value{
The smallest number of possible merge operations as an index list.
}
\description{
Greedily identifies the smallest number of possible sequences merges that can be performed.
}
\keyword{internal}
