% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_eval.R
\name{merge.template.decisions}
\alias{merge.template.decisions}
\title{Merge Template Decisions.}
\usage{
\method{merge}{template.decisions}(eval.t)
}
\arguments{
\item{eval.t}{List with evaluated template constraints}
}
\value{
List with merged boolean decisions.
}
\description{
Merges the results for multiple template evaluations.
}
\keyword{internal}
