% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{mismatch.info}
\alias{mismatch.info}
\title{Information about Mismatches.}
\usage{
mismatch.info(primer, seqs)
}
\arguments{
\item{primer}{Primer character vector.}

\item{seqs}{Template binding sequences of primers as a \code{XStringsView} object.}
}
\value{
List with positions and number of mismatches of the \code{primer} in the 
\code{seqs}. The list contains the field \code{mm.pos} containing a list
with the positions of the mismatches and the field \code{Nbr} containing
a numeric vector with the number of mismatches per template binding event.
}
\description{
Computes information about mismatch binding events.
}
\keyword{internal}
