% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{mismatch.string.to.list}
\alias{mismatch.string.to.list}
\title{Conversion of Mismatch Postions String to List.}
\usage{
mismatch.string.to.list(mismatches)
}
\arguments{
\item{mismatches}{A character vector where parenthesis give mismatches for a template binding event.}
}
\value{
A list with the mismatches for every template for every primer.
}
\description{
Conversion of Mismatch Postions String to List.
}
\keyword{internal}
