% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_view.R
\name{modify.col.rep}
\alias{modify.col.rep}
\title{Modification of Column Names.}
\usage{
modify.col.rep(template.df, for.shiny = TRUE)
}
\arguments{
\item{template.df}{The data frame whose column names are to be modified.}

\item{for.shiny}{Whether formatting should be for shiny.}
}
\value{
\code{template.df} with modified column names.
}
\description{
Modifies column names for frontend output.
}
\keyword{internal}
