% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguity.R
\name{my.disambiguate}
\alias{my.disambiguate}
\title{Disambiguation of Sequences.}
\usage{
my.disambiguate(template.seqs, gap.char = "-", degen.cutoff = 2^10)
}
\arguments{
\item{template.seqs}{A \code{DNAStringSet} object with sequences to disambiguate.}

\item{gap.char}{The character indicating gaps in alignments.}

\item{degen.cutoff}{The maximal degeneration of sequences to be disambiguated.}
}
\value{
A \code{DNAStringSetList} object with disambiguated sequences.
}
\description{
Disambiguates the input sequences, but does not disambiguate highly generate sequences.
}
\keyword{internal}
