% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{my.read.fasta}
\alias{my.read.fasta}
\title{Read FASTA File.}
\usage{
my.read.fasta(fasta.file, NTs)
}
\arguments{
\item{fasta.file}{The path to a FASTA file.}

\item{NTs}{The allowed set of nucleotides.}
}
\value{
List with vectors of chars.
}
\description{
Reads the input FASTA file.
}
\keyword{internal}
