% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{my.warning}
\alias{my.warning}
\title{Custom Warning.}
\usage{
my.warning(subclass, message, call = sys.call(-1), ...)
}
\arguments{
\item{subclass}{String giving the specific type of error.}

\item{message}{Message to be displayed to the user.}

\item{call}{Environment where the error ocurred.}

\item{...}{Other arguments to the condition function.}
}
\value{
Generates a custom warning.
}
\description{
Creates a warning with a custom class.
}
\keyword{internal}
