% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{my_ggsave}
\alias{my_ggsave}
\title{Wrapper for the ggplot2::ggsave function.}
\usage{
my_ggsave(filename, plot = ggplot2::last_plot(), ...)
}
\arguments{
\item{filename}{The filename to store the plot.}

\item{plot}{The ggplot object.}

\item{...}{Further arguments to the ggplot2 ggsave function.}
}
\value{
Stores \code{p} in \code{fname}.
}
\description{
Saves a plot using ggplot2's ggsave function.
}
\keyword{internal}
