% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_repeats.R
\name{nbr.of.repeats}
\alias{nbr.of.repeats}
\title{Number of Repeats}
\usage{
nbr.of.repeats(x)
}
\arguments{
\item{x}{Input sequence strings.}
}
\value{
The maximal number of dinucleotide repeats in \code{x}.
}
\description{
Computes the number of dinucleotide repeats in the input sequences.
}
\keyword{internal}
