% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_runs.R
\name{nbr.of.runs}
\alias{nbr.of.runs}
\title{Number of Runs}
\usage{
nbr.of.runs(x)
}
\arguments{
\item{x}{Primer character sequences.}
}
\value{
The longest repeat of a single character in \code{x}.
}
\description{
Computes the longest run of a single character in the input sequence.
}
\keyword{internal}
