% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{methods}
\name{optiLimits}
\alias{optiLimits}
\alias{optiLimits,DesignSettings-method}
\title{Getter for Optimization Constraint Limits.}
\usage{
optiLimits(x)

\S4method{optiLimits}{DesignSettings}(x)
}
\arguments{
\item{x}{A \code{DesignSettings} object.}
}
\description{
Gets the limits for the constraints
that are applied just before the optimization procedure
using the \code{Input_Constraint_Boundaries} slot of the provided
\code{DesignSettings} object \code{x}.
}
\keyword{internal}
