% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_template_secondary_structures.R
\name{optimize.template.binding.regions.dir}
\alias{optimize.template.binding.regions.dir}
\title{Optimization of Binding Regions}
\usage{
optimize.template.binding.regions.dir(template.df,
  annealing.temperature = NULL, primer.lengths,
  mode.directionality = c("fw", "rev", "both"))
}
\arguments{
\item{template.df}{Template data frame.}

\item{annealing.temperature}{Temperature at which to compute secondary structures.}

\item{primer.lengths}{Target length of primers that are to be used.}

\item{mode.directionality}{Direction of primers.}
}
\value{
List with intervals indicating improved primer binding regions.
}
\description{
Optimizes the template binding regions.
}
\keyword{internal}
