% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_template_secondary_structures.R
\name{optimize.template.binding.regions.single}
\alias{optimize.template.binding.regions.single}
\title{Optimization of Template Binding}
\usage{
optimize.template.binding.regions.single(template.df, annealing.temperature,
  primer.lengths, mode.directionality = c("fw", "rev"))
}
\arguments{
\item{template.df}{Template data frame.}

\item{annealing.temperature}{Temperature at which to compute secondary structures.}

\item{primer.lengths}{Target length of primers that are to be used.}

\item{mode.directionality}{Direction of primers.}
}
\value{
List with new binding intervals for every template.
}
\description{
Optimizes template binding regions according to secondary structures.
}
\keyword{internal}
