% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{pair_primers}
\alias{pair_primers}
\title{Pairing of Forward and Reverse Primers.}
\usage{
pair_primers(primer.df, template.df)
}
\arguments{
\item{primer.df}{An object of class \code{Primers}.}
}
\value{
An object of class \code{Primers} containing the paired primers.
}
\description{
Pairs forward and reverse primers such that coverage is maximized 
for every pair.
}
\keyword{internal}
