% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{parallel_setup}
\alias{parallel_setup}
\title{Setup the Parallel Backend.}
\usage{
parallel_setup(cores = NULL)
}
\arguments{
\item{cores}{A numeric providing the number of cores to use. The default is \code{NULL}
such that half the number of available cores are used.}
}
\value{
Returns \code{NULL}
}
\description{
Registers the specified number of cores with the parallel backend.
}
\examples{
# Use two cores for parallel processing:
parallel_setup(2)
}
