% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{parse.constraints}
\alias{parse.constraints}
\title{Parse XML Constraint Data.}
\usage{
parse.constraints(xml_data)
}
\arguments{
\item{xml_data}{XML object from a parsed XML file.}
}
\value{
List with constraint settings.
}
\description{
Parses the constraint settings contained in an XML object.
}
\keyword{internal}
