% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{parse.header}
\alias{parse.header}
\title{Parse FASTA headers}
\usage{
parse.header(hdr, delim, hdr.str, id.column)
}
\arguments{
\item{hdr}{The headers (> entries) of a FASTA File.}

\item{delim}{The delimiter used to separate distinct fields in the headers.
For example, \code{|} for headers such as > E.coli | GeneX | ...}

\item{hdr.str}{Names of the fields appearing in the header.}

\item{id.column}{Field in the header to be used used as an identifier for the sequences.}
}
\value{
Data frame with structured information from the headers.
}
\description{
Parses the headers of a FASTA file.
}
\keyword{internal}
