% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{parse.IMGT.gene.groups}
\alias{parse.IMGT.gene.groups}
\title{Parser for IMGT Groups.}
\usage{
parse.IMGT.gene.groups(IDs)
}
\arguments{
\item{IDs}{Group information strings to be parsed.}
}
\value{
Data frame with structured group information.
}
\description{
Parses IMGT group information contained in FASTA headers.
}
\keyword{internal}
