% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{parse.oligo.results}
\alias{parse.oligo.results}
\title{Parser for OligoArrayAux Dimerization Data.}
\usage{
parse.oligo.results(deltaG.file, struct.file)
}
\arguments{
\item{deltaG.file}{A path to a file with OligoArrayAux energies.}

\item{struct.file}{A path to a file with OligoArrayAux structures.}
}
\value{
A data frame with structures and free energies.
}
\description{
Parses the free energies and structures of OligoArrayAux.
}
\keyword{internal}
