% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{methods}
\name{PCR}
\alias{PCR}
\alias{PCR,DesignSettings-method}
\alias{PCR<-}
\alias{PCR<-,DesignSettings-method}
\alias{PCR,DesignSettings-method}
\title{Getter/Setter for the PCR Conditions.}
\usage{
PCR(x)

\S4method{PCR}{DesignSettings}(x)

PCR(x) <- value

\S4method{PCR}{DesignSettings}(x) <- value
}
\arguments{
\item{x}{A \code{DesignSettings} object.}

\item{value}{A named list providing PCR conditions 
The permissible fields of the list and their types
are documented in the \code{\link{PCR_Conditions}} class.}
}
\description{
Gets the PCR conditions that are defined in the provided
\code{DesignSettings} object \code{x}.

Sets the PCR conditions that are defined in the provided
\code{DesignSettings} object \code{x}.
}
\examples{
# Load some settings
data(Ippolito)
# View the active PCR conditions
PCR(settings)
# Evaluate primers with a fixed annealing temperature
PCR(settings)$annealing_temperature <- 50 # celsius
# View available PCR conditions
settings
}
\seealso{
Other settings functions: \code{\link{ConstraintOptions-class}},
  \code{\link{ConstraintSettings-class}},
  \code{\link{CoverageConstraints-class}},
  \code{\link{DesignSettings-class}},
  \code{\link{PCR_Conditions-class}},
  \code{\link{conOptions}}, \code{\link{constraintLimits}},
  \code{\link{constraints}}, \code{\link{cvg_constraints}},
  \code{\link{read_settings}}
}
\keyword{Settings}
