% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\docType{class}
\name{PCR_Conditions-class}
\alias{PCR_Conditions-class}
\alias{PCR_Conditions}
\title{Class for PCR Conditions.}
\value{
A \code{PCR_Conditions} object.
}
\description{
The \code{PCR_Conditions} class encapsulates the PCR conditions
for the computation of primer properties.
}
\section{Slots}{

\describe{
\item{\code{status}}{Named boolean vector indicating 
which of the possible options are active (\code{TRUE})
and which are not (\code{FALSE}).}

\item{\code{settings}}{Named list with constraint options.
The following fields are possible: 
\describe{
\item{\code{use_taq_polymerase}:}{A logical identifying whether you are performing PCR with a Taq polymerase (\code{TRUE}) or not (\code{FALSE}).}
\item{\code{annealing_temp}:}{The annealing temperature in Celsius that is to be used for evaluating the
constraints defined in the \code{\link{ConstraintSettings}} object.
If the annealing temperature field
is not provided, a suitable annealing temperature is automatically computed using a rule of thumb (i.e. subtracting 5 from the melting temperature).}
\item{\code{Na_concentration}:}{The molar concentration of monovalent sodium ions.}
\item{\code{Mg_concentration}:}{The molar concentration of divalent magnesium ions.}
\item{\code{K_concentration}:}{The molar concentration of monovalent potassium ions.}
\item{\code{Tris_concentration}:}{The molar concentration of the Tris(hydroxymethyl)-aminomethan buffer. 
Note that this value is the buffer concentration. To determine corresponding Tris ion concentration, the value of the buffer concentration is halved.}
\item{\code{primer_concentration}:}{The molar concentration of the PCR primers.}
\item{\code{template_concentration}:}{The molar concentration of the PCR templates.}
}}
}}

\examples{
# Initialize a new 'PCR_Conditions' object:
PCR.conditions <- new("PCR_Conditions")
# Retrieving the PCR conditions from a 'DesignSettings' object:
data(Ippolito) # loads a 'DesignSettings' object into 'settings'
PCR(settings)
# Modifying the PCR conditions:
PCR(settings)$use_taq_polymerase <- FALSE
}
\seealso{
Other settings functions: \code{\link{ConstraintOptions-class}},
  \code{\link{ConstraintSettings-class}},
  \code{\link{CoverageConstraints-class}},
  \code{\link{DesignSettings-class}}, \code{\link{PCR}},
  \code{\link{conOptions}}, \code{\link{constraintLimits}},
  \code{\link{constraints}}, \code{\link{cvg_constraints}},
  \code{\link{read_settings}}
}
\keyword{Classes}
