% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_algo.R
\name{plot.all.filtering.stats}
\alias{plot.all.filtering.stats}
\title{Plot Filtering Stats.}
\usage{
\method{plot}{all.filtering.stats}(results.loc, sample, excluded.df,
  filtered.stats, stats.relax, template.df)
}
\arguments{
\item{results.loc}{Location where the filtering results are stored.}

\item{sample}{Primer design run identifier.}

\item{excluded.df}{Data frame with excluded primers.}

\item{filtered.stats}{Filtering statistics data frame.}

\item{stats.relax}{Filtering statistics after relaxation.}

\item{template.df}{Template data frame.}
}
\value{
Write-out of filtering results.
}
\description{
Plots filtering statistics.
}
\keyword{internal}
