% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_template_secondary_structures.R
\name{plot.Delta.DeltaG}
\alias{plot.Delta.DeltaG}
\title{Delta DeltaG Plot}
\usage{
\method{plot}{Delta.DeltaG}(constrained.foldings, stratify = FALSE)
}
\arguments{
\item{constrained.foldings}{Data frame with info from constrained foldings.}

\item{stratify}{Stratify according to template groups?}
}
\value{
Plot of Delta DeltaG.
}
\description{
Plots the difference between the free energy of constrained 
and unconstrained foldings.
}
\keyword{internal}
