% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\name{plot.dimer.dist}
\alias{plot.dimer.dist}
\title{Plot Dimer DeltaG}
\usage{
\method{plot}{dimer.dist}(dimer.data, deltaG.cutoff)
}
\arguments{
\item{dimer.data}{Data frame with dimerization information.}

\item{deltaG.cutoff}{Free energy cutoff for dimerization.}
}
\value{
A plot of dimerization free energies.
}
\description{
Plot the distribution of dimerization free energies.
}
\keyword{internal}
