% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{plot.excluded.hist}
\alias{plot.excluded.hist}
\title{Plot of Excluded Primers}
\usage{
\method{plot}{excluded.hist}(excluded.df, filtered.stats, template.df)
}
\arguments{
\item{excluded.df}{Data frame with excluded primers.}

\item{filtered.stats}{Data frame with statistics of the filtering procedure.}

\item{template.df}{Template data frame.}
}
\value{
A plot of excluded primers.
}
\description{
Plots histogram of excluded primers.
}
\keyword{internal}
