% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_filtering.R
\name{plot.filtering.runtime}
\alias{plot.filtering.runtime}
\title{Plot Filtering Runtimes}
\usage{
\method{plot}{filtering.runtime}(filtered.stats)
}
\arguments{
\item{filered.stats}{Stats from filtering.}
}
\value{
A plot of the runtime for each filtering step.
}
\description{
Plots the runtimes of individual evaluation steps in the filtering procedure.
}
\keyword{internal}
