% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_filtering.R
\name{plot.filtering.stats.cvg}
\alias{plot.filtering.stats.cvg}
\title{Plot of Filtering Stats for Coverage.}
\usage{
\method{plot}{filtering.stats.cvg}(stats, stats.relax = NULL)
}
\arguments{
\item{stats}{Statistics of the filtering procedure.}

\item{stats.relax}{Statistic of the relaxation procedure.}
}
\value{
A plot showing the possible coverage after each filtering step.
}
\description{
Plots the remaining coverage after each filtering step.
}
\keyword{internal}
