% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_filtering.R
\name{plot.filtering.stats}
\alias{plot.filtering.stats}
\title{Plot of Overall Filtering Stats.}
\usage{
\method{plot}{filtering.stats}(stats, stats.relax = NULL)
}
\arguments{
\item{stats}{Statistics on the filtering procedure}

\item{stats.relax}{Statistic on the filtering procedure after relaxation.}
}
\value{
A plot for the number of primers after filtering.
}
\description{
Plots the number of primers remaining after each filtering step.
}
\keyword{internal}
