% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{plot_conservation}
\alias{plot_conservation}
\title{Plot of Template Sequence Conservation.}
\usage{
plot_conservation(entropy.df, alignments, template.df, gap.char = "-")
}
\arguments{
\item{entropy.df}{A data frame with entropies.
Each row gives the entropies of a group of related template
sequences for all columns of the alignment.}

\item{alignments}{A list with \code{DNABin} alignment objects
corresponding to the groups (rows) in the alignment.}

\item{template.df}{The \code{Templates} object for which 
the conservation has been determined.}

\item{gap.char}{The gap char in the alignments. By default,
\code{gap.char} is set to "-".}
}
\value{
A plot showing the degree of sequence conservation in the templates.
}
\description{
Plots the template sequence conservation (range [0,1]) according to
the Shannon entropy of the sequences.
}
\note{
Computing the conservation scores for the plot requires the MAFFT software for multiple alignments (http://mafft.cbrc.jp/alignment/software/).
}
\examples{
\dontrun{
data(Ippolito)
# Select binding regions for every group of templates and plot:
template.df <- select_regions_by_conservation(template.df, win.len = 30)
plot_conservation(attr(template.df, "entropies"), attr(template.df, "alignments"), template.df)
# Select binding regions for all templates and plot:
data(Ippolito)
template.df <- select_regions_by_conservation(template.df, by.group = FALSE)
plot_conservation(attr(template.df, "entropies"), attr(template.df, "alignments"), template.df)
} 
}
