% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\docType{methods}
\name{plot_constraint,Primers-method}
\alias{plot_constraint,Primers-method}
\title{Histogram of Constraint Values.}
\usage{
\S4method{plot_constraint}{Primers}(primers, settings, active.constraints)
}
\arguments{
\item{primers}{An evaluated object of class \code{Primers}.}

\item{settings}{A \code{DesignSettings} object containing the
settings for the constraints to be plotted.}

\item{active.constraints}{Identifiers of constraints to be plotted.
provided settings are used to visualize the desired ranges of constraints.
If \code{active.constraints} is not provided, the plotting method
will automatically try to plot all constraints defined in \code{settings}.}
}
\value{
A histogram of constraint values for the properties specified by
\code{constraints}.
}
\description{
Plots a histogram of constraint values.
}
\keyword{internal}
