% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\name{plot_constraint.histogram.nbr.mismatches}
\alias{plot_constraint.histogram.nbr.mismatches}
\title{Histogram of Number of Mismatches.}
\usage{
plot_constraint.histogram.nbr.mismatches(primer.df, allowed.mismatches)
}
\arguments{
\item{primer.df}{Primer data frame.}

\item{allowed.mismatches}{Number of allowed mismatches.}
}
\value{
A plot of the number of primer mismatches.
}
\description{
Plots a histogram of mismatches.
}
\keyword{internal}
