% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\name{plot_constraint.histogram.primer.efficiencies}
\alias{plot_constraint.histogram.primer.efficiencies}
\title{Histogram of Efficiencies}
\usage{
plot_constraint.histogram.primer.efficiencies(primer.df, opti.constraints)
}
\arguments{
\item{primer.df}{Primer data frame.}

\item{opti.constraints}{List with constraint settings.}
}
\value{
A plot of primer efficiencies.
}
\description{
Plots a histogram of primer efficiencies.
}
\keyword{internal}
