% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\name{plot_constraint}
\alias{plot_constraint}
\title{Plot of Constraint Values.}
\usage{
plot_constraint(primers, settings,
  active.constraints = names(constraints(settings)), ...)
}
\arguments{
\item{primers}{Either an evaluated object of class \code{Primers}
or a list of \code{Primers} objects.}

\item{settings}{A \code{DesignSettings} object containing the
settings for the constraints to be plotted.}

\item{active.constraints}{Identifiers of constraints to be plotted.
If \code{active.constraints} is not provided, the plotting method
automatically plots all constraints defined in \code{settings}
that are annotated in \code{primers}.}

\item{...}{\code{highlight.set} (a character vector identifying the set
that is to be highlighted when \code{primers} is a list).}
}
\value{
A plot showing the distribution of primer properties.
}
\description{
Shows the distribution of the primer properties.
The current constraint settings are indicated with dashed lines in the plot.
}
\examples{
# Plot histogram of constraints for a single primer set
data(Ippolito)
plot_constraint(primer.df, settings)
# Compare constraints across multiple primer sets
data(Comparison)
plot_constraint(primer.data, settings)
}
\seealso{
Other constraint visualizations: \code{\link{plot_constraint_deviation}},
  \code{\link{plot_constraint_fulfillment}}
}
