% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\docType{methods}
\name{plot_constraint_deviation,list-method}
\alias{plot_constraint_deviation,list-method}
\title{Plot of Constraint Deviations for Multiple Primer Sets.}
\usage{
\S4method{plot_constraint_deviation}{list}(primer.data, settings,
  active.constraints, deviation.per.primer = FALSE)
}
\arguments{
\item{settings}{A \code{DesignSettings} object
containing the target ranges for the primer properties.}

\item{active.constraints}{Constraint identifiers to be plotted.}

\item{deviation.per.primer}{Whether to show the deviation per primer or per constraint.}

\item{constraint.df}{An evaluated object of class \code{Primers}.}
}
\value{
A boxplot of deviations
}
\description{
Plots a box plot of the absolute mean deviation of each primer 
for comparing multiple primer sets.
}
\keyword{internal}
