% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\name{plot_constraint_deviation}
\alias{plot_constraint_deviation}
\title{Plot of Constraint Deviations.}
\usage{
plot_constraint_deviation(primer.data, settings,
  active.constraints = names(constraints(settings)), ...)
}
\arguments{
\item{primer.data}{An evaluated object of class \code{Primers} or a list with \code{Primers} objects.}

\item{settings}{A \code{DesignSettings} object
containing the target ranges for the primer properties.}

\item{active.constraints}{Constraint identifiers to be plotted. By default,
all constraints found in \code{settings} are plotted.}

\item{...}{\code{deviation.per.primer} (a boolean indicating whether 
the deviations should be plotted per primer rather than per constraint
if \code{primer.data} is a list)}
}
\value{
A plot showing the deviations of the primer properties from the targets.
}
\description{
Shows the deviation of primer properties from
the target ranges.
}
\details{
Deviations are computed in the following way. Let the
minimum and maximum allowed constraint values be given by
the interval \eqn{[s, e]} and the observed value be \eqn{p}. Then,
if \eqn{p < s}, we output \eqn{-p/|s|}, if \eqn{p > e} we output \eqn{p/|e|},
and otherwise, i.e. if \eqn{s <= p <= e}, we output 0.
}
\examples{
# Deviations for a single primer set
data(Ippolito)
plot_constraint_deviation(primer.df, settings)
# Deviations for multiple primer sets
data(Comparison)
plot_constraint_deviation(primer.data, settings)
}
\seealso{
Other constraint visualizations: \code{\link{plot_constraint_fulfillment}},
  \code{\link{plot_constraint}}
}
