% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\docType{methods}
\name{plot_constraint_fulfillment,Primers-method}
\alias{plot_constraint_fulfillment,Primers-method}
\title{Overview of Constraint Fulfillment.}
\usage{
\S4method{plot_constraint_fulfillment}{Primers}(primers, settings,
  active.constraints, plot.p.vals = TRUE)
}
\arguments{
\item{primers}{A \code{Primers} object.}

\item{settings}{A \code{DesignSettings} object.}

\item{active.constraints}{The identifiers of constarints to be plotted for fulfillment.}

\item{plot.p.vals}{Show p-value from Fisher's exact test 
for the significance
of primer constraint fulfillment in comparison to reference
primer sets.}
}
\value{
A data frame with statistics on fulfilled constraints.
}
\description{
Plots an overview of which primers passed the filtering constraints and which
primers did not.
}
\keyword{internal}
