% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\name{plot_constraint_fulfillment}
\alias{plot_constraint_fulfillment}
\title{Constraint Fulfillment Plot.}
\usage{
plot_constraint_fulfillment(primers, settings,
  active.constraints = names(constraints(settings)), plot.p.vals = FALSE,
  ...)
}
\arguments{
\item{primers}{Either an object of class \code{Primers} or a list of such objects.}

\item{settings}{A \code{DesignSettings} object containing
the constraints to be evaluated.}

\item{active.constraints}{The identifiers of constraints to be plotted for fulfillment. By default \code{active.constraints} is set according to 
all active constarints defined in \code{settings}.}

\item{plot.p.vals}{An optional logical argument indicating whether
p-values computed via \code{\link{primer_significance}} should be annotated in the plot. The default is \code{FALSE}.}

\item{...}{The optional arguments \code{ncol} (a numeric indicating the number of facet columns if \code{primers} is a list),
\code{highlight.set} (the identifier of the primer set to be highlighted if \code{primers} is a list)}
}
\value{
A plot indicating the constraints that fulfilled by the input primers.
}
\description{
Visualizes which which primers pass the constraint settings and which
primers break the constraints.
}
\examples{
# Plot fulfillment for a single primer set:
data(Ippolito)
plot_constraint_fulfillment(primer.df, settings)
# Plot fulfillment for multiple primer sets:
data(Comparison)
plot_constraint_fulfillment(primer.data[1:5], settings)
}
\seealso{
Other constraint visualizations: \code{\link{plot_constraint_deviation}},
  \code{\link{plot_constraint}}
}
