% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\docType{methods}
\name{plot_cvg_constraints,list-method}
\alias{plot_cvg_constraints,list-method}
\title{Plot for Comparing Primer Coverage Constraints.}
\usage{
\S4method{plot_cvg_constraints}{list}(primers, settings, active.constraints,
  highlight.set = NULL)
}
\arguments{
\item{primers}{List with objects of class \code{Primers}.}

\item{settings}{A \code{DesignSettings} object.}

\item{active.constraints}{Names of the coverage constraints to be plotted.}

\item{highlight.set}{Primer sets to highlight in the plot.}
}
\value{
Plot of primer coverage constraints for multiple sets.
}
\description{
Plot for Comparing Primer Coverage Constraints.
}
\keyword{internal}
