% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\docType{methods}
\name{plot_cvg_constraints,Primers-method}
\alias{plot_cvg_constraints,Primers-method}
\title{Histogram of Coverage Constraints.}
\usage{
\S4method{plot_cvg_constraints}{Primers}(primers, settings, active.constraints)
}
\arguments{
\item{primers}{A \code{Primers} object.}

\item{settings}{A \code{DesignSettings} object.}

\item{active.constraints}{Names of coverage constraints to be plotted.}
}
\value{
A plot of coverage constraints.
}
\description{
Plots a histogram of coverage constraint values.
}
\keyword{internal}
