% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\name{plot_cvg_constraints}
\alias{plot_cvg_constraints}
\title{Plot of Coverage Constraints.}
\usage{
plot_cvg_constraints(primers, settings,
  active.constraints = names(cvg_constraints(settings)), ...)
}
\arguments{
\item{primers}{A \code{Primers} object or a list with objects of class \code{Primers}.}

\item{settings}{A \code{DesignSettings} object.}

\item{active.constraints}{Names of coverage constraints to be plotted.
By default, all active coverage constraints in \code{settings} are plotted.}

\item{...}{\code{highlight.set} (a character vector identifying the set
that is to be highlighted when \code{primers} is a list).}
}
\value{
A plot showing the distribution of the coverage constraint values.
}
\description{
Plots the distribution of the coverage constraint values.
}
\examples{
# Plot coverage constraints of a single primer set
data(Ippolito)
plot_cvg_constraints(primer.df, settings)
# Plot coverage constraints for mulitple primer sets
data(Comparison)
plot_cvg_constraints(primer.data, settings)
}
