% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{plot_cvg_vs_set_size}
\alias{plot_cvg_vs_set_size}
\title{Plot of Primer Coverage Ratio vs Set Size.}
\usage{
plot_cvg_vs_set_size(primer.data, template.data, show.labels = TRUE,
  highlight.set = NULL)
}
\arguments{
\item{primer.data}{List with objects of class \code{Primers} containing
the primer sets that are to be compared.}

\item{template.data}{List with objects of class \code{Templates} containing
the templates corresponding to \code{primer.data}.}

\item{show.labels}{Whether the identifiers of the primer sets
should be annotated in the plot. The default is \code{TRUE}.}

\item{highlight.set}{A character vector providing the identifiers
of primer sets to highlight. By default, \code{highlight.set} is \code{NULL}
such that no highlighting takes place.}
}
\value{
A plot of coverage vs set size.
}
\description{
Plots the coverage ratios of the input primer sets 
against the size of the sets.
}
\examples{
data(Comparison)
plot_cvg_vs_set_size(primer.data, template.data)
}
\seealso{
Other comparison visualizations: \code{\link{plot_penalty_vs_set_size}}
}
