% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{plot_penalty_vs_set_size}
\alias{plot_penalty_vs_set_size}
\title{Plot of Primer Penalties vs Set Size.}
\usage{
plot_penalty_vs_set_size(primer.data, settings,
  active.constraints = names(constraints(settings)), alpha = 0)
}
\arguments{
\item{primer.data}{List with objects of class \code{Primers}.}

\item{settings}{An object of class \code{DesignSettings}.}

\item{active.constraints}{A character vector with constraint identifiers
to be considered for generating the plot.}

\item{alpha}{A numeric in the range [0,1] defining the trade-off between
the maximal deviation of a constraint (large code{alpha}) and
all constraint deviations (large \code{alpha}).
By default, \code{alpha} is set to 0 such that the absolute
deviation across all constraints is considered.}
}
\value{
A plot showing the association between 
primer penalties and the size of the primer sets.
}
\description{
Plots the penalties of the input primer sets 
against the number of primers contained in each set.
The penalties are computed using \code{\link{score_primers}}
where more information is provided on how to set \code{alpha}.
}
\examples{
data(Comparison)
plot_deviation_vs_set_size(primer.data, settings)
}
\seealso{
Other comparison visualizations: \code{\link{plot_cvg_vs_set_size}}
}
