% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_comparison.R
\name{plot_primer.comparison.mismatches}
\alias{plot_primer.comparison.mismatches}
\title{Plot Primer Mismatches}
\usage{
plot_primer.comparison.mismatches(primer.data, template.data,
  allowed.mismatches, highlight.set = NULL)
}
\arguments{
\item{primer.data}{List with primer data frames.}

\item{template.data}{List with template data frames.}

\item{allowed.mismatches}{Allowed mismatches.}

\item{highlight.set}{Primer sets to highlight in the plot.}
}
\value{
Plot of mismatches for comparison.
}
\description{
Plots primer mismatches for every set.
}
\keyword{internal}
