% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{plot_primer}
\alias{plot_primer}
\title{Primer View Plot.}
\usage{
plot_primer(primer.df, template.df, identifier = NULL, relation = c("fw",
  "rev"), region.names = c("Binding region", "Amplification region"))
}
\arguments{
\item{primer.df}{An object of class \code{Primers} containing
primers with evaluated primer coverage.}

\item{template.df}{An object of class \code{Templates} with template sequences
corresponding to \code{primer.df}.}

\item{identifier}{Identifiers of primers that are to be considered.
If \code{identifier} is set to \code{NULL} (the default), all primers are considered.}

\item{relation}{Compute binding positions relative to forward ("fw") or reverse ("rev") binding regions.
The default is "fw".}

\item{region.names}{Character vector of length 2 providing the names
of the binding and amplification region.}
}
\value{
A plot of primer binding sites in the templates.
}
\description{
Visualizes the binding positions of every primer relative to
the target binding region in the corresponding template sequences.
}
\examples{
data(Ippolito)
plot_primer(primer.df[1,], template.df[1:30,])
}
\seealso{
Other coverage visualizations: \code{\link{plot_primer_binding_regions}},
  \code{\link{plot_primer_cvg}},
  \code{\link{plot_primer_subsets}},
  \code{\link{plot_template_cvg}}
}
