% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\docType{methods}
\name{plot_primer_binding_regions,list,list-method}
\alias{plot_primer_binding_regions,list,list-method}
\title{Plot of Primer Binding Regions for Multiple Sets.}
\usage{
\S4method{plot_primer_binding_regions}{list,list}(primers, templates,
  direction = c("both", "fw", "rev"), group = NULL, relation = c("fw",
  "rev"), region.names = c("Binding region", "Amplification region"),
  highlight.set = NULL)
}
\arguments{
\item{primers}{List with primer data frames.}

\item{templates}{List with template data frames.}

\item{direction}{Direction of primers.}

\item{group}{Template groups to plot.
This defaults to plotting all groups.}

\item{relation}{Plot binding region relative to forward binding region or reverse?}

\item{region.names}{Names for the primer binding region and the amplified region.}

\item{highlight.set}{Primer sets to highlight in the plot.}
}
\value{
A plot for primer binding region comparison.
}
\description{
Plots the primer binding regions for every primer set.
}
\keyword{internal}
