% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{plot_primer_binding_regions}
\alias{plot_primer_binding_regions}
\title{Plot of Primer Binding Regions.}
\usage{
plot_primer_binding_regions(primers, templates, direction = c("both", "fw",
  "rev"), group = NULL, relation = c("fw", "rev"),
  region.names = c("Binding region", "Amplification region"), ...)
}
\arguments{
\item{primers}{Either a single \code{Primers} object or 
a list with \code{Primers} objects.}

\item{templates}{If \code{primers} is a \code{primers} object,
please supply a \code{Templates} object.
If \code{primers} is a list, please supply a corresponding list of
\code{Templates} objects.}

\item{direction}{The directionality of primers to be plotted. This can either
be "both" to plot primers of both directions (the default), "fw" to plot
only forward primers, or "rev" to plot only reverse primers.}

\item{group}{Optional identifiers of template groups for which binding events should
be determined. By default, \code{group} is set to \code{NULL} such that
all templates are considered.}

\item{relation}{An optional character vector specifying whether binding region data shall
be plotted relative to the forward (\code{fw}) or reverse (\code{rev}) 
target binding regions.}

\item{region.names}{An optional, two-component character vector specifying
the identifiers for the primer binding region and the amplified region.}

\item{...}{\code{highlight.set} (the identifiers of primer sets
to be highlighted, if \code{primers} is a list)}
}
\value{
A plot for primer binding region comparison.
}
\description{
Visualizes the number of binding events of the input primers
with respect to the allowed binding regions in the templates.
}
\examples{
# Primer binding regions of a single primer set
data(Ippolito)
plot_primer_binding_regions(primer.df, template.df)
# Primer binding regions of multiple primer sets
data(Comparison)
plot_primer_binding_regions(primer.data[1:3], template.data[1:3])
}
\seealso{
Other coverage visualizations: \code{\link{plot_primer_cvg}},
  \code{\link{plot_primer_subsets}},
  \code{\link{plot_primer}},
  \code{\link{plot_template_cvg}}
}
