% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\docType{methods}
\name{plot_primer_cvg,list,list-method}
\alias{plot_primer_cvg,list,list-method}
\title{Plot Multiple Primer Coverages.}
\usage{
\S4method{plot_primer_cvg}{list,list}(primers, templates,
  per.mismatch = FALSE)
}
\arguments{
\item{primers}{List with \code{Primers} objects.}

\item{templates}{List with \code{Templates} objects.}

\item{per.mismatch}{Whether the coverage should
be broken down for individual settings of allowed mismatches.}
}
\value{
A bar plot showing the coverage of individual primers.
}
\description{
Plots the coverage of individual primers for multiple sets.
}
\keyword{internal}
