% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\docType{methods}
\name{plot_primer_cvg,Primers,Templates-method}
\alias{plot_primer_cvg,Primers,Templates-method}
\title{Plot Individual Primer Coverage.}
\usage{
\S4method{plot_primer_cvg}{Primers,Templates}(primers, templates,
  per.mismatch = FALSE, groups = NULL)
}
\arguments{
\item{per.mismatch}{Whether to stratify by allowed mismatches.}

\item{p.df}{Primer data frame.}

\item{template.df}{Template data frame.}

\item{excluded.seqs}{Identifiers of templates that should not be considered.}

\item{per.mismatch}{Whether the coverage should
be broken down for individual settings of allowed mismatches.}
}
\value{
A bar plot showing the coverage of individual primers.
}
\description{
Shows which templates are covered by individual primers.
}
\keyword{internal}
