% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{plot_primer_cvg}
\alias{plot_primer_cvg}
\title{Plot of Primer Coverage.}
\usage{
plot_primer_cvg(primers, templates, per.mismatch = FALSE, ...)
}
\arguments{
\item{primers}{An object of class \code{Primers} or a list with with objects
of class \code{Primers}.}

\item{templates}{If \code{primers} is an object of class \code{Primers},
please supply a \code{Templates} object. If \code{primers} is a list,
please supply a corresponding list with \code{Templates} objects.}

\item{per.mismatch}{A logical identifiying whether the coverage should
be plotted for individual settings of allowed mismatches. By default
\code{per.mismatch} is set to \code{FALSE} such that the overall
coverage is plotted.}

\item{...}{\code{groups} (the identifiers of template groups
to be excluded from the plot if \code{primers} is a single primer set)}
}
\value{
A plot showing the coverage of individual primers.
}
\description{
Shows which groups of templates are covered by individual primers.
}
\examples{
# Plot expected coverage per primer
data(Ippolito)
plot_primer_cvg(primer.df, template.df)
# Plot coverage stratified by allowed mismatches:
plot_primer_cvg(primer.df, template.df, per.mismatch = TRUE)
# Plot coverage of multiple primer sets
data(Comparison)
plot_primer_cvg(primer.data[1:3], template.data[1:3])
}
\seealso{
Other coverage visualizations: \code{\link{plot_primer_binding_regions}},
  \code{\link{plot_primer_subsets}},
  \code{\link{plot_primer}},
  \code{\link{plot_template_cvg}}
}
