% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{plot_primer_cvg_unstratified}
\alias{plot_primer_cvg_unstratified}
\title{Plot Individual Primer Coverage.}
\usage{
plot_primer_cvg_unstratified(p.df, template.df, groups = NULL)
}
\arguments{
\item{p.df}{Primer data frame.}

\item{template.df}{Template data frame.}

\item{groups}{Optional identifiers of template groups to be considered.
If not provided, all template groups are considered.}
}
\value{
A bar plot showing the coverage of individual primers.
}
\description{
Plots the coverage of individual primers.
}
\keyword{internal}
