% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{plot_primer_subsets}
\alias{plot_primer_subsets}
\title{Plot of Primer Subset Coverage.}
\usage{
plot_primer_subsets(primer.subsets, template.df, required.cvg = 1)
}
\arguments{
\item{primer.subsets}{A list with optimal primer subsets, each of class \code{Primers}. The \emph{k}-th list entry should correspond to an object of class \code{Primers}
representing the primer subset of size \emph{k} whose coverage ratio
is the largest among all possible subsets of size \emph{k}.}

\item{template.df}{An object of class \code{Templates} containing the
template sequences corresponding to the primers specified in \code{primer.subsets}.}

\item{required.cvg}{The required coverage ratio.
The default is 100\%; this value is plotted as a horizontal line.}
}
\value{
Plot of the coverages of the primer subsets in \code{primer.subsets}.
}
\description{
Visualizes the coverage of optimized primer subsets.
}
\details{
The input for the \code{primer.subsets} argument can be computed using
\code{\link{subset_primer_set}}. 
The line plot indicates the ratio of covered templates when considering
all primers in a primer set of a given size. The bar plots indicate
the coverage ratios of individual primers in a set. The target coverage
ratio is indicated by a horizontal line. Bars exceeding the target ratio
possibly indicate the existence of redundant coverage events.
}
\examples{
data(Ippolito)
primer.subsets <- subset_primer_set(primer.df, template.df)
plot_primer_subsets(primer.subsets, template.df)
}
\seealso{
Other coverage visualizations: \code{\link{plot_primer_binding_regions}},
  \code{\link{plot_primer_cvg}}, \code{\link{plot_primer}},
  \code{\link{plot_template_cvg}}
}
