% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\docType{methods}
\name{plot_template_cvg,list,list-method}
\alias{plot_template_cvg,list,list-method}
\title{Templates Coverage for Multiple Primer Sets.}
\usage{
\S4method{plot_template_cvg}{list,list}(primers, templates, per.mismatch,
  highlight.set = NULL)
}
\arguments{
\item{primers}{List with primer data frames.}

\item{templates}{List with template data frames.}

\item{highlight.set}{Primer sets to be highlighted.}

\item{colors}{Color for every primer set.}
}
\value{
A plot for comparing primer coverage.
}
\description{
Plots the coverage of multiple primer sets.
}
\keyword{internal}
