% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{plot_template_cvg}
\alias{plot_template_cvg}
\title{Bar Plot of Template Coverage.}
\usage{
plot_template_cvg(primers, templates, per.mismatch = FALSE, ...)
}
\arguments{
\item{primers}{Either a \code{Primers} object with evaluated primer coverage
or a list containing \code{Primers} objects.}

\item{templates}{If \code{primers} is a \code{Primers} object, \code{templates} should be a \code{Templates} object.
If \code{primers} is a list, then \code{templates} should be a list of \code{Templates} objects.}

\item{per.mismatch}{A logical specifying whether the visualization should be stratified
according to the allowed number of mismatches. By default,
\code{per.mismatch} is set to \code{FALSE} such that the overall coverage
is plotted.}

\item{...}{Optional arguments \code{groups} (a character vector of groups to be plotted when \code{primers} is a single primer set), \code{highlight.set} (the identifier of a primer set to be highlighted when \code{primers} is a list)}
}
\value{
A plot showing the number of covered template sequences.
}
\description{
Creates a bar plot showing the coverage for every group of template sequences.
}
\examples{
# Visualize the template coverage of a single primer set
data(Ippolito)
plot_template_cvg(primer.df, template.df)
# Stratify by allowed mismatches:
plot_template_cvg(primer.df, template.df, per.mismatch = TRUE)
# Compare the coverage of multiple primer sets
data(Comparison)
plot_template_cvg(primer.data[1:3], template.data[1:3])
# Stratify by allowed mismatches:
plot_template_cvg(primer.data[1:3], template.data[1:3], per.mismatch = TRUE)
}
\seealso{
Other coverage visualizations: \code{\link{plot_primer_binding_regions}},
  \code{\link{plot_primer_cvg}},
  \code{\link{plot_primer_subsets}},
  \code{\link{plot_primer}}
}
