% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{plot_template_cvg_comparison_mismatch}
\alias{plot_template_cvg_comparison_mismatch}
\title{Templates Coverage for Multiple Primer Sets.}
\usage{
plot_template_cvg_comparison_mismatch(primers, templates,
  highlight.set = NULL, show.percentages = FALSE)
}
\arguments{
\item{primers}{List with primer data frames.}

\item{templates}{List with template data frames.}

\item{highlight.set}{Primer sets to be highlighted.}

\item{show.percentages}{Whether to show the total coverage percentage above the bars.}
}
\value{
A plot for comparing primer coverage.
}
\description{
Plots the coverage of multiple primer sets.
}
\keyword{internal}
