% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{plot_template_cvg_comparison_unstratified}
\alias{plot_template_cvg_comparison_unstratified}
\title{Templates Coverage for Multiple Primer Sets.}
\usage{
plot_template_cvg_comparison_unstratified(primers, templates,
  highlight.set = NULL)
}
\arguments{
\item{primers}{List with primer data frames.}

\item{templates}{List with template data frames.}

\item{highlight.set}{Primer sets to be highlighted.}
}
\value{
A plot for comparing primer coverage.
}
\description{
Plots the coverage of multiple primer sets.
}
\keyword{internal}
