% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{plot_template_cvg_unstratified}
\alias{plot_template_cvg_unstratified}
\title{Bar Plot of Template Coverage.}
\usage{
plot_template_cvg_unstratified(primers, templates, groups = NULL)
}
\arguments{
\item{primers}{A \code{Primers} object with evaluated primer coverage.}

\item{templates}{A \code{Templates} object containing the template sequences.}

\item{groups}{Identifiers of template groups for which plot should be created. By default, \code{groups} is set to \code{NULL} such that all 
templates are considered.
according to the number of mismatches between primer-template pairs.}
}
\value{
A plot showing the number of covered template sequences.
}
\description{
Creates a bar plot showing the coverage for every group of template sequences.
}
\keyword{internal}
