% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\name{plot_template_structure}
\alias{plot_template_structure}
\title{Plot of Template Folding Energies.}
\usage{
plot_template_structure(fold.df)
}
\arguments{
\item{fold.df}{A data frame with free energies for the
template regions.}
}
\value{
A plot of DeltaDeltaG.
}
\description{
Plots the DeltaDeltaG of template folding, which is 
the difference between the free energy change of
the unconstrained folding and the free energy change
of the constrained folding.
}
\keyword{internal}
