% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pos.to.range}
\alias{pos.to.range}
\title{Conversion of Positions to Ranges.}
\usage{
pos.to.range(pos1, pos2)
}
\arguments{
\item{pos1}{The first value.}

\item{pos2}{The second value.}
}
\value{
A character vector range.
}
\description{
Converts two numeric values to a range.
}
\keyword{internal}
