% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{predict_coverage}
\alias{predict_coverage}
\title{Prediction of Primer Coveragee.}
\usage{
predict_coverage(primer.df, template.df, settings, mode = c("on_target",
  "off_target"), updateProgress = NULL)
}
\arguments{
\item{primer.df}{A \code{Primers} data frame.}

\item{template.df}{A \code{Templates} data frame.}

\item{settings}{A \code{DesignSettings} object.}

\item{mode}{Whether on-target or off-target events shall be considered.}
}
\value{
The predictions for primer coverage
}
\description{
Predicts primer coverage using a logistic regression model.
Converts coverage probabilities to expected false positive rate
for a given probability.
}
\keyword{internal}
