% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_primers_tree.R
\name{prefilter.primer.candidates}
\alias{prefilter.primer.candidates}
\title{Identication of Short Primers.}
\usage{
prefilter.primer.candidates(primer.candidates, min.len)
}
\arguments{
\item{primer.candidates}{Primer alignment.}

\item{min.len}{Minimal primer length.}
}
\value{
The index of proposed primers that are shorter than \code{min.len}.
}
\description{
Identify initial primers that are too short.
}
\keyword{internal}
