% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_comparison.R
\name{prepare.constraint.plot}
\alias{prepare.constraint.plot}
\title{Preparation of Comparison Plot for Evaluation.}
\usage{
prepare.constraint.plot(primer.data, constraint.settings, plot.p.vals = FALSE)
}
\arguments{
\item{primer.data}{List with objects of class \code{Primers}. Each 
list entry corresponds to a single primer set.}

\item{constraint.settings}{List with settings for each constraint.
If \code{NULL} (the default), use the available evaluation results
in \code{primer.data}.}

\item{plot.p.vals}{Whether p-values from Fisher's exact test
should be annotated for every primer set.}
}
\value{
Plot indicating the ratio of primers
fulfilling the constraints specified in \code{constraint.settings}
for each primer set in \code{primer.data}.
}
\description{
Preparation of Comparison Plot for Evaluation.
}
\keyword{internal}
