% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{prepare.dimer.seqs}
\alias{prepare.dimer.seqs}
\title{Preparation of Input for Dimerization.}
\usage{
prepare.dimer.seqs(s1, s2)
}
\arguments{
\item{s1}{Nucleotide character vectors (5' to 3')}

\item{s2}{Nucleotide character vectors  (5' to 3')}
}
\value{
A list with two fields containing character vectors.
}
\description{
Preparation of Input for Dimerization.
}
\keyword{internal}
