% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{prepare_mm_plot}
\alias{prepare_mm_plot}
\title{Data Preparation for Mismatch Plot.}
\usage{
prepare_mm_plot(primer.df, template.df, mode = c("on_target", "off_target"))
}
\arguments{
\item{primer.df}{A \code{Primers} object.}

\item{template.df}{A \code{Templates} object.}

\item{mode}{Whether to compute for on-target or off-target events.}
}
\value{
A data frame with binding information for every primer.
}
\description{
Data Preparation for Mismatch Plot.
}
\keyword{internal}
