% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{prepare_template_cvg_mm_data}
\alias{prepare_template_cvg_mm_data}
\title{Preparation of Data for Plotting Mismatch Template Coverage.}
\usage{
prepare_template_cvg_mm_data(primer.df, template.df,
  allowed.mismatches = NULL)
}
\arguments{
\item{primer.df}{A \code{Primers} object.}

\item{template.df}{A \code{Templates} object.}

\item{allowed.mismatches}{An optional numeric specifying the number of mismatches to be considered at most for plotting.
If not provided, the maximal number of mismatches found for the input primer 
set is used.}
}
\value{
A data frame for creating a plot.
}
\description{
Creates a data frame for plotting a bar plot for the covered templates per allowed mismatches.
}
\keyword{internal}
