% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{primer.binding.regions.data}
\alias{primer.binding.regions.data}
\title{Primer Binding Region Data}
\usage{
primer.binding.regions.data(primer.df, template.df, direction = c("both",
  "fw", "rev"), group = NULL, relation = c("fw", "rev"))
}
\arguments{
\item{primer.df}{Primer data frame.}

\item{template.df}{Template data frame.}

\item{direction}{Primer direction}

\item{group}{The groups for which binding data shall be retrieved.}

\item{relation}{Binding region data relative to forward/reverse binding region?}
}
\value{
Data frame with primer binding data.
}
\description{
Collects all data concerning primer binding regions.
}
\keyword{internal}
