% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_ILP.R
\name{primer.coverage.for.groups}
\alias{primer.coverage.for.groups}
\title{Determination of Primer Coverage for Groups.}
\usage{
primer.coverage.for.groups(primer.df, template.df, groups)
}
\arguments{
\item{primer.df}{Primer data frame.}

\item{template.df}{Template data frame.}

\item{groups}{Template groups for which coverage should be determined.}
}
\value{
\code{primer.df} with coverage considered only for \code{groups}.
}
\description{
Modifies a primer data frame to retain only coverage events
relating to the selected groups of templates.
}
\keyword{internal}
