% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\name{primer.set.parameter.stats}
\alias{primer.set.parameter.stats}
\title{Primer Set Statistics}
\usage{
primer.set.parameter.stats(primer.df, mode.directionality, lex.seq)
}
\arguments{
\item{primer.df}{Primer data frame.}

\item{mode.directionality}{Direction of primers.}

\item{lex.seq}{Template data frame.}
}
\value{
A data frame with statistics.
}
\description{
Creates an overview of all primer set constraint values.
}
\keyword{internal}
