% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer_significance.R
\name{primer_significance}
\alias{primer_significance}
\title{Significance of a Primer Set.}
\usage{
primer_significance(primer.df, set.name = NULL, active.constraints = NULL)
}
\arguments{
\item{primer.df}{An object of class \code{Primers} for which the
significance of physicochemical properties shall be determined.}

\item{set.name}{An identifier for the input primers. If \code{NULL},
the run identifier is used.}

\item{active.constraints}{Identifiers of the constraints contained in 
\code{primer.df} to consider.
By default (\code{NULL}) all constraints available in \code{primer.df}
are considered for determining the significance.}
}
\value{
The p-value of the primer set according to Fisher's exact test.
The returned value has the following attributes: 
\describe{
\item{\code{test}}{The results of the significance test}
\item{\code{tab}}{The confusion matrix for Fisher's exact test}
\item{\code{constraints}}{The names of the considered constraints}
}
}
\description{
Uses Fisher's exact test to determine the significance
of a primer set according to its ratio of fulfilled 
constraints on the primer properties.
}
\details{
The significance is computed by comparing
the total count of fulfilled and failed constraints
with the corresponding counts of primer sets from the literature.
Significant p-values indicate primer sets whose rate of constraint 
fulfillment is higher compared to the reference sets.
}
\examples{
data(Ippolito)
p.data <- primer_significance(primer.df, "Ippolito")
attr(p.data,"tab") # the confusion matrix
attr(p.data, "test") # results from Fisher's test
attr(p.data, "constraints") # considered constraints for the test
}
\seealso{
Other primer functions: \code{\link{Primers-class}},
  \code{\link{check_constraints}},
  \code{\link{check_restriction_sites}},
  \code{\link{create_report}},
  \code{\link{design_primers}},
  \code{\link{filter_primers}},
  \code{\link{get_initial_primers}},
  \code{\link{score_degen}}, \code{\link{write_primers}}
}
