% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\docType{class}
\name{Primers-class}
\alias{Primers-class}
\alias{Primers}
\title{The Primers Class.}
\usage{
Primers(...)
}
\arguments{
\item{...}{A data frame fulfilling the structural
requirements for initializing a \code{Primers} object.}
}
\value{
A \code{Primers} object, an instance of a data frame.
}
\description{
The \code{Primers} class encapsulates a data frame
representing a set of primers. Objects of this class
store all properties associated with a set of primers,
for example the results from evaluating the properties
of a primer set or from determining its coverage.
}
\section{Basic columns}{

In the following you can find a description of the most
important columns that can be found in objects of class \code{Primers}. 
Note that angular brackets indicate the existence of multiple possibilities.
The following columns are present when a set of primers
is loaded from a FASTA file using \code{\link{read_primers}}:
\describe{
\item{\code{ID}}{The identifiers of the primers.}
\item{\code{Identifier}}{The internal identifiers of the primers.}
\item{\code{Forward}}{The sequences of forward primers.}
\item{\code{Reverse}}{The sequences of reverse primers.}
\item{\code{primer_length<fw|rev>}}{The lengths of
forward and reverse primer sequences, respectively.}
\item{\code{Direction}}{Either 'fw' for forward primers,
'rev' for reverse primers, or 'both' for a primer pair.}
\item{\code{Degeneracy_<fw|rev>}}{The degeneracy (ambiguity) of
forward and reverse primers, respectively.}
\item{\code{Run}}{An identifier describing the primer set.}
}
}

\section{Coverage-related columns}{

The following columns are only available after primer coverage
has been computed, that is after \code{\link{check_constraints}}
has been called with the active \code{primer_coverage} constraint. Computed coverage
values relating solely to string matching are indicated by the prefix
\code{Basic_}, while columns without this prefix relate to the coverage after
applying the constraints formulated via \code{CoverageConstraints}.
Information on off-target coverage events are indicated by
the \code{Off_} prefix, while on-target coverage events do not carry
this prefix.

\describe{
\item{\code{primer_coverage}}{The number of templates that are
covered by the primers. Note that if a primer set contains
primers of both directions, a template is only considered covered
if it is covered by primers of both directions.}
\item{\code{Coverage_Ratio}}{The ratio of templates that are covered by the primers.}
\item{\code{Binding_Position_Start_<fw|rev>}}{The upstream position in the 
templates where forward and reverse primers respectively bind.}
\item{\code{Binding_Position_End_<fw|rev>}}{The downstream position in the templates where forward and reverse primers respectively bind.}
\item{\code{Relative_<Forward|Reverse>_Binding_Position_<Start|End>_<fw|rev>}}{
The binding upstream (\code{Start}) or downstream (\code{End}) positions 
of the primers relative to the forward (\code{Forward})
or reverse (\code{Reverse}) binding regions, either for 
forward (\code{fw}) or reverse primers (\code{rev}).}
\item{\code{Binding_Region_Allowed}}{Whether a coverage event
occurred in the target binding region or not. If the allowed
off-target ratio was set to 0 only coverage events within the 
the target region are reported.}
\item{\code{Nbr_of_mismatches_<fw|rev>}}{The number of mismatches
of forward and reverse primer coverage events, respectively.}
\item{\code{Mismatch_pos_<fw|rev>}}{The position of mismatches
for forward and reverse coverage events, respectively. Mismatch
positions are reported relative to the 3' end, that is, position
1 indicates a mismatch in the last base of a primer.}
\item{\code{primer_specificity}}{The specificity of a primer
as determined by its ratio of off-target binding events.}
}
}

\section{Constraint-related columns}{

Each constraint that is considered when calling \code{\link{check_constraints}}
gives rise to at least one column in the provided \code{Primers} object.
Due to the large number of possible constraints, we will limit our description
to the \code{gc_clamp} constraint. Once the GC clamp property has been computed,
the \code{gc_clamp_fw} column contains the length of the GC clamp for forward 
primers and \code{gc_clamp_rev} the corresponding length for reverse primers.
Whether the desired extent of the GC clamp was obtained by a primer
is indicated by the \code{EVAL_gc_clamp} column. It contains \code{TRUE} when
the GC clamp constraint was fulfilled and \code{FALSE} when it was broken.
To identify whether all required constraints were fulfilled by a primer,
the \code{constraints_passed} column can be used. It contains \code{TRUE}
if all \code{active.constraints} used by \code{\link{check_constraints}} were fulfilled
and \code{FALSE} otherwise.
}

\examples{
primer.location <- system.file("extdata", "IMGT_data", "primers", "IGHV", 
                     "Ippolito2012.fasta", package = "openPrimeR")
primer.df <- read_primers(primer.location, "_fw", "_rev")
}
\seealso{
\code{\link{read_primers}} for loading a primer set,
\code{\link{score_degen}} for scoring the degeneracy of a primer,
\code{\link{primer_significance}} for determining the significance
of a primer set,
\code{\link{get_initial_primers}} for computing an initial set of primers,
\code{\link{design_primers}} for designing primer sets,
\code{\link{check_constraints}} for determining the properties of a primer set,
\code{\link{filter_primers}} for filtering a primer set,
\code{\link{check_restriction_sites}} to search for restriction sites,
\code{\link{get_cvg_ratio}} to determine the coverage ratio of a primer set,
\code{\link{create_report}} to create a PDF report for a primer set.

Other primer functions: \code{\link{check_constraints}},
  \code{\link{check_restriction_sites}},
  \code{\link{create_report}},
  \code{\link{design_primers}},
  \code{\link{filter_primers}},
  \code{\link{get_initial_primers}},
  \code{\link{primer_significance}},
  \code{\link{score_degen}}, \code{\link{write_primers}}
}
\keyword{Classes}
