% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rbind.primer.data}
\alias{rbind.primer.data}
\title{Rbind for Primer Data Frames.}
\usage{
\method{rbind}{primer.data}(primer.data)
}
\arguments{
\item{primer.data}{List with primer data frames.}
}
\value{
A data frame containing all data in \code{primer.data}.
}
\description{
Merges all primer data frames in \code{primer.data} into one data frame.
}
\keyword{internal}
