% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{rbind.Primers}
\alias{rbind.Primers}
\title{rbind for Primers class.}
\usage{
\method{rbind}{Primers}(...)
}
\arguments{
\item{...}{Parameters for rbind function.}
}
\value{
Row-binded Primers data frame.
}
\description{
Ensures that the rbind result has the appropriate class.
}
\examples{
data(Ippolito)
primer.df <- rbind(primer.df, primer.df)
}
\keyword{internal}
