% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{read.leaders}
\alias{read.leaders}
\title{Read Individual Binding Regions}
\usage{
read.leaders(fasta.file, direction = c("fw", "rev"), rm.keywords = NULL,
  gap.character)
}
\arguments{
\item{fasta.file}{Path to a FASTA file with binding regions.}

\item{direction}{String identifying whether the FASTA file
contains information pertaining to the binding region of forward
or reverse primers.}

\item{rm.keywords}{A vector of keywords that are used to remove templates whose headers contain any of the keywords.}

\item{gap.character}{The character for indicating gaps in sequences.}
}
\value{
A data frame with individual binding regions.
}
\description{
Reads individual binding regions into a data frame.
}
\keyword{internal}
