% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_secondary_structures.R
\name{read.secondary.structure.raw}
\alias{read.secondary.structure.raw}
\title{Read a Secondary Structure}
\usage{
read.secondary.structure.raw(fw.out)
}
\arguments{
\item{fw.out}{Path to a ViennaRNA secondary structure output file.}
}
\value{
Data frame with information on secondary structures.
}
\description{
Reads the secondary structure output of ViennaRNA.
}
\keyword{internal}
