% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{read.sequences}
\alias{read.sequences}
\title{Read Sequences.}
\usage{
read.sequences(fasta.file, gap.character)
}
\arguments{
\item{fasta.file}{The path to a FASTA file.}

\item{The}{character indicating gaps in sequences.}
}
\value{
A data frame with sequences.
}
\description{
Reads an input FASTA file.
}
\keyword{internal}
