% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{read_primers.internal}
\alias{read_primers.internal}
\title{Internal Function for Reading Primers}
\usage{
read_primers.internal(primer.seqs, headers, fw.id, rev.id,
  merge.ambig = c("none", "merge", "unmerge"), max.degen, sample.name)
}
\arguments{
\item{primer.seqs}{Primer sequences.}

\item{headers}{Headers of the primer FASTA file.}

\item{fw.id}{Identifier of forward primers in the headers.}

\item{rev.id}{Identifier of reverse primers in the headers.}

\item{merge.ambig}{Should ambiguous primers be merged?}

\item{max.degen}{Maximum allowed degeneracy}
}
\value{
A data frame with primer sequences.
}
\description{
Reads the given primer sequences into a data frame.
}
\keyword{internal}
