% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{read_primers}
\alias{read_primers}
\title{Input of Primers.}
\usage{
read_primers(primer.location, fw.id = "_fw", rev.id = "_rev",
  merge.ambig = c("none", "merge", "unmerge"), max.degen = 16,
  template.df = NULL, adapter.action = c("warn", "rm"),
  sample.name = NULL, updateProgress = NULL)
}
\arguments{
\item{primer.location}{Path to a single or multiple primer FASTA or CSV files.}

\item{fw.id}{For FASTA input, the identifier for forward primers in the FASTA headers.}

\item{rev.id}{For FASTA input, the identifier for reverse primers in the FASTA headers.}

\item{merge.ambig}{Indicates whether similar primers should be merged
("merge") using IUPAC ambiguity codes or whether primers should 
be disambiguated ("unmerge").
By default \code{merge.ambig} is set to "none", leaving primers as they are.}

\item{max.degen}{A scalar numeric providing the maximum
allowed degeneracy for merging primers if
\code{merge.ambig} is set to "merge".
Degeneracy is defined by the number of disambiguated sequences 
that are represented by a degenerate primer.}

\item{template.df}{An object of class \code{Templates}.
If \code{template.df} is provided the primers are checked
for restriction sites upon input. By default
\code{template.df} is \code{NULL} such that the primers are not
checked for restriction sites.}

\item{adapter.action}{The action to be performed when \code{template.df} is
provided for identifying adapter sequences.
Either "warn" to issue warning about adapter sequences or
"rm" to remove identified adapter sequences. The default is "warn".}

\item{sample.name}{An identifier for the input primers.}

\item{updateProgress}{A Shiny progress callback function. This is
\code{NULL} by default such that no progress is tracked.}
}
\value{
An object of class \code{Primers} for a single 
\code{primer.location} or a list of such objects for multiple locations.
}
\description{
Reads one or multiple input files with primer sequences. The input can either be in FASTA
or in CSV format.
}
\details{
The input arguments \code{fw.id}, \code{rev.id}, \code{merge.ambig}, and \code{max.degen} are only used for loading primers from a FASTA file.
If you want to load a FASTA file, please ensure that \code{fw.id} and 
\code{rev.id} are set according to the keywords indicating
the primer directionalities in the FASTA file.
When loading a CSV file, the format of the file should adhere
to the structure defined by the \code{\link{Primers}} class. 
You can easily store a \code{Primers} objects as a CSV file using the
\code{\link{write_primers}} function.
}
\examples{
primer.fasta <- system.file("extdata", "IMGT_data", "primers", "IGHV", 
                     "Ippolito2012.fasta", package = "openPrimeR")
primer.df <- read_primers(primer.fasta, "_fw", "_rev")
# Read multiple FASTA files
fasta.files <- list.files(system.file("extdata", "IMGT_data", "primers", 
                 "IGHV", package = "openPrimeR"), pattern = "*\\\\.fasta",
                 full.names = TRUE)[1:3]
primer.data <- read_primers(fasta.files)
# Read primers from a CSV file
primer.csv <- system.file("extdata", "IMGT_data", "comparison", 
             "primer_sets", "IGL", "IGL_openPrimeR2017.csv",  package = "openPrimeR")
primer.df <- read_primers(primer.csv)
# Read multiple primer CSV files
primer.files <- list.files(path = system.file("extdata", "IMGT_data", "comparison", 
                         "primer_sets", "IGH", package = "openPrimeR"),
                          pattern = "*\\\\.csv", full.names = TRUE)[1:3]
primer.data <- read_primers(primer.files)
# Read a mixture of FASTA/CSV files:
mixed.primers <- c(primer.fasta, primer.csv)
primer.data <- read_primers(mixed.primers)
}
\keyword{Primers}
