% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{read_primers_csv}
\alias{read_primers_csv}
\title{Read Primer CSV File.}
\usage{
read_primers_csv(file)
}
\arguments{
\item{file}{The path to a csv file containing the primer data.}
}
\value{
A \code{Primers} object, an instance of a data frame.
}
\description{
Reads a primer data frame stored in a CSV file.
}
\keyword{internal}
