% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{read_primers_multiple}
\alias{read_primers_multiple}
\title{Input of Multiple Primer Sets.}
\usage{
read_primers_multiple(filenames, fw.id, rev.id, merge.ambig, max.degen,
  template.df, adapter.action, sample.name, updateProgress)
}
\arguments{
\item{filenames}{The paths to multiple primer CSV/FASTA files.}
}
\value{
A list containing objects of class \code{Primers}.
}
\description{
Reads multiple CSV files representing stored objects of class \code{Primers}.
}
\keyword{internal}
