% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{read_settings}
\alias{read_settings}
\title{Loading of Analysis Settings.}
\usage{
read_settings(filename = list.files(system.file("extdata", "settings", package
  = "openPrimeR"), pattern = "*.xml", full.names = TRUE), frontend = FALSE)
}
\arguments{
\item{filename}{Path to a valid XML file containing the 
primer analysis settings. By default, \code{filename} is set
to all settings that are shipped with openPrimeR and the lexicographically
first file is loaded.}

\item{frontend}{Indicates whether settings shall be loaded for the Shiny frontend. 
In this case no unit conversions for the PCR settings are performed.
The default setting is \code{FALSE} such that the correct units are used.}
}
\value{
An object of class \code{DesignSettings}.
}
\description{
Loads primer analysis settings from an XML file.
}
\details{
If \code{filename} is not provided,
a default XMl settings file is loaded. Please review the 
function's examples to learn more about the default settings. If you want
to load custom settings, you can store a modified \code{DesignSettings}
object as an XML file using \code{\link{write_settings}}.
}
\examples{
# Select the available settings
#' the available supplied settings by calling  
available.settings <- list.files(
     system.file("extdata", "settings", package = "openPrimeR"), 
     pattern = "*.xml", full.names = TRUE)
# Select one of the settings and load them
filename <- available.settings[1]
settings <- read_settings(filename)
# Modify, store, and read a settings object:
constraints(settings)$gc_clamp <- c("min" = 0, "max" = 5)
out.file <- tempfile("my_settings", fileext = ".xml")
write_settings(settings, out.file)
my_settings <- read_settings(out.file)
}
\seealso{
Other settings functions: \code{\link{ConstraintOptions-class}},
  \code{\link{ConstraintSettings-class}},
  \code{\link{CoverageConstraints-class}},
  \code{\link{DesignSettings-class}},
  \code{\link{PCR_Conditions-class}}, \code{\link{PCR}},
  \code{\link{conOptions}}, \code{\link{constraintLimits}},
  \code{\link{constraints}}, \code{\link{cvg_constraints}}
}
