% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{read_templates}
\alias{read_templates}
\title{Input of Template Sequences.}
\usage{
read_templates(template.file, hdr.structure = NULL, delim = NULL,
  id.column = NULL, rm.keywords = NULL, remove.duplicates = FALSE,
  fw.region = c(1, 30), rev.region = c(1, 30), gap.character = "-",
  run = NULL)
}
\arguments{
\item{template.file}{Path to one or multiple FASTA or CSV files 
containing the template sequences.}

\item{hdr.structure}{A character vector describing the information contained in the FASTA headers. In case that the headers of \code{fasta.file} contain
template group information, please include the keyword "GROUP" in
\code{hdr.structure}. If the numer of elements provided via \code{hdr.structure}
is shorter than the actual header structure, the missing fields are ignored.}

\item{delim}{Delimiter for the information in the FASTA headers.}

\item{id.column}{Field in the header to be used as the identifier of individual
template sequences.}

\item{rm.keywords}{A vector of keywords that are used to remove templates whose headers contain any of the keywords.}

\item{remove.duplicates}{Whether duplicate sequence shall be removed.}

\item{fw.region}{The positional interval from the template 5' end specifying the
binding sites for forward primers. The default \code{fw.region} is set to
the first 30 bases of the templates.}

\item{rev.region}{The positional interval from the template 3' end specifying
the binding sites for reverse primers. The default \code{rev.region}
is set to the last 30 bases of the templates.}

\item{gap.character}{The character in the input file representing gaps. 
Gaps are automatically removed upon input and the default character is "-".}

\item{run}{An identifier for the set of template sequences. By default,
\code{run} is \code{NULL} and its value is set via \code{template.file}.}
}
\value{
An object of class \code{Templates}.
}
\description{
Read one or multiple files with template sequences in FASTA or CSV format.
}
\details{
When supplying a FASTA file with template sequences,
the input arguments \code{hdr.structure}, \code{delim}, \code{id.column}, 
\code{rm.keywords}, \code{remove.duplicates}, \code{fw.region}, 
\code{rev.region}, \code{gap.character}, and \code{run} are utilized.
Most importantly, \code{hdr.structure} and \code{delim} should
match the FASTA header structure.
To learn more about setting the primer binding regions, consider the
\code{\link{assign_binding_regions}} function.
In contrast, when supplying a CSV file with template sequences,
the data are loaded without performing any modifications because the CSV file
should represent an object of class \code{\link{Templates}}, which 
can be stored using the \code{\link{write_templates}} function.
}
\examples{
# Load templates from a FASTA file
fasta.file <- system.file("extdata", "IMGT_data", "templates", 
          "Homo_sapiens_IGH_functional_exon.fasta", package = "openPrimeR")
hdr.structure <- c("ACCESSION", "GROUP", "SPECIES", "FUNCTION")
template.df.fasta <- read_templates(fasta.file, hdr.structure, "|", "GROUP")
# Load mutliple FASTA files
fasta.files <- c(fasta.file, fasta.file)
template.df.fastas <- read_templates(fasta.files, hdr.structure, "|", "GROUP")
# Load templates from a previously stored CSV file
csv.file <- system.file("extdata", "IMGT_data", "comparison", 
               "templates", "IGH_templates.csv", package = "openPrimeR")
template.df.csv <- read_templates(csv.file)
# Load multiple CSV files:
csv.files <- c(csv.file, csv.file)
template.df.csvs <- read_templates(csv.files)
# Load a mixture of FASTA/CSV files:
mixed.files <- c(csv.file, fasta.file)
template.data <- read_templates(mixed.files)
}
\seealso{
Other template functions: \code{\link{Templates-class}},
  \code{\link{assign_binding_regions}},
  \code{\link{update_template_cvg}},
  \code{\link{write_templates}}
}
\keyword{Templates}
