% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{read_templates_csv}
\alias{read_templates_csv}
\title{Read Template CSV File}
\usage{
read_templates_csv(fname)
}
\arguments{
\item{fname}{The filename of the input template CSV file.}
}
\value{
A template data frame.
}
\description{
Reads an input template CSV file.
}
\keyword{internal}
