% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{read_templates_multiple}
\alias{read_templates_multiple}
\title{Input of Multiple Templates.}
\usage{
read_templates_multiple(filenames, hdr.structure = NULL, delim = NULL,
  id.column = NULL, rm.keywords = NULL, remove.duplicates = FALSE,
  fw.region = c(1, 30), rev.region = c(1, 30), gap.character = "-",
  run = NULL)
}
\arguments{
\item{filenames}{Names of FASTA/CSV files containing template data.}

\item{hdr.structure}{A character vector describing the information contained in the FASTA headers. In case that the headers of \code{fasta.file} contain
template group information, please include the keyword "GROUP" in
\code{hdr.structure}.}

\item{delim}{Delimiter for the information in the FASTA headers.}

\item{id.column}{Field in the header to be used as the identifier.}

\item{rm.keywords}{A vector of keywords that are used to remove templates whose headers contain any of the keywords.}

\item{remove.duplicates}{Whether duplicate sequence shall be removed.}

\item{fw.region}{The positional interval from the template 5' end specifying the 
binding sites for forward primers.}

\item{rev.region}{The positional interval from the template 3' end specifying
the binding sites for reverse primers.}

\item{gap.character}{The character in the input file representing gaps.
Gaps are automatically removed upon input.}

\item{run}{An identifier for the template sequences.}
}
\value{
A list containing objects of class \code{Templates}.
}
\description{
Reads multiple template CSV/FASTA files.
}
\keyword{internal}
