% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefCoverage.R
\docType{data}
\name{RefCoverage}
\alias{RefCoverage}
\alias{feature.matrix}
\alias{ref.data}
\title{Experimental Coverage Data.}
\format{The \code{feature.matrix} data frame contains the properties
of the primer set from Tiller et al. as well as a primer set
that was designed by openPrimeR. The column \code{Experimental_Coverage}
indicates the experimentally determined coverage, while the other
columns relate to properties of the primers that were computed
with openPrimeR. 
The \code{ref.data} list contains the raw experimental coverage
of individual primers from the primer sets from Tiller and openPrimeR, which
both target templates from the IGH locus.
The rows of the data frames indicate primers and the columns indicate
IGH templates for which experimental coverage was determined.
The cell entries are hex codes. Each hex code represents a color 
indicating a certain experimental coverage status. 
Hex codes representing red shades indicate no or little amplification, 
while hex codes for green shades indicate high yields.}
\usage{
data(RefCoverage)
}
\description{
Experimental Coverage Data.
}
\examples{
data(RefCoverage)
}
\references{
Tiller, Thomas, et al. "Efficient generation of monoclonal antibodies from single human B cells by single cell RT-PCR and expression vector cloning." Journal of immunological methods 329.1 (2008): 112-124.
}
\keyword{datasets}
