% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{relax.constraints}
\alias{relax.constraints}
\title{Relaxation of Constraints}
\usage{
relax.constraints(settings, filtered.df, excluded.df, stat.df, template.df,
  mode.directionality = c("fw", "rev"), required.cvg, target.temps = NULL,
  results.loc = NULL)
}
\arguments{
\item{settings}{A \code{DesignSettings} object.}

\item{filtered.df}{Data set of primers that fulfilled all constraints of the filtering procedure.}

\item{excluded.df}{Data frame with excluded primers from the first filtering round.}

\item{stat.df}{Data frame with statistics from filtering.}

\item{template.df}{Template data frame.}

\item{mode.directionality}{Primer direction.}

\item{required.cvg}{Required ratio of covered templates.}

\item{target.temps}{Target melting temperature values.}

\item{results.loc}{The location where intermediary results should be stored.}
}
\value{
A list containing information about the relaxation as well as the filtered primers.
}
\description{
Relax constraints trying to reach the target coverage ratio.
}
\keyword{internal}
