% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_algo.R
\name{relax.opti.constraints}
\alias{relax.opti.constraints}
\title{Relaxation of Optimization Constraints}
\usage{
relax.opti.constraints(cur.opti.constraints, initial.opti.limits,
  initial.opti.constraints)
}
\arguments{
\item{cur.opti.constraints}{List with optimization constraint settings.}

\item{initial.opti.limits}{Initial optimization limits.}

\item{initial.opti.constraints}{Initial optimization constraints.}
}
\value{
Relaxed optimization constraints.
}
\description{
Relax optimization constraints.
}
\keyword{internal}
