% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_ILP.R
\name{remove.redundant.cols}
\alias{remove.redundant.cols}
\title{Removal of Redundant Primers.}
\usage{
remove.redundant.cols(S, cvg.matrix)
}
\arguments{
\item{S}{Indices of primers that are selected in an optimal solution.}

\item{cvg.matrix}{Binary matrix of coverage events.}
}
\value{
Updated indices of selected primers \code{S} such that indices
representing primers with redundant coverage are removed.
}
\description{
Removes redundant primers from an optimal solution.
}
\details{
An optimal solution can contain primers with redundant coverage when using presolve or greedy optimization.
}
\keyword{internal}
