% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{remove.seqs.by.keyword}
\alias{remove.seqs.by.keyword}
\title{Removal of Partial Sequences.}
\usage{
remove.seqs.by.keyword(template.df, keyword = "partial")
}
\arguments{
\item{template.df}{Template data frame.}

\item{Header}{keywords indiating templates that should be excluded.}
}
\value{
Template data frame with partial sequences removed.
}
\description{
Removes partial template sequences.
}
\keyword{internal}
