% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{rename.constraint.options}
\alias{rename.constraint.options}
\title{Renaming of Constraint Options.}
\usage{
rename.constraint.options(constraint.options)
}
\arguments{
\item{constraint.options}{A list with constraint options.}
}
\value{
A list with renamed constraint options.
}
\description{
Renames the input list with constraint options.
}
\keyword{internal}
