% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{render_report}
\alias{render_report}
\title{Renders an rmarkdown file using Pandoc.}
\usage{
render_report(params, report_template, out.file)
}
\arguments{
\item{params}{A list with parameters for the R markdown parser.}

\item{report_template}{A character vector giving the basename of the
Rmarkdown template to use for report creation.}

\item{out.file}{The filename of the report PDF to be created.}
}
\value{
Creates a PDF in \code{out.file} if successful.
}
\description{
Creates a PDF report using rmarkdown and Pandoc by passing the specified
\code{params} to the markdown file given by \code{report_template} and 
storing the PDF in \code{out.file}.
}
\keyword{internal}
