% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_greedy.R
\name{reorder.primer.table}
\alias{reorder.primer.table}
\title{Reorder Primers}
\usage{
\method{reorder}{primer.table}(filtered.primers, primer.ID.order)
}
\arguments{
\item{filtered.primers}{Primer data frame.}

\item{primer.ID.order}{new ordering of IDs in the data frame.}
}
\value{
Reordered primer data frame.
}
\description{
Reorders a primer set according to the IDs of primers.
}
\keyword{internal}
