% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{restriction_hits}
\alias{restriction_hits}
\title{Identification of Restriction Sites.}
\usage{
restriction_hits(bad.regions, DB)
}
\arguments{
\item{DB}{A data frame with restriction enzyme sites.}

\item{bad.region}{IRanges with possible adapter sites.}
}
\value{
A boolean data frame indicating the presence
of adapters for all considered restriction sites.
}
\description{
Identifies restriction sites in a list with putative
restriction sites provided by \code{bad.regions} using
a data frame of restriction sites given by \code{DB}.
}
\keyword{internal}
