% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{restriction_match}
\alias{restriction_match}
\title{Identification of Sequence Matches.}
\usage{
restriction_match(primer.seqs, template.seqs)
}
\arguments{
\item{primer.seqs}{Primer sequences.}

\item{template.seqs}{Template sequences.}
}
\value{
A vector with the template regions matching the 
\code{primer.seqs} best.
}
\description{
Determines the most similar template sequence for every
input primer sequence. Used to identify regions for
alignment for the identification of restriction sites.
}
\keyword{internal}
