% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{retrieve.leader.region}
\alias{retrieve.leader.region}
\title{Retrieval of Binding Regions}
\usage{
retrieve.leader.region(template.df, direction = c("fw", "rev"), start, end,
  gap.char)
}
\arguments{
\item{template.df}{Template data frame.}

\item{direction}{Identify forward and reverse.}

\item{start}{Start positions.}

\item{end}{End positions.}

\item{gap.char}{The character for gaps in alignments.}
}
\value{
Data frame with information on allowed binding regions.
}
\description{
Retrives information about individual binding regions.
}
\keyword{internal}
