% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguity.R
\name{rev.comp.sequence}
\alias{rev.comp.sequence}
\title{Reverse complement of a sequence}
\usage{
\method{rev}{comp.sequence}(seq)
}
\arguments{
\item{seq}{the input strings}
}
\value{
The reverse complement of the input sequences.
}
\description{
Computes the reverse complement of the input sequences.
}
\keyword{internal}
