% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguity.R
\name{rev.sequence}
\alias{rev.sequence}
\title{Reversion of a sequence}
\usage{
\method{rev}{sequence}(seq)
}
\arguments{
\item{seq}{the input sequence.}
}
\value{
The input sequence in reverse order.
}
\description{
Reverses the input sequences.
}
\keyword{internal}
