% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startApp.R
\name{runTutorial}
\alias{runTutorial}
\title{The openPrimeR Tutorial.}
\usage{
runTutorial(dev = FALSE)
}
\arguments{
\item{dev}{A logical indicating whether to start the development version of the tutorial (default: \code{FALSE}).}
}
\value{
Opens the openPrimeR tutorial in a web browser.
}
\description{
Starts a Shiny app containing the openPrimeR tutorial, which
was built using the learnr package. The application
starts locally and should open a new tab in your default
browser. If no browser is opened, please consider the console
output to identify the local port on which the server is running.
}
\note{
The Shiny app can be started only if you fulfill all 
of the suggested package dependencies for the Shiny framework,
so please ensure that you've installed openPrimeR including
all suggested dependencies.
}
\examples{
# Open the tutorial
\dontrun{
runTutorial()
}
}
