% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{sanitize_path}
\alias{sanitize_path}
\title{Sanitiziation of Filename.}
\usage{
sanitize_path(path, suffix = "", sub.char = "_")
}
\arguments{
\item{path}{The path to the file to be sanitized, without file extension.}

\item{suffix}{The suffix (e.g. ".png") of a file.}

\item{sub_char}{The character used to replacing disallowed chars.}
}
\value{
The sanitized filename
}
\description{
Ensures that a filename is valid for the file system.
}
\keyword{internal}
