% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_primers_tree.R
\name{score.conservation}
\alias{score.conservation}
\title{Conservation Scores}
\usage{
score.conservation(primer.range, ali.entropy)
}
\arguments{
\item{primer.range}{A data frame with starts/ends of primers.}

\item{ali.entropy}{Entropies corresponding to the alignment}
}
\value{
Entropies indicating conservation (similarity) of regions.
}
\description{
Scores the conservation of alignment regions.
}
\keyword{internal}
