% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{score_conservation}
\alias{score_conservation}
\title{Scoring of Template Conservation.}
\usage{
score_conservation(template.df, gap.char = "-", win.len = 30,
  by.group = TRUE)
}
\arguments{
\item{template.df}{A \code{Templates} object providing the sequence
conservation shall be determined.}

\item{gap.char}{The alignment gap character. By default, this is set to "-".}

\item{win.len}{The size of a window for evaluating conservation.
The default window size is set to 30.}

\item{by.group}{Whether the determination of binding regions 
should be stratified according to the groups defined in \code{template.df}.
The default is \code{TRUE}.}
}
\value{
A list containing \code{Entropies} and \code{Alignments}.
Entropies are given as a data frame with conservation scores. 
Each column indicates a position in the alignment of template sequences
and each row gives the entropies of the sequences 
belonging to a specific group of template sequences.
Alignments are given as lists of \code{DNABin} objects, where each
object gives the alignment corresponding to one group of template sequences.
}
\description{
Determines the sequence conservation scores of a set of templates
using Shannon entropy.
}
\note{
Requires the MAFFT software for multiple alignments (http://mafft.cbrc.jp/alignment/software/).
}
\examples{
\dontrun{
data(Ippolito)
entropy.data <- score_conservation(template.df, gap.char = "-", win.len = 18, by.group = TRUE)
}
}
