% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_primers_tree.R
\name{score_degen}
\alias{score_degen}
\title{Primer Degeneration Score.}
\usage{
score_degen(seq, gap.char = "-")
}
\arguments{
\item{seq}{A list of vectors with single characters.}

\item{gap.char}{The gap character in sequences.}
}
\value{
The number of unambiguous sequences represented by \code{seq}.
}
\description{
Determines the degeneration score of a sequence.
}
\details{
The degeneration of an ambiguous sequence is defined as the number
of unambiguous sequences that the ambiguous sequence represents.
Let a sequence \code{S} of length \code{n} be represented by a collection of sets such that
\deqn{S = {s_1, s_2, \ldots, s_n}}
where \eqn{s_i} indicates the set of unambiguous bases found
at position \eqn{i} of the primer. Then the degeneracy \eqn{D} of a primer
can be defined as 
\deqn{D = \prod_i{|s_i|}}
where \eqn{|s_i|} provides the number of disambiguated bases at position \eqn{i}.
}
\seealso{
Other primer functions: \code{\link{Primers-class}},
  \code{\link{check_constraints}},
  \code{\link{check_restriction_sites}},
  \code{\link{create_report}},
  \code{\link{design_primers}},
  \code{\link{filter_primers}},
  \code{\link{get_initial_primers}},
  \code{\link{primer_significance}},
  \code{\link{write_primers}}
}
