% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{select.allowed.binding.events}
\alias{select.allowed.binding.events}
\title{Selection of Binding Events}
\usage{
select.allowed.binding.events(bound.fw, bound.to.allowed.region.fw,
  allowed.other.binding.ratio)
}
\arguments{
\item{bound.fw}{Indices of covered templates of a single primer.}

\item{bound.to.allowed.region.fw}{Corresponding allowed binding regions.}

\item{allowed.other.binding.ratio}{The ratio of other binding events. If
this is different from 0, disallowed binding events will also be reported.}
}
\value{
The indices of the allowed binding events.
}
\description{
Selects primer binding events that are within the allowed binding regions.
}
\keyword{internal}
