% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_ILP.R
\name{select.best.ILP}
\alias{select.best.ILP}
\title{Selection of Best ILP}
\usage{
select.best.ILP(ILP.df)
}
\arguments{
\item{ILP.df}{Data frame with ILP result properties.}
}
\value{
The index of the best solution.
}
\description{
Selects the best solution from multiple solved ILP instances.
}
\keyword{internal}
