% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_greedy.R
\name{select.best.opti.result}
\alias{select.best.opti.result}
\title{Selection of Best Greedy Result}
\usage{
select.best.opti.result(opti.results, template.df)
}
\arguments{
\item{opti.results}{List with primer data frames for different target melting temperatures.}

\item{template.df}{Template data frame.}
}
\value{
The index of the best primer data set.
}
\description{
Selects best greedy primer data set.
}
\keyword{internal}
