% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_algo.R
\name{select.best.primer.set}
\alias{select.best.primer.set}
\title{Selection of Best Primer Set.}
\usage{
select.best.primer.set(stats)
}
\arguments{
\item{stats}{Statistics of the primer sets to be evaluated.}
}
\value{
The index of the best primer set.
}
\description{
Selects the best primer set according to coverage and melting temperature differences among primers in the set.
}
\keyword{internal}
