% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{select.binding.events}
\alias{select.binding.events}
\title{Selection of Individual Binding Events}
\usage{
select.binding.events(fw.binding.filtered, p.idx)
}
\arguments{
\item{fw.binding.filtered}{IRanges with binding events.}

\item{p.idx}{Index of binding events to keep.}
}
\value{
An IRanges object containing only the selected binding events.
}
\description{
Selects only binding events of interest.
}
\keyword{internal}
