% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{select.constraints}
\alias{select.constraints}
\title{Selection of Constraints.}
\usage{
select.constraints(active.constraints)
}
\arguments{
\item{active.constraints}{A vector whose names give the constraints to be checked.}
}
\value{
A vector of useable constraint identifiers.
}
\description{
Selects constraints that can be computed according to installed third-party software.
This function is only used for initializing the 'constraint_order' option.
}
\keyword{internal}
