% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{select.min.cross.idx}
\alias{select.min.cross.idx}
\title{Selection of cross dimer index}
\usage{
select.min.cross.idx(deltaG, primers)
}
\arguments{
\item{deltaG}{Data frame with thermodynamic info.}

\item{primers}{The corresponding primers.}
}
\value{
The indices with smallest DeltaG for every primer.
}
\description{
Select the index with the smallest DeltaG value.
}
\keyword{internal}
