% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_primers_tree.R
\name{select.primer.region.by.conservation}
\alias{select.primer.region.by.conservation}
\title{Selection by Conservation}
\usage{
select.primer.region.by.conservation(primer.range, ali.entropy, conservation,
  bin)
}
\arguments{
\item{primer.range}{Data frame with primer starts/stops.}

\item{ali.entropy}{Entropy values for the alignment.}

\item{conservation}{Desired ratio of primer conservation.
Only regions with a conservation of at least \code{conservation}
are considered for the initialization of primers.}

\item{bin}{\code{DNABin} alignment of templates.}
}
\value{
Updated primer regions according to the desired conservation.
}
\description{
Selects primer regions for initialization of primers according
to their conservation scores.
}
\details{
The conservation scores are computed using the entropies computed from
the alignment of the template sequence regions of interests.
}
\keyword{internal}
