% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_primer_coverage.R
\name{select_best_binding}
\alias{select_best_binding}
\title{Selection of Best (smallest number of mismatches) Binding Event per Template Coverage Event.}
\usage{
select_best_binding(binding, fw.mm.info)
}
\arguments{
\item{binding}{Binding information.}

\item{fw.mm.info}{Info about mismatches.}
}
\value{
A list with entries 'fw' and 'rev' giving the best indices of primers.
}
\description{
Selection of Best (smallest number of mismatches) Binding Event per Template Coverage Event.
}
\keyword{internal}
