% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{select_regions_by_conservation}
\alias{select_regions_by_conservation}
\title{Selection of Primer Binding Regions by Conservation.}
\usage{
select_regions_by_conservation(template.df, gap.char = "-", win.len = 30,
  by.group = TRUE, direction = c("both", "fw", "rev"))
}
\arguments{
\item{template.df}{A \code{Templates} object containing
the template sequences for which the binding regions
shall be determined according to conservation.}

\item{gap.char}{The alignment gap character. This is "-" by default.}

\item{win.len}{The extent of the desired primer binding region.
This should be smaller than the \code{allowed.region}. The default is 30.}

\item{by.group}{Shall the determination of binding regions be stratified
according to the groups defined in \code{template.df}. By default,
this is set to \code{TRUE}.}

\item{direction}{Whether regions shall be selected for primers of
both directions ("both"), forward primers ("fw"), or reverse primers ("rev").
The default is "both".}
}
\value{
A \code{Templates} object with adjusted binding regions.
The attribute \code{entropies} gives a data frame with positional entropies
and the attribute \code{alignments} gives the alignments of the templates.
}
\description{
Computes Shannon entropy for putative binding regions
and determines the most conserved regions.
}
\note{
Requires the MAFFT software for multiple alignments (http://mafft.cbrc.jp/alignment/software/).
}
\examples{
\dontrun{
data(Ippolito)
new.template.df <- select_regions_by_conservation(template.df)
}
}
