% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{selenium.installed}
\alias{selenium.installed}
\title{Determination if Selenium is installed.}
\usage{
selenium.installed()
}
\value{
\code{TRUE} is selenium for python is available,
\code{FALSE} otherwise.
}
\description{
Checks whether selenium module for python is installed on the system.
}
\keyword{internal}
