% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{set.new.constraint.value}
\alias{set.new.constraint.value}
\title{Update Constraint Settings.}
\usage{
set.new.constraint.value(values, relax.speed, cur.limits, cur.setting)
}
\arguments{
\item{values}{Observed constraint values considered for the update.}

\item{relax.speed}{The speed at which the constraints should be relaxed.
This value should be in the interval [0,1].}

\item{cur.limits}{List with current relaxation limits.}

\item{cur.setting}{List with current constraint settings.}
}
\value{
Relaxed constraint settings according to the given \code{values} and \code{cur.limits}.
}
\description{
Updates the constraint settings with new values. Sets to the maximal
observed values within the limits or less if \code{relax.speed} is less than 1.
}
\keyword{internal}
