% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{set.new.limits}
\alias{set.new.limits}
\title{Relaxation of Constraint Limits.}
\usage{
set.new.limits(cur.limits, initial.limits, initial.constraints,
  con.name = NULL)
}
\arguments{
\item{cur.limits}{List with current constraint settings.}

\item{initial.limits}{List with initial coonstraint limits.}

\item{initial.constraints}{List with initial constraint settings before relaxing.}

\item{con.name}{The constraint for which the settings are to be changed.}
}
\value{
A list with relaxed constraint limits.
}
\description{
Relaxes the constraint limits by moving according to the difference
between the \code{initial.limits} and \code{initial.constraints}.
}
\keyword{internal}
