% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_primers_tree.R
\name{shannon.entropy}
\alias{shannon.entropy}
\title{Shannon Entropy}
\usage{
shannon.entropy(ali)
}
\arguments{
\item{ali}{An alignment of primer sequences.}
}
\value{
The Shannon entropy for the alignment.
}
\description{
Computation of Shannon entropy for an alignment.
}
\keyword{internal}
