% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stop_codons.R
\name{split_str_by_index}
\alias{split_str_by_index}
\title{Split a sequence}
\usage{
split_str_by_index(target, index)
}
\arguments{
\item{target}{The target string.}

\item{index}{The position for the split.}
}
\value{
List with splitted strings
}
\description{
Splits a sequence at a specified positions
}
\keyword{internal}
