% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startApp.R
\name{startApp}
\alias{startApp}
\title{The openPrimeR Shiny Application.}
\usage{
startApp()
}
\value{
Opens the Shiny app in a web browser.
}
\description{
Starts the openPrimeR Shiny application. 
A new tab should open in your default
browser. If no browser is opened, please consider the console
output to identify the local port on which the server is running
and manually open the shown URL.
}
\note{
The Shiny app can be started only if you fulfill all 
of the suggested package dependencies for the Shiny framework,
so please ensure that you've installed openPrimeR including
all suggested dependencies.
}
\examples{
# Start the shiny app
\dontrun{
startApp()
}
}
