% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_filtering.R
\name{stats_plot_data}
\alias{stats_plot_data}
\title{Combination of Filtering Stats.}
\usage{
stats_plot_data(stats, stats.relax)
}
\arguments{
\item{stats}{Statistics of the filtering procedure.}

\item{stats.relax}{Statistic of the relaxation procedure.}
}
\value{
A data frame combinin filtering/relaxation stats.
}
\description{
Summarizes filtering/relaxation statistics for plotting.
}
\keyword{internal}
