% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{store.filtering.sets}
\alias{store.filtering.sets}
\title{Writes Filtering Data Sets to Disk.}
\usage{
store.filtering.sets(filtered.df, excluded.df, results.loc, tag = "",
  stat.df = NULL, settings = NULL)
}
\arguments{
\item{filtered.df}{A filtered \code{Primers} set.}

\item{excluded.df}{A set of \code{Primers} that were excluded.}

\item{results.loc}{The location where to store the data.}

\item{tag}{A tag for the output files.}

\item{stat.df}{Data frame with statistics of the filtering procedure.}

\item{settings}{A \code{DesignSettings} object.}
}
\value{
No return value, writes output to disk.
}
\description{
Writes Filtering Data Sets to Disk.
}
\keyword{internal}
