% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_view.R
\name{string.list.format}
\alias{string.list.format}
\title{Format a String List.}
\usage{
string.list.format(values, order.mode = c("percentage", "value"))
}
\arguments{
\item{values}{The string list to format.}

\item{order.mode}{How the result should be ordered.
For "percentage", the strings are ordered by their percentages, while
for "value", the strings are ordered by their values.}
}
\value{
A formatted string with percentage annotations.
}
\description{
Formats a string list, summarizing values with percentages.
}
\keyword{internal}
