% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{string.list.format.total}
\alias{string.list.format.total}
\title{Format Strings}
\usage{
string.list.format.total(values)
}
\arguments{
\item{values}{A comma-separated string with values.}
}
\value{
A percentage-formatted representation of the input string.
}
\description{
Changes the representation of the comma-separated string input.
}
\keyword{internal}
