% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_constraints.R
\name{string.to.IQR}
\alias{string.to.IQR}
\title{Conversion of Comma-Separated String to IQR String}
\usage{
string.to.IQR(string.values)
}
\arguments{
\item{string.values}{A vector of comma-separated numeric strings.}
}
\value{
The IQR corresponding to the input string.
}
\description{
Conversion of Comma-Separated String to IQR String
}
\keyword{internal}
