% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_ILP.R
\name{subset.ILP}
\alias{subset.ILP}
\title{Subset ILP Constructor}
\usage{
\method{subset}{ILP}(primer.df, template.df, k)
}
\arguments{
\item{primer.df}{Primer data frame to be subsetted.}

\item{template.df}{Template data frame.}

\item{k}{Required number of primers to be selected.}
}
\value{
An ILP for choosing the primer subset of size \code{k} with the largest coverage.
}
\description{
Constructs an ILP for selecting optimal primer subsets.
}
\details{
Here, "optimal" refers to a subset of a certain size that maximizes the coverage.
}
\keyword{internal}
