% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_ILP.R
\name{subset_primer_set}
\alias{subset_primer_set}
\title{Optimal Primer Subsets.}
\usage{
subset_primer_set(primer.df, template.df, k = 1, groups = NULL,
  identifier = NULL, cur.results.loc = NULL)
}
\arguments{
\item{primer.df}{An objectc of class \code{Primers} providing the
primers for which optimal subsets should be constructed.}

\item{template.df}{An object of class \code{Templates} providing the
template sequences that are targeted by \code{primer.df}.}

\item{k}{The spacing between generated primer subset sizes. By default,
\code{k} is set to 1 such that all primer subsets are constructed.}

\item{groups}{The identifiers of template groups according to which 
coverage should be determined. By default, \code{groups} is set to 
\code{NULL} such that all all covered templates are considered.}

\item{identifier}{An identifier for storing the primer set. By default,
\code{identifier} is set to  \code{NULL}.}

\item{cur.results.loc}{Directory for storing the results. By default,
\code{cur.results.loc} is set to \code{NULL}, which means that
no results are stored.}
}
\value{
A list with optimal primer subsets, each of class \code{Primers}.
}
\description{
Determines subsets of the input primer set that 
are optimal with regard to the number of covered template 
sequences.
}
\details{
The optimal subsets are identified by solving an integer-linear program.
Since the quality of the primers (in terms of properties) is not taken into
account when creating the subsets, this method should only be used
for primer sets that are already of high quality.
}
\examples{
data(Ippolito)
primer.subsets <- subset_primer_set(primer.df, template.df)
}
