% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\docType{class}
\name{Templates-class}
\alias{Templates-class}
\alias{Templates}
\title{The Templates Class.}
\usage{
Templates(...)
}
\arguments{
\item{...}{A data frame fulfilling the structural requirements
for initializing a \code{Templates} object.}
}
\value{
A \code{Templates} object, an instance of a data frame.
}
\description{
The \code{Templates} class encapsulates a data frame
containing the sequencs of the templates, their binding regions,
as well as additional information (e.g. template coverage).
}
\details{
In the following you can find a description
of the most important columns that can be found
in an object of class \code{Templates}. Note that
angle brackets in the column names 
indicate the existence of multiple possibilities.
\describe{
\item{\code{ID}}{The identifiers of the templates.}
\item{\code{Identifier}}{The internal identifiers of the templates.}
\item{\code{Group}}{The identifiers of the groups that the templates belong to.}
\item{\code{Allowed_Start_<fw|rev>}}{The start of the interval in the templates
where binding is allowed for forward and reverse primers, respectively.}
\item{\code{Allowed_End_<fw|rev>}}{The end of the interval in the templates
where binding is allowed for forward and reverse primers, respectively.}
\item{\code{Allowed_<fw|rev>}}{The template sequence where binding
is allowed for forward and reverse primers, respectively.}
\item{\code{Run}}{An identifier for the set of template sequences.}
\item{\code{Covered_By_Primers}}{The identifiers of primers covering the templates,
when the template coverage has been annotated.}
\item{\code{primer_coverage}}{The number of primers covering the templates,
when the template coverage has been annotated.}
}
}
\examples{
fasta.file <- system.file("extdata", "IMGT_data", "templates", 
     "Homo_sapiens_IGH_functional_exon.fasta", package = "openPrimeR")
hdr.structure <- c("ACCESSION", "GROUP", "SPECIES", "FUNCTION")
template.df <- read_templates(fasta.file, hdr.structure, "|", "GROUP")
}
\seealso{
\code{\link{read_templates}} for loading template sequences,
\code{\link{assign_binding_regions}} for adjusting the primer binding regions,
\code{\link{update_template_cvg}} for setting the template coverage,
\code{\link{plot_template_cvg}} for plotting the template coverage,

Other template functions: \code{\link{assign_binding_regions}},
  \code{\link{read_templates}},
  \code{\link{update_template_cvg}},
  \code{\link{write_templates}}
}
\keyword{Classes}
