% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\docType{methods}
\name{cbind2,Templates,ANY-method}
\alias{cbind2,Templates,ANY-method}
\alias{rbind2,Templates,ANY-method}
\alias{[,Templates,ANY,ANY,ANY-method}
\alias{[,Templates-method}
\alias{$<-,Templates-method}
\title{S4 cbind for Templates.}
\usage{
\S4method{cbind2}{Templates,ANY}(x, y, ...)

\S4method{rbind2}{Templates,ANY}(x, y, ...)

\S4method{[}{Templates,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{$}{Templates}(x, name) <- value
}
\arguments{
\item{x}{The Templates data frame.}

\item{y}{Another data frame.}

\item{...}{Other arguments to the slice operator.}

\item{i}{The row index.}

\item{j}{The column index.}

\item{drop}{Simplify data frame?}

\item{name}{The name of the column.}

\item{value}{The values of the column.}

\item{x}{The Template data frame.}
}
\value{
Cbinded template data frame.

Rbinded template data frame.

Subsetted template data frame.

Templates with replaced column.
}
\description{
S4 cbind function for Templates data frame.

S4 rbind function for templates.

Slicing of Templates data frame object.

Set a column in a Templates data frame.
}
\examples{
data(Ippolito)
template.df <- cbind2(template.df, seq_len(nrow(template.df)))
data(Ippolito)
template.df <- rbind2(template.df, template.df)
data(Ippolito)
template.df <- template.df[1:2,]
data(Ippolito)
template.df$ID[1] <- "newID"
}
\keyword{internal}
