% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tiller.R
\docType{data}
\name{Tiller}
\alias{Tiller}
\alias{tiller.template.df}
\alias{tiller.primer.df}
\alias{tiller.settings}
\title{Evaluated Primer Data from Tiller et al.}
\format{\code{tiller.primer.df} provides a \code{Primers} object, \code{tiller.template.df} provides the corresponding \code{Templates} object, and \code{tiller.settings} provides the \code{DesignSettings} object that was used
for evaluating \code{tiller.primer.df}.
DesignSettings object \code{tiller.settings}.}
\usage{
data(Tiller)
}
\description{
Primer and template data for IGHV from Tiller et al.
}
\examples{
data(Tiller)
tiller.primer.df
tiller.template.df
constraints(tiller.settings)
}
\references{
Tiller, Thomas, et al. "Efficient generation of monoclonal antibodies from single human B cells by single cell RT-PCR and expression vector cloning." Journal of immunological methods 329.1 (2008): 112-124.
}
\keyword{datasets}
