% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{ungap_sequence}
\alias{ungap_sequence}
\title{Ungapping of Sequences.}
\usage{
ungap_sequence(seqs, gap.char = "-")
}
\arguments{
\item{seqs}{The input character vector with sequences}

\item{gap.char}{The character used to represent gaps.}
}
\value{
\code{seqs} with gaps removed.
}
\description{
Removes gaps from the input sequences.
}
\keyword{internal}
