% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{unify.leaders}
\alias{unify.leaders}
\title{Unification of Leaders}
\usage{
unify.leaders(l.seq.fw, l.seq.rev, lex.seq, gap.char)
}
\arguments{
\item{l.seq.fw}{Data frame with binding information for forward primers.}

\item{l.seq.rev}{Data frame with binding information for reverse primers.}

\item{lex.seq}{Template data frame.}

\item{gap.char}{The character for indicating alignment gaps.}
}
\value{
Template data frame with annotated binding regions.
}
\description{
Unifies individual binding regions for forward and reverse primers.
}
\keyword{internal}
