% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{update.binding.ranges.by.conservation}
\alias{update.binding.ranges.by.conservation}
\title{Updates Binding Region in the Alignment by conservation.}
\usage{
\method{update}{binding.ranges.by.conservation}(template.df, bins, entropy.df,
  gap.char = "-", win.len = 30, direction = c("fw", "rev"))
}
\arguments{
\item{template.df}{A \code{Templates} object.}

\item{bins}{A list with \code{DNAbin} alignments, one for each group of template sequences.}

\item{entropy.df}{A data frame with entropy information.}

\item{gap.char}{The gap character for alignments.}

\item{win.len}{The desired length of the new binding region.}

\item{direction}{The direction for which the binding range shall be adjusted.}
}
\value{
A \code{Templates} object with modified binding regions.
}
\description{
Updates Binding Region in the Alignment by conservation.
}
\keyword{internal}
