% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{update.binding.regions}
\alias{update.binding.regions}
\title{Update of Binding Regions.}
\usage{
\method{update}{binding.regions}(template.df, opti.regions)
}
\arguments{
\item{template.df}{An object of class \code{Templates}.}

\item{opti.regions}{List with new binding intervals.
The list can contain the components \code{fw} and \code{rev} providing
numeric vectors of length 2 providing the start and end
of the binding regions in the templates, for forward
and reverse binding regions, respectively.}
}
\value{
A \code{Templates} object with updated binding regions.
}
\description{
Updates the binding regions in the templates by providing
new intervals for forward and reverse binding regions.
}
\keyword{internal}
