% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_eval.R
\name{update.constraint.values}
\alias{update.constraint.values}
\title{Update of Primer Constraints.}
\usage{
\method{update}{constraint.values}(constraint.df, constraint.values)
}
\arguments{
\item{constraint.df}{Primer data frame.}

\item{constraint.values}{Data frame with computed constraint values.}
}
\value{
A primer data frame with updated columns.
}
\description{
Updates the input primer data frame with the computed constraint values.
}
\keyword{internal}
