% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{update.individual.binding.region}
\alias{update.individual.binding.region}
\title{Adjustment of Existing Binding Regions for one Direction.}
\usage{
\method{update}{individual.binding.region}(min, max, template.df,
  mode.directionality)
}
\arguments{
\item{min}{Position where binding should start.}

\item{max}{End position of binding.}

\item{mode.directionality}{Directionality of primers.}

\item{Template}{data frame.}
}
\value{
Template data frame with updated binding regions.
}
\description{
Adjusts the existing annotation of binding regions by specifying
an interval relative to the existing binding region.
}
\details{
Position \code{0} indicates the first position after the existing
binding region. Hence, negative positions adjust the binding region towards
the existing binding regions and non-negative positionis extend
the existing binding region definition aways from the existing target region.
}
\keyword{internal}
