% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_algo.R
\name{update.opti.results}
\alias{update.opti.results}
\title{Augmentation of Optimized Primer Data.}
\usage{
\method{update}{opti.results}(primer.df, settings, template.df)
}
\arguments{
\item{primer.df}{A primer data frame.}

\item{settings}{A \code{DesignSettings} object.}
}
\value{
An updated primer data frame.
}
\description{
Adds melting_temp_diff and cross_dimerization info to optimized sets.
}
\keyword{internal}
