% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{update_primer_cvg}
\alias{update_primer_cvg}
\title{Updates the Primer Coverage.}
\usage{
update_primer_cvg(primer.df, template.df, allowed.mismatches,
  cvg.definition = c("constrained", "basic"))
}
\arguments{
\item{primer.df}{A \code{Primers} object.}

\item{template.df}{A \code{Templates} object.}

\item{allowed.mismatches}{A numeric giving the maximal number of allowed.mismatches.}

\item{cvg.definition}{The definition of coverage to be used, either "constrained" or "basic".}
}
\value{
A primer data frame with modified coverage information.
}
\description{
Updates the most important columns in a primer data frame according to the selected
coverage definition. Only coverage events with less or equal than the allowed number of
mismatches according to the selected coverage definition will be retained.
}
\keyword{internal}
