% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{update_template_cvg}
\alias{update_template_cvg}
\title{Annotation of Template Coverage.}
\usage{
update_template_cvg(template.df, primer.df, mode.directionality = NULL)
}
\arguments{
\item{template.df}{An object of class \code{Templates}.}

\item{primer.df}{An object of class \code{Primers} containing
primers with annotated coverage that are to be used to update 
the template coverage in \code{template.df}.}

\item{mode.directionality}{Directionality of primers.
The default is \code{NULL}, which means that the directionality
of primers is identified automatically.}
}
\value{
An object of class \code{Templates} with updated coverage columns.
}
\description{
Annotates the templates with coverage information.
}
\examples{
data(Ippolito)
template.df <- update_template_cvg(template.df, primer.df)
}
\seealso{
Other template functions: \code{\link{Templates-class}},
  \code{\link{assign_binding_regions}},
  \code{\link{read_templates}},
  \code{\link{write_templates}}
}
