% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{validate_primers}
\alias{validate_primers}
\title{Validates a Primers Object.}
\usage{
validate_primers(object)
}
\arguments{
\item{object}{An input data frame to be checked for being a primer data frame.}
}
\value{
\code{TRUE}, if the object is valid, FALSE otherwise.
}
\description{
Checks whether a Primers object is valid or not.
}
\keyword{internal}
