% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{validate_templates}
\alias{validate_templates}
\title{Validates a Template Object.}
\usage{
validate_templates(object)
}
\arguments{
\item{object}{An input data frame to be checked for being a template data frame.}
}
\value{
\code{TRUE}, if the object is valid, \code{FALSE} otherwise.
}
\description{
Checks whether a Templates object is valid or not.
}
\keyword{internal}
