% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_view.R
\name{view.cvg.primers}
\alias{view.cvg.primers}
\title{View the Evaluated Primers.}
\usage{
view.cvg.primers(primer.df, template.df, mode.directionality,
  view.cvg.individual = c("active", "inactive"), for.shiny = TRUE)
}
\arguments{
\item{primer.df}{A \code{Primers} object.}

\item{template.df}{A \code{Templates} object.}

\item{mode.directionality}{The direction of the primers.}

\item{view.cvg.individual}{Whether information on individual coverage events should be retrained.}

\item{for.shiny}{Whether the table is intended for Shiny (HTML) or not.}
}
\value{
A formatted primer table.
}
\description{
Creates a formatted primers table.
}
\keyword{internal}
