% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_dimerization.R
\name{view.dimer.df}
\alias{view.dimer.df}
\title{Formatted dimerization data.}
\usage{
view.dimer.df(dimers, type = c("Self", "Cross"))
}
\arguments{
\item{dimers}{Dimerization data frame.}

\item{type}{Type of dimerization.}
}
\value{
A data frame whose columns are formatted in a user-readable way.
}
\description{
Format a dimerization data frame for frontend output.
}
\keyword{internal}
