% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_view.R
\name{view.input.primers}
\alias{view.input.primers}
\title{View the Input Primers.}
\usage{
view.input.primers(primer.df, mode.directionality, for.shiny = TRUE)
}
\arguments{
\item{primer.df}{A \code{Primers} object.}

\item{mode.directionality}{The direction of the primers.}

\item{for.shiny}{Whether output is intended for Shiny.}
}
\value{
A formatted primer table.
}
\description{
Creates a formatted primers table.
}
\keyword{internal}
