% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO_view.R
\name{view.primers.report}
\alias{view.primers.report}
\title{View the Evaluated Primers in the Report.}
\usage{
view.primers.report(primer.df, template.df)
}
\arguments{
\item{primer.df}{A \code{Primers} object.}

\item{template.df}{A \code{Templates} object.}
}
\value{
A formatted primer table.
}
\description{
Creates a formatted primers table for the report PDF.
}
\keyword{internal}
