% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_algo.R
\name{visualize.all.results}
\alias{visualize.all.results}
\title{Visualization of Design Results.}
\usage{
visualize.all.results(sample, filtering.results.loc, opti.results.loc,
  primer_conc, na_salt_conc, mg_salt_conc, k_salt_conc, tris_salt_conc,
  settings, mode.directionality, used.settings, required.cvg)
}
\arguments{
\item{sample}{Identifier of the design run.}

\item{filtering.results.loc}{Location of filtering results.}

\item{opti.results.loc}{Location of optimization results.}

\item{primer_conc}{Primer concentration.}

\item{na_salt_conc}{Sodium ion concentration.}

\item{mg_salt_conc}{Magensium ion concentration.}

\item{k_salt_conc}{Potassium ion concentration.}

\item{tris_salt_conc}{Tris ion concentration.}

\item{settings}{The \code{DesignSettings} object.}

\item{mode.directionality}{Strand direction for which primers were designed.}

\item{used.settings}{A list with the used settings for optimization (fields "fw" and "rev").}

\item{required.cvg}{The required coverage.}

\item{template_conc}{Template concentration.}
}
\value{
Writes visualizations to files in.
}
\description{
Visualizes all results from designing primers.
}
\keyword{internal}
