% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_algo.R
\name{visualize.filtering.results}
\alias{visualize.filtering.results}
\title{Visualization of Filtering Results.}
\usage{
visualize.filtering.results(sample, results.loc, mode.directionality,
  excluded.df, template.df, filtered.df, filtered.stats, stats.relax,
  primer_conc, na_salt_conc, mg_salt_conc, k_salt_conc, tris_salt_conc,
  settings, required.cvg)
}
\arguments{
\item{sample}{Primer design run identifier.}

\item{results.loc}{Location where the filtering results are stored.}

\item{mode.directionality}{Design direction.}

\item{excluded.df}{Data frame with excluded primers.}

\item{template.df}{Template data frame.}

\item{filtered.df}{Primer data frame containing the primers that passed the constraints.}

\item{stats.relax}{Filtering statistics after relaxation.}

\item{primer_conc}{Primer concentration.}

\item{na_salt_conc}{Sodium ion concentration.}

\item{mg_salt_conc}{Magensium ion concentration.}

\item{k_salt_conc}{Potassium ion concentration.}

\item{tris_salt_conc}{Tris ion concentration.}

\item{settings}{A \code{DesignSettings} object.}

\item{required.cvg}{The required coverage.}

\item{template_conc}{Template concentration.}
}
\value{
Write-out of filtering results.
}
\description{
Visualizes the filtering results.
}
\keyword{internal}
