% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_eval.R
\name{were.constraints.relaxed}
\alias{were.constraints.relaxed}
\title{Check for Relaxation}
\usage{
were.constraints.relaxed(used.constraints, input.constraints)
}
\arguments{
\item{used.constraints}{The constraints that were used during the optimization.}

\item{input.constraints}{The user-specified constraints.}
}
\value{
If \code{input.constraints} was relaxed TRUE is returned, otherwise FALSE.
}
\description{
Determines whether constraints where relaxed or not.
}
\keyword{internal}
