% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{write_primers}
\alias{write_primers}
\title{Storing Primers to Disk.}
\usage{
write_primers(primer.df, fname, ftype = c("FASTA", "CSV"))
}
\arguments{
\item{primer.df}{An object of class \code{Primers} to be stored to disk.}

\item{fname}{The path to the file where the primers should be stored.}

\item{ftype}{A character vector giving the type of the file.
This can either be "FASTA" or "CSV" (default: "FASTA").}
}
\value{
Stores primers to \code{fname}.
}
\description{
Writes a set of primers to disk, either as a FASTA or CSV file.
}
\examples{
data(Ippolito)
# Store primers as FASTA
fname.fasta <- tempfile("my_primers", fileext = ".fasta")
write_primers(primer.df, fname.fasta)
# Store primers as CSV
fname.csv <- tempfile("my_primers", fileext = ".csv")
write_primers(primer.df, fname.csv, "CSV")
}
\seealso{
Other primer functions: \code{\link{Primers-class}},
  \code{\link{check_constraints}},
  \code{\link{check_restriction_sites}},
  \code{\link{create_report}},
  \code{\link{design_primers}},
  \code{\link{filter_primers}},
  \code{\link{get_initial_primers}},
  \code{\link{primer_significance}},
  \code{\link{score_degen}}
}
\keyword{Primers}
