% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{write_settings}
\alias{write_settings}
\title{Storing Design Settings to Disk.}
\usage{
write_settings(settings, filename)
}
\arguments{
\item{settings}{A \code{DesignSettings} object to be stored to disk.}

\item{filename}{A character vector specifying the location 
where the settings should be stored as an XML file.}
}
\value{
Outputs the return status from closing the connection.
}
\description{
Stores primer analysis settings to a file in XML format.
}
\examples{
xml <- settings.xml <- system.file("extdata", "settings", 
       "C_Taq_PCR_high_stringency.xml", package = "openPrimeR")
settings <- read_settings(xml)
out.file <- tempfile("my_settings", fileext = ".xml")
write_settings(settings, out.file)
}
\keyword{Settings}
