% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{write_templates}
\alias{write_templates}
\title{Storing Templates to Disk.}
\usage{
write_templates(template.df, fname, ftype = c("FASTA", "CSV"))
}
\arguments{
\item{template.df}{An object of class \code{Templates} to be stored to disk.}

\item{fname}{The path to the file where the templates should be stored.}

\item{ftype}{A character vector giving the filetype.
This can either be "FASTA" or "CSV" (default: "FASTA").}
}
\value{
Stores templates to \code{fname}.
}
\description{
Stores a set of templates as a FASTA or CSV file.
}
\examples{
data(Ippolito)
# Store templates as FASTA
fname.fasta <- tempfile("my_templates", fileext = ".fasta")
write_templates(template.df, fname.fasta)
# Store templates as CSV
fname.csv <- tempfile("my_templates", fileext = ".csv")
write_templates(template.df, fname.csv, "CSV")
}
\seealso{
Other template functions: \code{\link{Templates-class}},
  \code{\link{assign_binding_regions}},
  \code{\link{read_templates}},
  \code{\link{update_template_cvg}}
}
\keyword{Templates}
