% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{xmlToChar}
\alias{xmlToChar}
\title{Conversion of XML to Character.}
\usage{
xmlToChar(xml)
}
\arguments{
\item{xml}{An xml object to be converted to character.}
}
\value{
A character vector.
}
\description{
Converts an XML object to a character string.
}
\keyword{internal}
