% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{create.options.table}
\alias{create.options.table}
\title{Creation of a Table for Constraint Options.}
\usage{
create.options.table(other.settings)
}
\arguments{
\item{other.settings}{List with constraint options}
}
\value{
A data frame.
}
\description{
Creation of a Table for Constraint Options.
}
\keyword{internal}
