% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization_algo.R
\name{evaluate.fw.rev.combinations}
\alias{evaluate.fw.rev.combinations}
\title{Evaluation of Set Combinations}
\usage{
evaluate.fw.rev.combinations(opti.fw, opti.rev, template.df, cur.target.temps)
}
\arguments{
\item{opti.fw}{List with optimized forward primer sets.}

\item{opti.rev}{List with optimized reverse primer sets.}

\item{template.df}{Template data frame for which primers were designed.}

\item{cur.target.temps}{Target melting temperatures for each primer data frame given
in \code{opti.fw} and \code{opti.rev}.}
}
\value{
List with information on the combinations of forward and reverse primers
as well as the combined data frames themselves.
}
\description{
Evaluates the combinations of forward and reverse primer sets.
}
\keyword{internal}
