% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{methods}
\name{opti}
\alias{opti}
\alias{opti,DesignSettings-method}
\title{Getter for Optimization Constraints.}
\usage{
opti(x)

\S4method{opti}{DesignSettings}(x)
}
\arguments{
\item{x}{A \code{DesignSettings} object.}
}
\description{
Gets the constraints on the physicochemical properties
that are applied just before the optimization procedure
using the \code{Input_Constraints} slot of the provided
\code{DesignSettings} object \code{x}.
}
\keyword{internal}
