% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_coverage.R
\name{plot_template_cvg_mismatches}
\alias{plot_template_cvg_mismatches}
\title{Bar Plot of Template Coverage for Mismatches.}
\usage{
plot_template_cvg_mismatches(primer.df, template.df)
}
\arguments{
\item{primer.df}{A \code{Primers} object.}

\item{template.df}{A \code{Templates} object.}
}
\value{
A plot showing the number of covered template sequences.
}
\description{
Creates a bar plot showing the coverage for every group of template sequences.
}
\keyword{internal}
