% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{update_primer_binding_regions}
\alias{update_primer_binding_regions}
\title{Update of Primer Binding Regions.}
\usage{
update_primer_binding_regions(primer.df, fw.region.old, fw.region,
  rev.region.old, rev.region, template.df)
}
\arguments{
\item{primer.df}{A \code{Primers} data frame.}

\item{fw.region.old}{List with two components containing the old forward binding region interval.}

\item{fw.region}{List with two components containing the new forward binding region interval.}

\item{rev.region.old}{<- List with two components containing the old reverse binding region interval.}

\item{rev.region}{List with two componenets containing the new reverse binding region interval.}

\item{template.df}{Templates in an order corresponding to the specified regions.}
}
\value{
A \code{Primers} object with updated relative binding positions.
}
\description{
Updates the relative primer binding sites in the templates
when the template binding regions have changed since the last
coverage computation.
}
\keyword{internal}
